% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{addList}
\alias{addList}
\title{Add a list to another}
\usage{
addList(lst1, lst2)
}
\arguments{
\item{lst1}{\itemize{
\item list to be added to
}}

\item{lst2}{\itemize{
\item list to add
}}
}
\value{
the expanded list
}
\description{
function to add a list to another.
}
\details{
An element of \code{lst2} with the same name as an element in \code{lst1} replaces that element.
Otherwise, the elements of \code{lst2} are appended to \code{lst1}.
}
