% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{advance}
\alias{advance}
\title{Advance (increment) the global counter `iln` by 1}
\usage{
advance()
}
\value{
NULL.
}
\description{
Function to advance (increment) the global counter `iln` by 1.
}
\details{
On return, `iln` in the global environment will have been advanced by 1
}
