% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllometry.R
\name{allom_pwrLaw1}
\alias{allom_pwrLaw1}
\title{Calculate weight-at-size}
\usage{
allom_pwrLaw1(pA, pB, z)
}
\arguments{
\item{pA}{\itemize{
\item weight (kg) when z = 1 mm
}}

\item{pB}{\itemize{
\item exponent on size
}}

\item{z}{\itemize{
\item sizes at which to calculate weights (mm-scale)
}}
}
\value{
weight in kg
}
\description{
Function to calculate weight-at-size.
}
\details{
The returned object is the same dimensions and type as the
input \code{z} object.The formula used is

W(z) = pA * z^{pB}

such that W(z=1mm) = pA in kg;
}
\examples{
# example code
z = seq(25,100,5);
wAtZ = allom_pwrLaw1(0.00027/1000,3.022134,z)

}
