% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllometry.R
\name{allom_pwrLaw2}
\alias{allom_pwrLaw2}
\title{Calculate weight-at-size}
\usage{
allom_pwrLaw2(pLnA, pLnS, pB, pZ0, z)
}
\arguments{
\item{pLnA}{\itemize{
\item log-scale weight when z = pZ0
}}

\item{pLnS}{\itemize{
\item log-scale units conversion to kg (fixed parameter)
}}

\item{pB}{\itemize{
\item exponent on relative size
}}

\item{pZ0}{\itemize{
\item reference size (fixed parameter)
}}

\item{z}{\itemize{
\item sizes at which to calculate weights
}}
}
\value{
weight in kg
}
\description{
Function to calculate weight-at-size.
}
\details{
The returned object is the same dimensions and type as the
input \code{z} object.The formula used is

W(z) = exp{pLnA-pLnS + pB*log(z/pZ0)}

such that W(z=pZ0) = exp(pLnA)/exp(pLnS);
}
\examples{
# example code
z = seq(25,100,5);
wAtZ = allom_pwrLaw2(20,1000,2.0,50.0,z)

}
