% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged_array.R
\name{as.vector.ragged_array}
\alias{as.vector.ragged_array}
\title{Convert a ragged_array to a vector}
\usage{
\method{as.vector}{ragged_array}(x, mode = "any")
}
\arguments{
\item{x}{- the ragged_array}

\item{mode}{- mode for resulting vector (default="any")}
}
\value{
vector of specified mode (default mode is the same as that of the underlying vector)
}
\description{
Function to convert a ragged_array to a vector.
}
\details{
Returns the vector underlying the ragged_array without the "dfrDims" attribute.
This is an S3 generic method, so just ue `as.vector(x)`
}
