% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{asclogistic}
\alias{asclogistic}
\title{Calculate an ascending logistic function parameterized by z50 and slope}
\usage{
asclogistic(z, params = NULL, pZ50 = NUL, pSlp = NULL, refZ = 0, debug = FALSE)

asclogistic(z, params = NULL, pZ50 = NUL, pSlp = NULL, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivities
}}

\item{params}{\itemize{
\item 2-element vector with selectivity function parameters
}}

\item{pZ50}{\itemize{
\item AD size at 50\\% selectivity
}}

\item{pSlp}{\itemize{
\item AD slope at \code{pZ50}
}}

\item{refZ}{\itemize{
\item reference size
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}

\item{pRefZ}{\itemize{
\item reference size (AD, but a fixed value)
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
vector of selectivity values at the elements of \code{z}

named vector with selectivity values at the elements of \code{z}
}
\description{
Function to calculate an ascending logistic function parameterized by z50 and slope.

Function to calculate an ascending logistic function parameterized by z50 and slope.
}
\details{
The parameter pRefZ is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.

If \code{pRefZ}>0, \code{pRefZ}=fully-selected size.
If \code{pRefZ}<0, function is normalized to max.
If \code{pRefZ}=0, no re-scaling is done.


}
