% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{asclogistic50D95}
\alias{asclogistic50D95}
\title{Calculate an ascending logistic function parameterized by z50 and D95 (=z95-z50)}
\usage{
asclogistic50D95(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivity values
}}

\item{params}{\itemize{
\item 2-element vector with selectivity function parameters
}}

\item{refZ}{\itemize{
\item reference size
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending logistic function parameterized by z50 and D95.
}
\details{

}
