% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{ascnormal}
\alias{ascnormal}
\title{Calculate an ascending normal selectivity curve}
\usage{
ascnormal(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivity values}

\item{params}{- 2-element vector of selectivity function parameters}

\item{refZ}{- reference size (not used)}

\item{debug}{- flag (T/F) to print debugging messages}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending normal selectivity curve.
}
\details{
The parameters are
\itemize{
\item{params[1] - width of ascending limb (standard deviation of a normal distribution)}
\item{params[2] - size at which ascending limb hits 1 (mean of a normal distribution)}
}
}
