% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{ascnormal1}
\alias{ascnormal1}
\title{Calculate an ascending normal selectivity curve}
\usage{
ascnormal1(
  z,
  params = NULL,
  ascWdZ = NULL,
  ascMdZ = NULL,
  refZ = 0,
  debug = FALSE
)

ascnormal1(
  z,
  params = NULL,
  ascWdZ = NULL,
  ascMdZ = NULL,
  refZ = 0,
  debug = FALSE
)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivity values
}}

\item{params}{\itemize{
\item 2-element vector of selectivity function parameters
}}

\item{refZ}{\itemize{
\item reference size (not used)
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}

\item{pAscZ1}{\itemize{
\item AD size at which ascending limb hits 1 (mean of a normal distribution)
}}

\item{pAscWdZ}{\itemize{
\item AD width of ascending limb (standard deviation of a normal distribution)
}}

\item{dZ}{\itemize{
\item bin size
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
vector with selectivity values at the elements of z

named vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending normal selectivity curve.

Function to calculate an ascending normal selectivity curve.
}
\details{
The parameter pRefZ is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.
This function uses a left square wave \code{\link[=squarewave_left]{squarewave_left()}} that rises
from 0 to 1 "at" pAscZ1 to create a differentiable function.

The parameters are
\itemize{
\item{params\link{1} - ascWdZ: width of ascending limb (standard deviation of a normal distribution)}
\item{params\link{2} - ascMdZ: size at which ascending limb hits 1 (mode of a normal distribution)}
}
}
