% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{ascnormal2a}
\alias{ascnormal2a}
\title{Calculate an ascending normal selectivity curve}
\usage{
ascnormal2a(z, params, refS = 0.5, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute function values}

\item{params}{- vector of function parameters}

\item{refS}{- reference selectivity value (default=0.5)}

\item{debug}{- flag (T/F) to print debugging messages}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending normal selectivity curve
}
\details{
The parameter vector has values
\itemize{
 \item{params[1]: size at which selectivity = refS}
 \item{params[2]: size at which ascending limb reaches 1}
}
}
