% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{ascnormal3}
\alias{ascnormal3}
\title{Calculates a 2-parameter ascending normal selectivity curve}
\usage{
ascnormal3(z, params, refZ = 0, debug = FALSE)

ascnormal3(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item dvector of sizes at which to compute function values
}}

\item{params}{\itemize{
\item dvar_vector of function parameters
}}

\item{refZ}{\itemize{
\item 2-element vector of reference sizes (see details)
}}

\item{pDZ1}{\itemize{
\item max possible size at which the curve could reach 1
}}

\item{pSatZ2}{\itemize{
\item selectivity at size=pRefZ2
}}

\item{pMxZ1}{\itemize{
\item max possible size at which the curve could reach 1 (constant)
}}

\item{pRefZ2}{\itemize{
\item reference size at which curve reaches the value of pSatZ2 (constant)
}}
}
\value{
\itemize{
\item vector of selectivity values
}

\itemize{
\item named vector of selectivity values
}
}
\description{
Function to calculate a 2-parameter ascending normal selectivity curve.

Function to calculate a 2-parameter ascending normal selectivity curve.
}
\details{
The parameters \code{pMxZ1} and \code{pRefZ2} are constanta and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.
This function uses a left square wave \code{\link[=squarewave_left]{squarewave_left()}} that rises
from 0 to 1 "at" \code{pMxZ1-pDZ1} to create a differentiable function.

Calculates 2-parameter normal function parameterized by
\itemize{
\item params\link{1}: delta from max possible size (refZ\link{1}) at which ascending limb could reach 1
\item params\link{2}: selectivity at size=refZ\link{2}
}:
\code{refZ} is a 2-element vector with elements
\itemize{
\item refZ\link{1} - max possible size at which the curve could reach 1
\item refZ\link{2} - reference size at which curve reaches the value of param\link{2}
}
}
