% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Matrices.R
\name{buildBlockDiagonalMat}
\alias{buildBlockDiagonalMat}
\title{Build a block-diagonal matrix from building block matrices}
\usage{
buildBlockDiagonalMat(..., force = FALSE)
}
\arguments{
\item{...}{\itemize{
\item individual matrices or a list of matrices
}}
}
\value{
block-diagonal matrix, either an "adsparse" matrix or inheriting from \link[Matrix:Matrix]{Matrix::Matrix} class \code{CsparseMatrix}.
}
\description{
Build a block-diagonal matrix from building block matrices.
}
