% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{buildTemplateQ.ar1}
\alias{buildTemplateQ.ar1}
\title{Build a template for a precision matrix with AR1 structure
across parameters within a group.}
\usage{
buildTemplateQ.ar1(np, ng, sequential = TRUE)
}
\arguments{
\item{np}{\itemize{
\item number of (RE) parameters per group
}}

\item{ng}{\itemize{
\item number of groups (levels) in grouping factor
}}
}
\value{
a sparse matrix \code{T} of type Matrix::dgCMatrix
}
\description{
Build a template for a precision matrix with AR1 structure
across parameters within a group.
}
\details{
In \code{T}, parameter ordering is considered to be within group, then by group.
If \code{sequential}=TRUE, the related (2*\code{np}-1)-element covariance parameter vector should be ordered as
c(sigma's, rho's) (i.e., sequentially). If \code{sequential}=FALSE, sigma, rho values should be
interleaved as c(\code{sigma[1]},\code{rho[1]},\code{sigma[2]},\code{rho[2]},...,\code{rho[np-1]},\code{sigma[np]}).
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
np = 5; #--number of (RE) parameters
ng = 2; #--number of groups (levels) in grouping factor
#--number of covariance parameters is 2/group (i.e., sigma, rho)
T = buildTemplateQ.ar1(np,ng);
T
T@x; #--covariance parameter vector indices

T = buildTemplateQ.ar1(np,ng,sequential=FALSE);
T
T@x; #--covariance parameter vector indices
\dontshow{\}) # examplesIf}
}
