% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{buildTemplateQ.diag}
\alias{buildTemplateQ.diag}
\title{Build a template for a diagonal precision matrix}
\usage{
buildTemplateQ.diag(np, ng)
}
\arguments{
\item{np}{\itemize{
\item number of (RE) parameters per group, as well as the number of variance parameters
}}

\item{ng}{\itemize{
\item #--number of groups (levels) in grouping factor
}}
}
\value{
a \eqn{(\code{np} \cdot \code{ng}) x (\code{np} \cdot \code{ng})} sparse diagonal matrix \code{T} of type Matrix::dgCMatrix
}
\description{
Function to build a template for a diagonal precision matrix.
}
\details{
Conceptually, a \eqn{(\code{np} x \code{np})} sparse diagonal matrix \code{T_1} is repeated \code{ng} times along
the block diagonal to form \code{T}.

Note that \code{idx = T@x} provides a mapping from the \code{np} covariance parameters \code{cp} to a sparse diagonal precision matrix \code{Q}
of size \eqn{(\code{np} \cdot \code{ng}) x (\code{np} \cdot \code{ng})} by \code{Q@x = cp[idx]}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
np = 2; #--number of (RE) parameters/group
ng = 5;  #--number of groups (levels) in grouping factor
isigma = rep(1:np,ng);
T = Matrix::.sparseDiagonal(q,x=isigma,shape="g"); #--template for Q
Tp = buildTemplateQ.diag(np,ng);
T-Tp
\dontshow{\}) # examplesIf}
}
