% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{buildTemplateQ.us}
\alias{buildTemplateQ.us}
\title{Build a template for a precision matrix with a general structure
across parameters within a group (i.e., an unstructured precision matrix)}
\usage{
buildTemplateQ.us(np, ng, byDiagonal = TRUE)
}
\arguments{
\item{np}{\itemize{
\item number of (RE) parameters per group
}}

\item{ng}{\itemize{
\item number of groups (levels) in grouping factor
}}
}
\value{
a sparse \eqn{(\code{np} \cdot \code{ng}) x (\code{np} \cdot \code{ng})} matrix \code{T} of type Matrix::dgCMatrix
}
\description{
Function to build a template for a precision matrix with a general structure
across parameters within a group (i.e., an unstructured precision matrix).
}
\details{
If \code{T_1} is the template for a single level, then the matrix with \code{T_1} repeated \code{ng} times
along the block diagonal is the template for the full precision matrix. The parameter ordering for \code{T_1}
is from the diagonal outward in one direction (\code{T_1} is symmetric, of course). Thus, the \eqn{(\begin{pmatrix}np+1 \\ 2 \end{pmatrix})}-element
covariance parameter vector should be ordered as
c(\eqn{\sigma}'s, lag1 \eqn{\rho}'s, lag 2 \eqn{\rho}'s, etc).
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
np = 5; #--number of (RE) parameters
ng = 2; #--number of groups (levels) in grouping factor
#--number of covariance parameters is `choose(np+1,2)` (i.e., \eqn{\sigma}, \eqn{\rho})
T = buildTemplateQ.us(np,ng);
T
T@x; #--covariance parameter vector indices
\dontshow{\}) # examplesIf}
}
