% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllometry.R
\name{calcAllometry}
\alias{calcAllometry}
\title{Calculate allometry for all model categories across time}
\usage{
calcAllometry(dims, info, params, verbose = FALSE, loopIC_ = TRUE)
}
\arguments{
\item{dims}{\itemize{
\item dimensions list
}}

\item{info}{\itemize{
\item info list (output list from \code{\link[=extractParamInfo_Allometry]{extractParamInfo_Allometry()}})
}}

\item{params}{\itemize{
\item RTMB parameters list with allometry-specific elements
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
3-d array with dimensions \verb{\[nYs, nSs,nCs\]}, where \code{nYs} is the number
of model years, \code{nSs} is the number of seasons/year, and \code{nCs} is the number of
population categories.
}
\description{
Function to calculate allometryfor all model categories across time.
}
