% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFisheriesCaptureRates.R
\name{calcFisheriesCaptureRates}
\alias{calcFisheriesCaptureRates}
\title{Calculate fully-selected fisheries capture rates for all model categories across time}
\usage{
calcFisheriesCaptureRates(dims, info, params, verbose = FALSE)
}
\arguments{
\item{dims}{\itemize{
\item dimensions list
}}

\item{info}{\itemize{
\item info list (output list from \code{\link[=extractParamInfo_FisheryCaptureRates]{extractParamInfo_FisheryCaptureRates()}})
}}

\item{params}{\itemize{
\item RTMB parameters list with elements specific to fisheries rates
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
A list by fishery fleet of fully-selected fishery captures as an array
}
\description{
Function to calculate fully-selected fisheries capture rates for all model categories across time.
}
\details{
If \code{f_}, \code{y_}, and \code{s_} are the fishery fleet, year, and season of interest,
then the fully-selected fisheries capture rates by population category \code{ic_} are:
fshCRs = lst[\link{f_}][y_,s_,ic_];
}
