% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFisheriesRates.R
\name{calcFisheriesRates}
\alias{calcFisheriesRates}
\title{Calculate size-specific fisheries capture and mortality rates for all model categories across time}
\usage{
calcFisheriesRates(dims, info, lstFCRs, lstHMRs, lstSels, verbose = FALSE)
}
\arguments{
\item{dims}{\itemize{
\item dimensions list
}}

\item{info}{\itemize{
\item info list (output list from \code{\link[=extractParamInfo_FisheriesRates]{extractParamInfo_FisheriesRates()}})
}}

\item{lstFCRs}{\itemize{
\item list of selectivity value arrays (output from \code{\link[=calcFisheriesCaptureRates]{calcFisheriesCaptureRates()}})
}}

\item{lstHMRs}{\itemize{
\item list of selectivity value arrays (output from \code{\link[=calcFisheriesHandlingMortalityRates]{calcFisheriesHandlingMortalityRates()}})
}}

\item{lstSels}{\itemize{
\item list of selectivity value arrays (output from \code{\link[=calcSelectivity]{calcSelectivity()}})
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}

\item{params}{\itemize{
\item RTMB parameters list with elements specific to survey catchability
}}
}
\value{
A list by fishery fleet of size-specific fishery capture, retention, discard,
and total mortality rates as an array
}
\description{
Function to calculate sie-specific fisheries capture and mortality rates for all model categories across time.
}
\details{
If \code{f_}, \code{y_}, and \code{s_} are the fishery fleet, year, and season of interest,
then the indices into the population vector and associated fishing rates are
given by \code{ic_} and \code{type_}:

fshRs = lst[\link{f_}][type_,y_,s_,ic_];

where \code{type_} is one of "FCR" (fishery capture rate),"RMR" (retention mortality rate),
"DMR" (discard mortality rate), "TFMR" (total fishing mortality rate)
}
