% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth_PrGr.R
\name{calcGrowth_PrGr}
\alias{calcGrowth_PrGr}
\title{Calculate the probability of undergoing transitions between size categories during growth, for all model categories across time}
\usage{
calcGrowth_PrGr(dims, info, params, verbose = FALSE)
}
\arguments{
\item{dims}{\itemize{
\item dimensions list
}}

\item{info}{\itemize{
\item info list (output list from \code{\link[=extractParamInfo_Growth_PrGr]{extractParamInfo_Growth_PrGr()}})
}}

\item{params}{\itemize{
\item RTMB parameters list with elements specific to the probability of undergoing a transition between size categories
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
a list of a list of matrices, indexed by year, then season
}
\description{
Function to calculate the probability of undergoing transitions between size categories during growth, for all model categories across time.
}
\details{
TBD
}
