% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth_PrM2M.R
\name{calcGrowth_PrM2M}
\alias{calcGrowth_PrM2M}
\title{Calculate the probability of undergoing a transition between maturity categories for all model categories across time}
\usage{
calcGrowth_PrM2M(dims, info, params, verbose = FALSE)
}
\arguments{
\item{dims}{\itemize{
\item dimensions list
}}

\item{info}{\itemize{
\item info list (output list from \code{\link[=extractParamInfo_Growth_PrM2M]{extractParamInfo_Growth_PrM2M()}})
}}

\item{params}{\itemize{
\item RTMB parameters list with elements specific to the probability of undergoing a transition between maturity categories
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
TODO: might want to return a list of a list of matrices
}
\description{
Function to calculate the probability of undergoing a transition between maturity categories for all model categories across time.
}
\details{
TBD
}
