% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth_PrMolt.R
\name{calcGrowth_PrMolt}
\alias{calcGrowth_PrMolt}
\title{Calculate the probability of undergoing a molt for all model categories across time}
\usage{
calcGrowth_PrMolt(dims, info, params, verbose = FALSE, loopIC_ = TRUE)
}
\arguments{
\item{dims}{\itemize{
\item dimensions list
}}

\item{info}{\itemize{
\item info list (output list from \code{\link[=extractParamInfo_Growth_PrMolt]{extractParamInfo_Growth_PrMolt()}})
}}

\item{params}{\itemize{
\item RTMB parameters list with elements specific to the probability of undergoing a molt
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
TODO: might want to return a list of a list of matrices
}
\description{
Function to calculate the probability of undergoing a molt for all model categories across time.
}
\details{
Application of the probability of molting vector, prM, to the
population vector during the growth season decomposes it into molting and
non-molting components.
}
