% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInitN.R
\name{calcInitN}
\alias{calcInitN}
\title{Calculate initial population abundance}
\usage{
calcInitN(dims, options, params, verbose)
}
\arguments{
\item{dims}{\itemize{
\item model dimensions list
}}

\item{options}{\itemize{
\item model options list
}}

\item{params}{\itemize{
\item model parameters list
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
(ad)vector with initial population abundance
}
\description{
Calculate initial population abundance.
}
\details{
Options for calculating the initial N vector include:
\itemize{
\item{zeroPop - all elements of the initial N vector are zero}
\item{inputPop - all elements of the initial N vector are given}
}
}
