% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanGrowthFunctions.R
\name{calcMnGrowthPwrLaw3}
\alias{calcMnGrowthPwrLaw3}
\title{Power law model 2 for mean growth}
\usage{
calcMnGrowthPwrLaw3(z, params, consts, debug = FALSE)

calcMnGrowthPwrLaw3(z, params, consts, debug = FALSE)
}
\arguments{
\item{z}{- vector of pre-molt sizes (bin mid-points)}

\item{params}{- vector of parameters (potentially estimated)}

\item{consts}{- vector of constants (not estimated)}

\item{debug}{- debugging flag (T/F)}
}
\value{
named vector of mean molt increments or post-molt size

named vector of mean molt increments or post-molt size
}
\description{
Function to calculate power law model 2 for mean growth

Function to calculate power law model 3 for mean growth
}
\details{
Mean growth (molt increment *or* post-molt size) `g` is described as
a power law function of pre-molt size `z` by
g(z|p, c) = p[1] * (z/c[1])^(log(p[2]/p[1])/log(c[2]/c[1]))
where `p` represents the parameter vector and `c` represents the consts vector.
`p[1]` is the mean growth at pre-molt size `c[1]`.
`p[2]` is the mean growth at pre-molt size `c[2]`.

The interpretation of `g` as representing a molt increment or post-molt size
is up to the user.

Mean growth (molt increment *or* post-molt size) `g` is described as
a power law function of pre-molt size `z` by
g(z|p, c) = p[1] * (z/c[1])^p[2]
where `p` represents the parameter vector and `c` represents the consts vector.
`p[1]` is the mean growth at pre-molt size `c[1]`.

The interpretation of `g` as representing a molt increment or post-molt size
is up to the user.
}
