% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Formulas.R
\name{calcModelMatrixECs}
\alias{calcModelMatrixECs}
\title{Calculate a model matrix for "environmental" covariates}
\usage{
calcModelMatrixECs(txt, dfrMF, ctrs = NULL, verbose = FALSE)
}
\arguments{
\item{txt}{\itemize{
\item model formula for environmental covariates as text
}}

\item{dfrMF}{\itemize{
\item basis dataframe for model frame
}}

\item{ctrs}{\itemize{
\item contrasts for any covariate factors
}}

\item{verbose}{\itemize{
\item flag to print debugging info
}}
}
\value{
list with number of parameters (\code{npars}) and, if \code{npars}>0, the model matrix and other information.
}
\description{
Function to calculate a model matrix (and associated information) for "environmental" covariates
}
\details{
This function calculates a model matrix based on the input formula (as text) and
model frame basis. The returned list has elements
\items{
\item{mtxRedMM - the reduced model matrix (for distinct factor and variable levels)}
\item{dfrRedMF - the reduced model (data)frame with distinct factor and variable levels; \code{row_idx} indicates corresponding row in mtxRedMM}
\item{dfrMdMtx - dataframe combining MdMtx and dfrRedMF}
\item{dfrMF - the basis dataframe for the model frame, with an appended column with each value indicating the corresponding row in mtxRedMM}
\item{f - the input formula (as text)}
\item{ns - the number of covariate parameters}
\item{npars - total number of RTMB parameters (same as \code{ns})}
}
The model matrix includes columns for both fixed effects (if any) and smooths (if any),
with fixed effect terms coming first. The number of columns reflects the number of RTMB (i.e., actual)
parameters associated with the model matrix. The model matrix has the same number of rows as the
model frame and \code{dfrMF}. The model frame is a dataframe consisting of a subset of the columns of \code{dfrMF} that
includes only the variables included in the formula,
with the columns corresponding to any smooth variables converted to numeric type and possibly preceded
by a column of 1's with name '(Intercept)' if the formula includes an intercept.

A version of the model matrix with only distinct rows is returned as \code{mtxRedMM}. \code{dfrRedMF} is a dataframe with
the rows corresponding to the (distinct) combination of variable values in the associated row in \code{mtxRedMM}, and
\code{dfrMdMtx} is the concatenated dataframe version of these. Given an appropriate RTMB parameter vector (\code{p}, say),
the values of the model parameter vector derived for each distinct combination of variable values in \code{dfrRedMF} can be
calculated as \code{mtxRedMM} \%*\% p. The basis dataframe (\code{dfrMF}) is included as part of the
returned list to allow

If \code{txt} is not a valid formula (as a character string), a list with only \code{npars}= 0 is returned.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
#--define mdoel DimsMap
vRs="EBS";
miz = as.character(seq(25,75,5));
mmz = as.character(seq(50,75,5));
fiz = as.character(seq(25,65,5));
fmz = as.character(seq(30,65,5));
vXs=list(male=list(imm=list(`new`=miz,
                             `old`=miz),
                    mat=list(`new`=mmz,
                             `old`=mmz,
                             `very`=mmz)
                  ),
          female=list(imm=list(`new`=fiz,
                               `old`=fiz),
                      mat=list(`new`=fmz,
                               `old`=fmz)
                     )
         ); attr(vXs,"dmnms")<-c("x","m","p","z");
ys = as.character(2015:2024); names(ys) = ys;
ss = as.character(1);         names(ss) = ss;
dmsYSC  = createSparseDimsMap(y=ys,s=ss,r=vRs,x=vXs);  #--DimsMap for combination
#--define time block for NMFS Q's
lstNMFS<-list(preCV =as.character(c(2015:2019)),       #--pre gear change
              preGC =as.character(c(2021:2022)),       #--pre gear change
              postGC=as.character(c(2023:2025)));      #--post gear change
tbNMFS = lstNMFS |> defineYearBlocks();
##--expand to dimensions map for NMFS Q's
dmsNMFS<-tbNMFS |>
         defineYearBlocks(dimsMdl=dmsYSC) |>
         dplyr::mutate(f="NMFS");
###--string defining a formula with fixed effects and a smooth on size (`z`)
txt = "~0 + yblk + s(z,k=10)";
##--calculate model matrix and associated information
ctrs = list(yblk="contr.sum");
res = calcModelMatrixFEs(txt,dmsNMFS,ctrs);
\dontshow{\}) # examplesIf}
}
