% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Formulas.R
\name{calcModelMatrixFEs}
\alias{calcModelMatrixFEs}
\title{Calculate a model matrix for fixed effects}
\usage{
calcModelMatrixFEs(txt, dfrMF, ctrs = NULL, verbose = FALSE)
}
\arguments{
\item{txt}{\itemize{
\item model formula as text
}}

\item{dfrMF}{\itemize{
\item basis dataframe for model frame
}}

\item{ctrs}{\itemize{
\item contrasts for linear fixed effects factors
}}

\item{verbose}{\itemize{
\item flag to print debugging info
}}
}
\value{
list with model matrix and other information, or NULL if not \code{txt} is not a formula
}
\description{
Function to calculate a fixed effects model matrix (and associated information)
}
\details{
This function calculates a model matrix based on the input formula (as text) and
model frame basis. The returned list has elements
\items{
\item{MdMtx - the reduced model matrix (for distinct factor and variable levels)}
\item{dfrMdMtx - dataframe with distinct factor and variable levels and reduced model matrix}
\item{fullMdMtx - the full model matrix}
\item{ModFrm - the full model frame}
\item{f - the input formula (as text)}
\item{ns - the number of fixed effects parameters}
\item{ks - an integer vector with the number of knots for each smooth (if any)}
\item{npars - the length of the resulting parameter vector}
\item{SmCons - a list of the smooth construct objects for each smooth (if any)}
}
The model matrix includes columns for both fixed effects (if any) and smooths (if any),
with fixed effect terms coming first. The number of columns reflects the number of actual
parameters associated with the model matrix. The model matrix has the same number of rows as the
model frame and \code{dfrMF}. The model frame is a dataframe consisting of a subset of the columns of \code{dfrMF},
with the columns corresponding to any smooth variables converted to numeric type and possibly preceded
by a column of 1's with name '(Intercept)' if the formula includes an intercept.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
#--define mdoel DimsMap
vRs="EBS";
miz = as.character(seq(25,75,5));
mmz = as.character(seq(50,75,5));
fiz = as.character(seq(25,65,5));
fmz = as.character(seq(30,65,5));
vXs=list(male=list(imm=list(`new`=miz,
                             `old`=miz),
                    mat=list(`new`=mmz,
                             `old`=mmz,
                             `very`=mmz)
                  ),
          female=list(imm=list(`new`=fiz,
                               `old`=fiz),
                      mat=list(`new`=fmz,
                               `old`=fmz)
                     )
         ); attr(vXs,"dmnms")<-c("x","m","p","z");
ys = as.character(2015:2024); names(ys) = ys;
ss = as.character(1);         names(ss) = ss;
dmsYSC  = createSparseDimsMap(y=ys,s=ss,r=vRs,x=vXs);  #--DimsMap for combination
#--define time block for NMFS Q's
lstNMFS<-list(preCV =as.character(c(2015:2019)),       #--pre gear change
              preGC =as.character(c(2021:2022)),       #--pre gear change
              postGC=as.character(c(2023:2025)));      #--post gear change
tbNMFS = lstNMFS |> defineYearBlocks();
##--expand to dimensions map for NMFS Q's
dmsNMFS<-tbNMFS |>
         defineYearBlocks(dimsMdl=dmsYSC) |>
         dplyr::mutate(f="NMFS");
###--string defining a formula with fixed effects and a smooth on size (`z`)
txt = "~0 + yblk + s(z,k=10)";
##--calculate model matrix and associated information
ctrs = list(yblk="contr.sum");
res = calcModelMatrixFEs(txt,dmsNMFS,ctrs);
\dontshow{\}) # examplesIf}
}
