% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Formulas.R
\name{calcModelMatrixREs}
\alias{calcModelMatrixREs}
\title{Calculate a model matrix for random effects}
\usage{
calcModelMatrixREs(txt, dfrMF, cov_type, verbose = FALSE)
}
\arguments{
\item{txt}{\itemize{
\item model formula as text
}}

\item{dfrMF}{\itemize{
\item basis dataframe for model frame
}}

\item{cov_type}{\itemize{
\item covariance type
}}

\item{verbose}{\itemize{
\item flag to print debugging info
}}
}
\value{
list with model matrix and other information, or NULL if not \code{txt} is not a formula
}
\description{
Function to calculate a random effects model matrix (and associated information)
}
\details{
TODO: fill in.
}
