% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNaturalMortality.R
\name{calcNaturalMortality}
\alias{calcNaturalMortality}
\title{Calculate natural mortality for all model categories across time}
\usage{
calcNaturalMortality(dims, info, params, verbose = FALSE, loopIC_ = TRUE)
}
\arguments{
\item{dims}{\itemize{
\item dimensions list
}}

\item{info}{\itemize{
\item info list (output list from \code{\link[=extractParamInfo_NaturalMortality]{extractParamInfo_NaturalMortality()}})
}}

\item{params}{\itemize{
\item RTMB parameters list with natural mortality-specific elements
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
3-d array with dimensions \verb{\[nYs, nSs,nCs\]}, where \code{nYs} is the number
of model years, \code{nSs} is the number of seasons/year, and \code{nCs} is the number of
population categories.
}
\description{
Function to calculate natural mortality for all model categories across time.
}
