% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsim_calcPrM2M.R
\name{calcPrM2M}
\alias{calcPrM2M}
\title{Calculate probability-at-size (pre-molt) for immature crab that a molt is to maturity, by sex and shell condition}
\usage{
calcPrM2M(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{showPlot}{- flag to show plots}
}
\value{
list with elements:
prM2M_cxz
prM2M_yxsz
}
\description{
Function to calculate probability-at-size (pre-molt) for immature crab that a molt is to maturity, by sex and shell condition.
}
\details{
None.
}
