% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsim_calcPrMolt.R
\name{calcPrMolt}
\alias{calcPrMolt}
\title{Calculate probability at size of molting for immature crab by sex, shell condition}
\usage{
calcPrMolt(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{\itemize{
\item model configuration object
}}

\item{showPlot}{\itemize{
\item flag to show plots
}}
}
\value{
prMolt_yxsz
}
\description{
Function to calculate probability at size of molting for immature crab by sex, shell condition.
}
\details{
None.
}
