% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRecruitment.R
\name{calcRecruitment}
\alias{calcRecruitment}
\title{Calculate recruitment-related quantitites}
\usage{
calcRecruitment(dims, options, params, verbose)
}
\arguments{
\item{dims}{- model dimensions list}

\item{options}{- model options list}

\item{params}{- model parameters list}

\item{verbose}{- flag to print diagnostic info}
}
\value{
list with
}
\description{
Function to calculate recruitment-related quantities.
}
\details{
Options for calculating recruitment-related quantities include:
\itemize{
 \item{?? - ??}
}
Returned list includes:
\itemize{
 \item{?? - ??}
}
}
