% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRecruitment_CategoryProportions.R
\name{calcRecruitment_CategoryProportions}
\alias{calcRecruitment_CategoryProportions}
\title{Calculate the category proportions at recruitment for all model categories across time}
\usage{
calcRecruitment_CategoryProportions(
  dims,
  info,
  params,
  verbose = FALSE,
  loopIC_ = FALSE
)
}
\arguments{
\item{dims}{\itemize{
\item dimensions list
}}

\item{info}{\itemize{
\item info list (output list from \code{\link[=extractParamInfo_Recruitment_CategoryProportions]{extractParamInfo_Recruitment_CategoryProportions()}})
}}

\item{params}{\itemize{
\item RTMB parameters list with elements specific to the probability of undergoing a molt
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
TODO: might want to return a list of a list of matrices
}
\description{
Function to calculate the category proportions at recruitment for all model categories across time.
}
\details{
Combination of the category proportions at recruitment vector, prZ, with other aspects of
recruitment yields recruitment by year, season, population category (excluding size) and size.
}
