% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRecruitment_TimeSeries.R
\name{calcRecruitment_TimeSeries}
\alias{calcRecruitment_TimeSeries}
\title{Calculate the bulk recruitment time series across time}
\usage{
calcRecruitment_TimeSeries(
  dims,
  info,
  params,
  verbose = FALSE,
  loopIC_ = FALSE
)
}
\arguments{
\item{dims}{\itemize{
\item dimensions list
}}

\item{info}{\itemize{
\item info list (output list from \code{\link[=extractParamInfo_Recruitment_TimeSeries]{extractParamInfo_Recruitment_TimeSeries()}})
}}

\item{params}{\itemize{
\item RTMB parameters list with elements specific to the recruitment time series
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
TODO: might want to return a list of a list of matrices
}
\description{
Function to calculate the bulk recruitment time series across time.
}
\details{
Combination of the recruitment time series, with other aspects of
recruitment for population category proportions and sizes, yields
recruitment by year, season, population category (excluding size) and size.
}
