% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsim_calcSelFcns.R
\name{calcSelFcns}
\alias{calcSelFcns}
\title{Calculate selectivity/retention functions.}
\usage{
calcSelFcns(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{showPlot}{- flag to show plots}
}
\value{
array sel_cz with dimensions 'pc', 'z'. The label associated with
each pc can be obtained as mc$dims$selfcns$lbls[pc].
}
\description{
Function to calculate selectivity/retention functions.
}
\details{
None.
}
