% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurveysCatchability.R
\name{calcSurveysCatchability}
\alias{calcSurveysCatchability}
\title{Calculate surveys catchability for all model categories across time}
\usage{
calcSurveysCatchability(dims, info, params, lstSels, verbose = FALSE)
}
\arguments{
\item{dims}{\itemize{
\item dimensions list
}}

\item{info}{\itemize{
\item info list (output list from \code{\link[=extractParamInfo_SurveysCatchability]{extractParamInfo_SurveysCatchability()}})
}}

\item{params}{\itemize{
\item RTMB parameters list with elements specific to survey catchability
}}

\item{lstSels}{\itemize{
\item list of selectivity value arrays (output from \code{\link[=calcSelectivity]{calcSelectivity()}})
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
A list of size-specific survey catchability arrays by survey fleet
}
\description{
Function to calculate surveys catchability for all model categories across time.
}
\details{
If \code{f_}, \code{y_}, and \code{s_} are the survey fleet, year, and season of interest,
then the indices into the population vector and associated selectivity values are
given by \code{ic_} and \code{selVals}
srvQs = lst[\link{f_}][y_,s_,ic_];
}
