% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsim_calcZTM.R
\name{calcZTM}
\alias{calcZTM}
\title{Calculate size transition matrices by year, sex, shell condition}
\usage{
calcZTM(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{\itemize{
\item model configuration object
}}
}
\value{
list with elements:
mnZAM_cxz: mean size after molt, by time block and sex
T_cxzz: 4d array with size transition matrices by time block/sex
T_yxszz: 5d array with size transition matrices by year/sex/shell condition
}
\description{
Function to calculate size transition matrices by year, sex, shell condition.
}
\details{
None.
}
