% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Txt.R
\name{concatText}
\alias{concatText}
\title{Concatenate text}
\usage{
concatText(..., concat = "+")
}
\arguments{
\item{...}{- text values to concatenate}

\item{concat}{- string to use as concatenator}
}
\value{
string
}
\description{
Function to concatenate text
}
\details{
The elements in ... are concateated into a single string using the
`concat` string separated by spaces.
}
\examples{
# example code
concatText(1,2,3)
concatText(1,NA,3)
concatText(1,2,NULL,4)
export

}
