% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{const_sel}
\alias{const_sel}
\title{Calculate a constant-valued selectivity curve}
\usage{
const_sel(z, params, ref = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item sizes at which to compute selectivity values
}}

\item{params}{\itemize{
\item 1-element parameter vector
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}

\item{refZ}{\itemize{
\item reference size (dummy input)
}}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate a constant-valued selectivity curve
}
\details{
The parameter vector has values
\itemize{
\item{params\link{1}: the constant value}
}
}
