% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Dataframe.R
\name{convertColsToNum}
\alias{convertColsToNum}
\title{Convert dataframe columns to numeric}
\usage{
convertColsToNum(dfr, cols)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe to convert
}}

\item{cols}{\itemize{
\item names of columns to convert to numeric
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
dataframe with requisite columns converted to numeric
}
\description{
Function to convert dataframe columns to numeric.
}
\details{
None.
}
