% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{convertCovMatToPrecMat}
\alias{convertCovMatToPrecMat}
\title{Convert a covariance matrix to a precision matrix}
\usage{
convertCovMatToPrecMat(C)
}
\arguments{
\item{C}{\itemize{
\item covariance matrix
}}
}
\value{
precision matrix
}
\description{
Convert a covariance matrix to a precision matrix
}
\details{
Uses \link{solve} to convert the covariance matrix to a precision matrix. If
C is a sparse matrix class, the resulting precision matrix may not be.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
##--R context
Q = mkPrecMat(5,c(0,0.5),"ar1",sparse=TRUE);
C = convertPrecMatToCovMat(Q);
t = getThetaFromCovMat(C);
Cp = getCorrMatFromTheta(t);
Qp = convertCovMatToPrecMat(Cp);

##--RTMB context
cov_pars = RTMB::AD(c(0,0.5),force=TRUE);   #--forcing to advector
Q = mkPrecMat(5,cov_pars,"ar1",sparse=TRUE);#--Q is adsparse
as.matrix(Q);
C = convertPrecMatToCovMat(Q);              #--C is adsparse
as.matrix(C);
\dontshow{\}) # examplesIf}
}
