% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth_PrGr.R
\name{convertParams_PwrLaw1to2}
\alias{convertParams_PwrLaw1to2}
\title{Convert parameters/constants values from \code{\link[=grwPwrLaw1]{grwPwrLaw1()}} to \code{\link[=grwPwrLaw2]{grwPwrLaw2()}}}
\usage{
convertParams_PwrLaw1to2(
  pGrA = NULL,
  pGrB = NULL,
  zGrA = NULL,
  zGrB = NULL,
  v = NULL
)
}
\arguments{
\item{pGrA}{\itemize{
\item grwPwrLaw1 parameter pGrA
}}

\item{pGrB}{\itemize{
\item grwPwrLaw1 parameter pGrB
}}

\item{zGrA}{\itemize{
\item grwPwrLaw1 constant zGrA
}}

\item{zGrB}{\itemize{
\item grwPwrLaw1 constant zGrB
}}

\item{v}{\itemize{
\item vector of grwPwrLaw1 parameters in above order
}}
}
\value{
vector of grwPwrLaw2 parameters pGrA and pGrB, plus zGrA and zGrB
}
\description{
Function to convert parameters/constants values from \code{\link[=grwPwrLaw1]{grwPwrLaw1()}} to \code{\link[=grwPwrLaw2]{grwPwrLaw2()}}
}
\details{
Converts parameters/constants for grwPwrLaw1 to those for grwPwrLaw2.
}
\examples{
par0 = c(32.2, 166.0,  25.0, 125.0); #--pGrA, pGrB, zGrA, zGrB
par1 = convertParams_PwrLaw1to2(v=par0);
par2 = convertParams_PwrLaw2to1(v=par1);
sum(abs(par2-par0));#--should be 0

}
