% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth_PrGr.R
\name{convertParams_PwrLaw2to1}
\alias{convertParams_PwrLaw2to1}
\title{Convert parameters/constants values from \code{\link[=grwPwrLaw2]{grwPwrLaw2()}} to \code{\link[=grwPwrLaw1]{grwPwrLaw1()}}}
\usage{
convertParams_PwrLaw2to1(
  pGrA = NULL,
  pGrB = NULL,
  zGrA = NULL,
  zGrB = NULL,
  v = NULL
)
}
\arguments{
\item{pGrA}{\itemize{
\item grwPwrLaw2 parameter pGrA
}}

\item{pGrB}{\itemize{
\item grwPwrLaw2 parameter pGrB
}}

\item{zGrA}{\itemize{
\item grwPwrLaw1 constant zGrA
}}

\item{zGrB}{\itemize{
\item grwPwrLaw1 constant zGrB
}}

\item{v}{\itemize{
\item vector of grwPwrLaw2 parameters + grwPwrLaw1 constants in above order
}}
}
\value{
vector of grwPwrLaw1 parameters pGrA and pGrB, and constants zGrA and zGrB
}
\description{
Function to convert parameters/constants values from \code{\link[=grwPwrLaw2]{grwPwrLaw2()}} to \code{\link[=grwPwrLaw1]{grwPwrLaw1()}}
}
\details{
Converts parameters/constants for grwPwrLaw2 to those for grwPwrLaw1.
}
\examples{
par0 = c(0.1919238, 1.0190025, 25.0, 125.0); #--pGrA, pGrB, zGrA, zGrB
par1 = convertParams_PwrLaw2to1(v=par0);
par2 = convertParams_PwrLaw1to2(v=par1);
sum(abs(par2-par0));#--should be 0

}
