% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{convertPrecMatToCovMat}
\alias{convertPrecMatToCovMat}
\title{Convert a precision matrix to a covariance matrix}
\usage{
convertPrecMatToCovMat(Q)
}
\arguments{
\item{Q}{\itemize{
\item precision matrix
}}
}
\value{
covariance matrix
}
\description{
Convert a precision matrix to a covariance matrix
}
\details{
Uses \link{solve} to convert the precision matrix to a covariance matrix. If
Q is a sparse matrix class, the resulting covariance matrix may not be.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
##--R context
Q = mkPrecMat(5,c(0,0.5),"ar1",sparse=TRUE);
C = convertPrecMatToCovMat(Q);
t = getThetaFromCovMat(C);
Cp = getCorrMatFromTheta(t);
Qp = convertCovMatToPrecMat(Cp);

##--RTMB context
cov_pars = RTMB::AD(c(0,0.5),force=TRUE);   #--force cov_pars to advector
Q = mkPrecMat(5,cov_pars,"ar1",sparse=TRUE);#--Q is adsparse
as.matrix(Q);
C = convertPrecMatToCovMat(Q);              #--C is adsparse
as.matrix(C);
\dontshow{\}) # examplesIf}
}
