% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsUtilities.R
\name{convertToDimsFactors}
\alias{convertToDimsFactors}
\title{Convert values to dimension factor levels in a dataframe}
\usage{
convertToDimsFactors(
  dfr,
  dfrDms,
  contrasts = NULL,
  all = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe to convert
}}

\item{dfrDms}{\itemize{
\item dimensions map dataframe/tibble (e.g. output from \code{\link[=createSparseDimsMap]{createSparseDimsMap()}})
}}

\item{contrasts}{\itemize{
\item character string, list, FALSE, or NULL defining contrasts to apply to (eventual) factor columns
}}

\item{all}{\itemize{
\item flag (T/F) to convert all columns to factors (default=TRUE)
}}
}
\value{
\itemize{
\item dataframe with values converted to factors
}
}
\description{
Function to convert dataframe values to factor levels in a dimensions map.
}
\details{
If a column name in \code{dfr} matches a dimension name,
factor or character columns are converted to a factor with the same levels as the dimension.
This also occurs if the column is numeric and \code{all} is TRUE. Otherwise, the
column is converted to a factor using \code{unique} to determine the levels
if it is a character column or if it numeric and \code{all} is TRUE.

If \code{contrasts} is \code{FALSE}, then no contrasts are applied are applied to any factor. If contrasts is NULL,
then the default contrast given by options("contrast") is applied  to each factor.
Otherwise, contrasts should be a list, with the contrast associated with each name in the list applied to
the corresponding factor.
}
