% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsFunctions.R
\name{createDenseDimsMap}
\alias{createDenseDimsMap}
\title{Create an index map using all (non-nested, fully-crossed) dimension levels}
\usage{
createDenseDimsMap(map, debug = FALSE)
}
\arguments{
\item{map}{- a DimsMap object created using [createSparseDimsMap()]}

\item{debug}{- flag (T/F) to print debugging info}
}
\value{
a DimsMap object, which is atibble (see [tibble::tibble()]) with class "DimsMap"
and attributes
\itemize{
\item{"dmnms" - vector of dimension names}
\item{"dmlvs" - list with non-nested dimension levels, by dimension name}
\item{"dmlns" - vector of dimension lengths, by dimension name}
}
These attributes have the same values as the corresponding ones in \code{map}.
}
\description{
Function to create an index map using all
(non-nested, fully-crossed) dimension levels
}
\details{
The sparse dims map \code{map} is used to identify all levels of
each dimension included in the tibble. These are then used to create a
fully-crossed (dense) dims map (i.e., ignoring any nesting of dimensions).
The first column ("dense_idx") contains the 1-d index associated
with each combination of the full dimension levels.
}
