% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsUtilities.R
\name{createDimsFactors}
\alias{createDimsFactors}
\title{Convert dimension values to factor levels in a dimensions map}
\usage{
createDimsFactors(dfr)
}
\arguments{
\item{dfr}{\itemize{
\item dimensions map dataframe/tibble (e.g. output from \code{\link[=createSparseDimsMap]{createSparseDimsMap()}})
}}
}
\value{
\itemize{
\item dimensions map with dimension values converted to factors
}
}
\description{
Function to convert dimension values to factor levels in a dimensions map.
}
\details{
Dimension columns that are already factors are not converted.
}
