% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFunctionInfo.R
\name{createFunctionInfo}
\alias{createFunctionInfo}
\title{Create a function info object}
\usage{
createFunctionInfo(fcn_id_, fcn_, model_frame, vars_, par_ids_)
}
\arguments{
\item{fcn_id_}{\itemize{
\item label used to identify particular instance of the function
}}

\item{fcn_}{\itemize{
\item the name of a function
}}

\item{model_frame}{\itemize{
\item the model frame which the function applies to
}}

\item{vars}{\itemize{
\item function variable (as character string)
}}

\item{par_ids}{\itemize{
\item character vector with par_ids
}}
}
\value{
a dataframe (at present--this will probably change)
}
\description{
Function to create a function info object.
}
\details{
The name of the function and the function id are appended to the model-frame. The
same function name, \code{fcn}, can be associated with multiple \code{fcn_id}'s, hence the necessity
for supplying a unique function "label".
}
