% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopDynamics.R
\name{createMatrixNextMaturityState}
\alias{createMatrixNextMaturityState}
\title{Create a matrix to advance maturity state to the next level}
\usage{
createMatrixNextMaturityState(dimMPZ)
}
\arguments{
\item{dimMPZ}{\itemize{
\item subset of a SparseDimsMap reflecting maturity states (m), post-molt age (p), and sizes (z) for a single sex
}}
}
\value{
a permutation and aggregation matrix of 1's and 0's to update a vector with maturity from \code{m} to \code{m+1}
}
\description{
Function to create a matrix to advance maturity state from \code{m} to \code{m}+1 for crab.
}
\details{
The resulting matrix can be used to update maturity state for crab that molted.

vp = m \%*\% v;

where \code{m} is the matrix, \code{v} is an abundance vector characterized by m,p,z, and \code{vp} is the resulting
abundance vector with the maturity state updated.
}
