% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopDynamics.R
\name{createMatrixPto0}
\alias{createMatrixPto0}
\title{Create a matrix to return post-molt category from \code{p} to 0 for crab that molted}
\usage{
createMatrixPto0(dimMPZ)
}
\arguments{
\item{dimMPZ}{\itemize{
\item subset of a SparseDimsMap reflecting maturity states (m), post-molt age (p), and sizes (z) for a single sex
}}
}
\value{
a permutation and aggregation matrix of 1's and 0's to update a vector with post-molt age states from \code{p} to \code{0}
}
\description{
Function to create a matrix to sreturn post-molt category from \code{p} to \code{0}for crab that molted.
}
\details{
The resulting matrix can be used to update post-molt age states for crab that molted.

vp = m \%*\% v;

where \code{m} is the matrix, \code{v} is an abundance vector characterized by m,p,z, and \code{vp} is the resulting
abundance vector with the post-molt states updated.
}
