% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopDynamics.R
\name{createMatrixPtoPplus1}
\alias{createMatrixPtoPplus1}
\title{Create a matrix to shift post-molt category from \code{p} to \code{p+1} (or max(\code{p}) as an accumulator state) for crab that didn't molt}
\usage{
createMatrixPtoPplus1(dimMPZ)
}
\arguments{
\item{dimMPZ}{\itemize{
\item subset of a SparseDimsMap reflecting maturity states (m), post-molt age (p), and sizes (z) for a single sex
}}
}
\value{
a permutation and aggregation matrix of 1's and 0's to update a vector with post-molt age states from \code{p} to \code{p}+1 for crab that didn't molt.
}
\description{
Function to create a matrix to shift post-molt category from \code{p} to \code{p+1} (or max(\code{p}) as an accumulator state) for crab that didn't molt.
}
\details{
The resulting matrix can be used to update post-molt age states for crab that did not undergo a molt.

vp = m \%*\% v;

where \code{m} is the matrix, \code{v} is an abundance vector characterized by m,p,z, and \code{vp} is the resulting
abundance vector with the post-molt states updated.
}
