% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Formulas.R
\name{createModelFrame}
\alias{createModelFrame}
\title{Create the model frame from a parameter formula and base dataframe}
\usage{
createModelFrame(formula, data)
}
\arguments{
\item{formula}{\itemize{
\item RHS formula describing model for which to create the model frame
}}

\item{data}{\itemize{
\item base dataframe from which to extract model frame
}}
}
\value{
A data.frame
}
\description{
Function to create the model frame from a parameter formula and base dataframe.
}
\details{
The formula is processed to substitute '+' for RE grouping symbols ('|' and '||')
and drop "specials" (see \link{.valid_covstruct} and \link{.valid_smooths}) from the formula.
Column names for derived columns (e.g., \code{cos(x)}) are a back-quoted version of the deriving formula.
}
