% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParamInfo_FEs.R
\name{createParamInfo_FEs}
\alias{createParamInfo_FEs}
\title{Get parameter information for fixed effects}
\usage{
createParamInfo_FEs(
  formula,
  model_frame,
  contrasts = NULL,
  drop.unused.levels = FALSE,
  offset = NULL,
  sparse = TRUE,
  debug = FALSE
)
}
\arguments{
\item{formula}{\itemize{
\item a formula from which to extract parameter info for fixed effects
}}

\item{model_frame}{\itemize{
\item the model frame
}}

\item{contrasts}{\itemize{
\item a list of contrasts
}}

\item{drop.unused.levels}{\itemize{
\item flag (T/F) to drop unused levels
}}

\item{offset}{\itemize{
\item the name of a column in \code{fr} to be used as a fixed offset
}}

\item{sparse}{\itemize{
\item flag (T/F) to return sparse model matrix?
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging info
}}
}
\value{
a list composed of
\itemize{
\item{formula}{input formula}
\item{model_frame - the model frame}
\item{fe_form}{the fixed effects formula}
\item{contrasts}{the list of specified contrasts (if any)}
\item{offset}{offset vector, or vector of zeros if offset not specified}
\item{vars}{fixed effects variables names}
\item{fe_pvs_lbls}{labels for fixed effects parameter values}
\item{X}{design matrix for fixed effects}
\item{model_matrix}{model matrix associated with \code{X}}
}
}
\description{
Function to get parameter information for fixed effects.
}
