% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParamInfo_REs.R
\name{createParamInfo_REs}
\alias{createParamInfo_REs}
\title{Create the parameters information list for a set of RE terms}
\usage{
createParamInfo_REs(
  formula,
  model_frame,
  contrasts = NULL,
  drop.unused.levels = TRUE,
  sparse = TRUE
)
}
\arguments{
\item{formula}{\itemize{
\item formula, possibly containing random effects
}}

\item{model_frame}{\itemize{
\item full model frame
}}

\item{contrasts}{a list of contrasts}

\item{drop.unused.levels}{\itemize{
\item flag (T/F) to drop unused levels
}}

\item{sparse}{\itemize{
\item flag (T/F) to return sparse model matrix?
}}
}
\value{
a list, or NULL (see details)
}
\description{
Function to create the parameters information list for a set of RE terms.
}
\details{
If \code{formula} does not contain any RE terms, NULL is returned. Otherwise, a list is returned with the following elements:
\itemize{
\item{re_form}{RE formula}
\item{reTerms}{RE terms as character strings}
\item{model_matrix}{model matrix for the REs}
\item{Z}{design matrix for the REs}
\item{re_pvs_idx}{a named vector with a cumulative index for the RE parameter values}
\item{re_pvs_lbls}{a list if labels for the RE parameter values}
\item{lstQtp}{a list of term-associated precision matrix templates (see \code{\link[=mkReTermInfoList]{mkReTermInfoList()}})}
\item{lstQfill}{a list of term-associated functions to fill the associated precision matrix (see \code{\link[=mkReTermInfoList]{mkReTermInfoList()}})}
}
}
