% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParamInfo_SMs.R
\name{createParamInfo_SMs}
\alias{createParamInfo_SMs}
\title{Create the parameters information list for a set of smooth terms TODO:under construction!!}
\usage{
createParamInfo_SMs(
  formula,
  model_frame,
  contrasts = NULL,
  drop.unused.levels = TRUE,
  sparse = TRUE
)
}
\arguments{
\item{formula}{\itemize{
\item formula, possibly containing smooth terms
}}

\item{model_frame}{\itemize{
\item full model frame
}}

\item{contrasts}{a list of contrasts}

\item{drop.unused.levels}{\itemize{
\item flag (T/F) to drop unused levels
}}

\item{sparse}{\itemize{
\item flag (T/F) to return sparse model matrix?
}}
}
\value{
a list, or NULL (see details)
}
\description{
Function to create the parameters information list for a set of smooth terms.
}
\details{
If \code{formula} does not contain any RE terms, NULL is returned. Otherwise, a list is returned with the following elements:
\itemize{
\item{smTerms}{smooth terms as character strings}
\item{Z}{design matrix for the smooths}
\item{sm_pvs_idx}{a named vector with a cumulative index for the SM parameter values}
\item{sm_pvs_lbls}{a list of labels for the SM parameter values}
}
}
