% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ParamsInfo.R
\name{createParamNamesForFEs}
\alias{createParamNamesForFEs}
\title{: Create "better" names for fixed effects parameters.}
\usage{
createParamNamesForFEs(fe_form, X, debug = TRUE)
}
\value{
dataframe with original (\code{orig}) and revised (\code{revd}) names.
}
\description{
: Create "better" names for fixed effects parameters.
}
\details{
The original parameter names (given by \code{colnames(X)})for those involving factors are combinations of
the factor names and the factor levels, with interactions indicated by a colon (":").
The revised names for these drop the factor names and split interactions with a "_").
}
