% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParamsInfo_FEs.R
\name{createParamsInfo_FEs}
\alias{createParamsInfo_FEs}
\title{Get parameter information for fixed effects}
\usage{
createParamsInfo_FEs(
  formula,
  model_frame,
  contrasts = NULL,
  offset = NULL,
  sparse = TRUE
)
}
\arguments{
\item{formula}{\itemize{
\item a formula from which to extract parameter info for fixed effects
}}

\item{model_frame}{\itemize{
\item the model frame
}}

\item{contrasts}{\itemize{
\item a list of contrasts (see ?glmmTMB)
}}

\item{offset}{\itemize{
\item the name of a column in \code{fr} to be used as a fixed offset
}}

\item{sparse}{\itemize{
\item (logical) return sparse model matrix?
}}
}
\value{
a list composed of
\itemize{
\item{p - a suitable parameter vector for the fixed effects}
\item{X}{design matrix for fixed effects}
\item{fe_form}{the fixed effects formula}
\item{contrasts}{the list of specified contrasts (if any)}
\item{offset}{offset vector, or vector of zeros if offset not specified}
\item{formula}{input formula}
\item{model_frame - the model frame}
\item{model_matrix - combined model frame and X}
}
}
\description{
Function to get parameter information for fixed effects.
}
