% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Parameters.R
\name{createParamsMap}
\alias{createParamsMap}
\title{Create TMB \code{map} for parameters from a parameter values dataframe}
\usage{
createParamsMap(dfr)
}
\description{
Create TMB \code{map} for parameters from a parameter values dataframe
}
