% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Formulas.R
\name{createParamValLabels}
\alias{createParamValLabels}
\title{: Create "better" labels for parameter values.}
\usage{
createParamValLabels(form, colnms, debug = TRUE)
}
\arguments{
\item{colnms}{\itemize{
\item column names from the associated design matrix
}}

\item{fe_form}{\itemize{
\item formula or term from formula
}}
}
\value{
dataframe with original (\code{orig}) and revised (\code{revd}) names.
}
\description{
: Create "better" labels for parameter values.
}
\details{
The original parameter names (\code{colnms}) from the design matrix for those columns involving factors are a combination of
the factor name and a factor level, with interactions indicated by a colon (":").
The revised labels for these add a "@" between a factor name and its level; interactions are still indicated with a ":".
}
