% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPostMoltAgeTMs.R
\name{createPostMoltAgeTMs}
\alias{createPostMoltAgeTMs}
\title{Create transition matrices between post-molt age categories}
\usage{
createPostMoltAgeTMs(dmsC)
}
\arguments{
\item{dmsC}{\itemize{
\item the model dimensions dataframe for population classes (i.e., dims$dmsC when setting up model)
}}
}
\value{
list with two transition matrices (\code{tmPMAD} and \code{tmPMAI}) between post-molt age categories (see Details below)
}
\description{
Function to create two transition matrices between post-molt age categories
}
\details{
This function creates transition matrices for crab classes that will (\code{tmPMAD}) or will not
(\code{tmPMAI}) undergo molting during a growth season. These matrices contain only constants (0's or 1's)
and thus can be created outside the objective function and should be included as part of the \code{inputs} list.

\code{tmPMAI} increments the post-molt age category for non-molting crab by one category (e.g., "new-shell" to "old_shell").
If nN is a vector of the numbers of crabs, by model population class, at the start of a growth season that will not molt,
then

$$nN^+ = tmPMAI \%*\% nN$$

"moves" crab from post-molt age class \code{p} to \code{p+1} (or leaves them in the last class, which acts as an aggregator).

\code{tmPMAD} decrements the post-molt age category for molting crab to the first category (i.e., post-molt age 0-e.g., "new_shell").
If mN is a vector of the numbers of crabs, by model population class, at the start of a growth season that will molt,
then

$$mN^+ = tmPMAD \%*\% mN$$

"moves" crab from post-molt age class \code{p} to the first class (post-molt age 0--e.g., "new_shell").
}
