% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRTMB_Params.R
\name{createRTMB_Params}
\alias{createRTMB_Params}
\title{Create the parameters information list for RTMB}
\usage{
createRTMB_Params(lst)
}
\arguments{
\item{-}{list from \code{\link[=readFile_CTL]{readFile_CTL()}}}
}
\value{
list
}
\description{
Function to create the parameters information list for RTMB.
}
\details{
returned list has elements
\itemize{
\item{lstRTMB_Params - list with initial parameter values}
\item{lstRTMB_TForms - list with parameter value-level transform function names}
\item{lstRTMB_LBs - list with parameter value lower bounds}
\item{lstRTMB_UBs - list with parameter value upper bounds}
\item{lstRTMB_Phases - list with parameter value estimation phases}
\item{lstRTMB_Mirrors - list with parameter values mirroring}
\item{lstRTMB_Xs - list with (sparse) design matrices}
\item{lstRTMB_MMs - list with model matrices (as tibbles)}
\item{lstRTMB_Map - RTMB "map" list}
\item{dfrMMs - tibble with model matrix for all parameters expanded to the largest extent present}
\item{vLinks - vector with parameter-level link function names}
}
}
