% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsFunctions.R
\name{createSparseDimsMap}
\alias{createSparseDimsMap}
\title{Create a sparse index map (a tibble) for a set of (possibly nested) dimension levels}
\usage{
createSparseDimsMap(..., convertCharsToFactors = FALSE, debug = FALSE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> name/value dimension pairs}

\item{convertCharsToFactors}{\itemize{
\item option to convert character dims to factors (default=FALSE)
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging info
}}
}
\value{
a \code{DimsMap} object, which is a tibble (see \code{\link[tibble:tibble]{tibble::tibble()}}) with
class "DimsMap" and attributes
\itemize{
\item{"dmtyp" - 'sparse': indicates DimsMap type}
\item{"dmnms" - vector of dimension names}
\item{"dmlvs" - list with non-nested dimension levels, by dimension name}
\item{"dmlns" - vector of dimension lengths, by dimension name}
}
}
\description{
Function to create a sparse 1-d index (a tibble) for a set of
(possibly nested) dimension levels. The index value can then be used to identify
the associated level for each dimension.
}
\details{
Each name in
\code{...} should be the name of the terminal dimension in the associated value (a
possibly-nested dimension list or a vector of levels). For each name, the
associated value is transformed to a tibble using \code{\link[=traverseDimsList]{traverseDimsList()}}, then all the
tibbles are crossed to create a tibble with only the desired combinations of
actual dimension levels. The first column ("i") contains the 1-d index associated
with each combination of actual dimension levels.

Alternatively, the \code{...} can be given by splicing a list:\cr
dimsMap = createSparseDimsMap(!!!lst)\cr
where \code{lst} is a named list of the inputs to the function.
This requires the \code{rlang} package be loaded.
}
\examples{
#--create some example dimensions
require(tibble);
#----years
vYs = 2001:2005;                  attr(vYs,"dmnms")<-"y";
#----seasons
vCs = c("fall","spring");         attr(vCs,"dmnms")<-"c";
#----regions
vRs = c("All EBS");               attr(vRs,"dmnms")<-"r";
#----sexes
vXs = c("MALE","FEMALE");         attr(vXs,"dmnms")<-"x";
#----maturity state nested in sex (working outward)
vMs = list(  MALE=c("IMMATURE","MATURE"),
           FEMALE=c("IMMATURE","MATURE"));
attr(vMs,"dmnms")<-c("x","m");
#----shell condition nested in maturity state/sex (working outward)
vSs = list(MALE=  list(IMMATURE=c("NEW SHELL"),
                        MATURE=c("NEW SHELL","OLD SHELL")),
          FEMALE=list(IMMATURE=c("NEW SHELL"),
                        MATURE=c("NEW SHELL","OLD SHELL")));
attr(vSs,"dmnms")<-c("x","m","s");
#----ages nested in shell condition/maturity state/sex (working outward)
vAs = list(MALE=  list(IMMATURE=list(`NEW SHELL`="NO AGEs"),
                         MATURE=list(`NEW SHELL`=c("PTM 0"),
                                     `OLD SHELL`=c("PTM 1","PTM 2"))),
          FEMALE=list(IMMATURE=list(`NEW SHELL`="NO AGEs"),
                        MATURE=list(`NEW SHELL`=c("PTM 0"),
                                    `OLD SHELL`=c("PTM 1","PTM 2"))));
attr(vAs,"dmnms")<-c("x","m","s","a");
#----sizes nested in age/shell condition/maturity state/sex (working outward)
vZs = list( MALE=  list(IMMATURE=list(`NEW SHELL`=list(`NO AGEs`=seq(40,60,5))),
                          MATURE=list(`NEW SHELL`=list(`PTM 0`=seq(80,120,10)),
                                      `OLD SHELL`=list(`PTM 1`=seq(80,120,10),
                                                       `PTM 2`=seq(80,120,10)))),
          FEMALE=list(IMMATURE=list(`NEW SHELL`=list(`NO AGEs`=seq(25,50,5))),
                        MATURE=list(`NEW SHELL`=list(`PTM 0`=seq(50,100,10)),
                                    `OLD SHELL`=list(`PTM 1`=seq(50,100,10),
                                                     `PTM 2`=seq(50,100,10)))));
attr(vZs,"dmnms")<-c("x","m","s","a","z");

#--create maps from the 1-d index to the dimension levels for various combinations
#----of the example dimensions.
require(rtmbGMACS);
dfrSparse = createSparseDimsMap(y=vYs);
dfrSparse = createSparseDimsMap(y=vYs,c=vCs);
dfrSparse = createSparseDimsMap(y=vYs,z=vZs);
dfrSparse = createSparseDimsMap(a=vAs);
dfrSparse = createSparseDimsMap(z=vZs);
dfrSparse = createSparseDimsMap(a=vAs,z=vZs);
dfrSparse = createSparseDimsMap(y=vYs,c=vCs,r=vRs,x=vXs,m=vMs,s=vSs,a=vAs,z=vZs);
}
