% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{dbllogistic}
\alias{dbllogistic}
\title{Calculate a double logistic selectivity curve}
\usage{
dbllogistic(z, params, refZ = 0, debug = FALSE)

dbllogistic(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivity curve
}}

\item{params}{\itemize{
\item 4-element vector of selectivity function parameters
}}

\item{refZ}{\itemize{
\item reference size
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}

\item{pAscZ50}{\itemize{
\item ascending limb size at which selectivity = 0.5 (logit-scale mean)
}}

\item{pAscSlp}{\itemize{
\item ascending limb slope at 50\\%-selected
}}

\item{pDscZ50}{\itemize{
\item descending limb size at which selectivity = 0.5 (logit-scale mean)
}}

\item{pDscSlp}{\itemize{
\item descending limb slope at 50\\%-selected
}}

\item{pRefZ}{\itemize{
\item reference size
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
vector with selectivity values at the elements of z

named vector with selectivity values at the elements of z
}
\description{
Function to calculate a double logistic selectivity curve.

Function to calculate a logistic selectivity curve.
}
\details{
The parameter pRefZ is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.

If \code{pRefZ}>0, \code{pRefZ}=fully-selected size.
if \code{pRefZ}<0, function is normalized to max.
if \code{pRefZ}=0, no re-scaling is done.


}
