% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{dbllogistic5095}
\alias{dbllogistic5095}
\title{Calculate a logistic function parameterized by z50 and D95=z95-z50 for ascending/descending limbs}
\usage{
dbllogistic5095(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivities
}}

\item{params}{\itemize{
\item 4-element parameter vector
}}

\item{refZ}{\itemize{
\item reference size
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate a logistic function parameterized by z50, z95 on ascending/descending limbs.
}
\details{

}
