% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{dblnormal4a}
\alias{dblnormal4a}
\title{Calculates a 4-parameter normal selectivity curve}
\usage{
dblnormal4a(z, params, refZ, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute function values}

\item{params}{- vector of function parameters}

\item{refZ}{- max possible size (e.g., max(z))}
}
\value{
- named vector of selectivity values
}
\description{
Function to calculate a 4-parameter normal selectivity curve.
}
\details{
Calculates 4-parameter normal function parameterized by
\itemize{
     \item params[1]: size at which ascending limb reaches 1
     \item params[2]: width of ascending limb
     \item params[3]: scaled increment to params[1] at which descending limb departs from 1
     \item params[4]: width of descending limb
}
}
