% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineYearBlocks.R
\name{defineYearBlocks}
\alias{defineYearBlocks}
\title{Define year blocks}
\usage{
defineYearBlocks(obj, dimsMdl = NULL, by = "y")
}
\arguments{
\item{obj}{\itemize{
\item named list of year blocks or data.frame
}}

\item{dimsMdl}{\itemize{
\item optional model dimensions tibble
}}

\item{by}{\itemize{
\item character vector of model dims used with \code{\link[dplyr:mutate-joins]{dplyr::inner_join()}} to join \code{dimMdl} to \code{obj} if \code{obj} is a data.frame
}}
}
\value{
dataframe with columns "yblk","y"
}
\description{
Function to create a dataframe defining a set of year blocks.
}
\details{
Creates a dataframe defining a set of year blocks. If \code{dimsMdl} is
not null, \code{dimsMdl} is inner-joined to this dataframe, mapping the year blocks to
the model dimensions.
}
\examples{
tbAll = list(all=1949:2022) |> defineYearBlocks();
tbRec = list(spinup=1949:1974,normal=1975:2022) |> defineYearBlocks();
tbNM = list(normal=c(1949:1979,1985:2022),
           heightened=1980:1984) |> defineYearBlocks();
tbTCFRetM<-list(early=c(1949:1983,1987:1989),
             mid  =c(1990:1996),
             PR   =c(2005:2009,2013:2015,2017:2018,2020:2022)) |> defineYearBlocks();
tbAll = tibble::tribble(~tb,~y,~m,
                        "all",1949:2022,"imm") |> defineYearBlocks();
}
