% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopDynamics.R
\name{doGrowth}
\alias{doGrowth}
\title{Run population dynamics}
\usage{
doGrowth(n_mpz, lst)
}
\arguments{
\item{n_mpz}{\itemize{
\item vector of abundance classified by m (maturity state), p (post-molt age), z (size)
}}

\item{lst}{\itemize{
\item growth info for crab in y,s,r,x (i.e., lstGrw[\link{y}][\link{s}][\link{r}][\link{x}])
}}

\item{dims}{\itemize{
\item model dimensions list
}}

\item{options}{\itemize{
\item model options list
}}

\item{params}{\itemize{
\item model parameters list
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
list with results from running the population dynamics model

list with element vector n_mpz after growth (molt/growth/maturation)
}
\description{
Run population dynamics for model time period.
}
\details{
At each season \code{s}, the population in year \code{y}is advanced using

np = T \%\emph{\% exp(-Z}dt) \%\emph{\% G \%}\% n + R if \code{dt} > 0 otherwise

np = T \%\emph{\% exp(-F)    \%}\% G \%*\% n + R if \code{dt} = 0

where \code{dt} is season length (as a fraction of a year),
\code{n} is the population abundance vector at the start of the season,
\code{np} is the population abundance vector at the end of the season,
\code{T} is a matrix representing movement among regions,
\code{Z} = \code{M} + \code{F} is a diagonal matrix representing total mortality \emph{rates},
where \code{M} represents natural mortality rates and \code{F} represents
total fishing mortality \emph{rates} in the case where the season is \emph{not} instantaneous,
\code{G} is a matrix representing growth, which encompasses the probability of molting,
the probability of maturing, and actual growth (conditioned on having molted),
and R is recruitment. Movement, growth, and recruitment are treated as instantaneous
processes occurring at the beginning of a season, after which mortality due to
natural processes and/or fishing occurs (depending on season length), and then
recruitment occurs at the end of the season.

Fishery capture processes are modeled by individual fleets as

a_f = c_f/(M+F) \%\emph{\% (1-exp(-Z}dt)) \%*\% n if \code{dt} >0

a_f = c_f/F \%\emph{\% (1-exp(-F)) \%}\% n if \code{dt} = 0

r_f = rho_f * a_f;

d_f = (1-rho_f) * af

m_f = r_f + h_f * d_f

where \code{a_f} is the abundance captured by fleet f,
\code{c_f} is the capture rate of fleet f,
\code{r_f} is the abundance retained by fleet f,
\code{rho_f} is the fraction retained by flet f,
\code{d_f} is the discard abundance (bycatch) by fleet f,
\code{m_f} is the abundance killed (fishing mortality) by fleet f,
and \code{h_f} is the handling mortality on discards in fleet f

"Survey" capture processes are modeled as

c_v = s_v \%*\% n or

c_v = s_v \%*\% np

depending on whether the "survey" occurs at the beginning or end of the season,
{TODO: how to handle CPUE indices from a fishery fleet when season is not instantaneous?}
where \code{c_v} is the abundance captured in the survey (scaled to the population)
and \code{s_v} is the catchability of the survey (which includes fully-selected catchability,
size-specific selectivity, and availability).

\code{lst} is a list in which the following elements may (or have to be) present:

\itemize{
\item{maturity - string indicating whether molt to maturity depends on "pre-molt size" or "post-molt size" \link{required}}
\item{mPtoP1 - dgCMatrix object to increment post-molt age by 1 for non-molting crab \link{required}}
\item{mPto0 - dgCMatrix object to decrement post-molt age to 0 for molting crab \link{required}}
\item{mImmToMat - dgCMatrix object to increment maturity state from immature to mature \link{required}}
\item{mT - dgCMatrix object representing combined molting/growth transitions \link{optional}}
\item{mPrGr - dgCMatrix object representing size transition matrix for molting crab when maturity is determined by post-molt size}
\item{mPrGrImm - dgCMatrix object representing size transition matrix for crab not undergoing the molt to maturity}
\item{mPrGrMat - dgCMatrix object representing size transition matrix for crab undergoing the molt to maturity}
\item{vPrMat - vector representing probability of molting to maturity}
\item{mPto0 - dgCMatrix object representing combined molting/growth transitions \link{optional}}
}

The value of \code{maturity} determines the order in which the probability of maturing is applied, either before growth occurs
("pre-molt size") or after growth occurs ("post-molt size"). If \code{mT} is present, then it is used to determine molting and growth under the assumption that the diagonal elements
represent the probability of not molting (and thus not growing or maturing) while the remaining upper triangle elements
represent the combined probability of molting and growing. If \code{mT} is not present, then \code{mPrGr} (if \code{maturity}=="post-molt size')
or both \code{mPrGrImm }and \code{mPrGrMat} (if \code{maturity}=="pre-molt size") must be present (the latter allows growth to differ between
crab that grow without maturing and those that undergo the molt to maturity).
}
