% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscConstants_Enums.R
\docType{data}
\name{.valid_covstruct}
\alias{.valid_covstruct}
\title{Named character vector of valid covariance structure names}
\format{
An object of class \code{character} of length 16.
}
\usage{
.valid_covstruct
}
\description{
Named character vector of valid covariance structure names
}
\details{
None.
}
\keyword{datasets}
