% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscConstants_Enums.R
\docType{data}
\name{.valid_family}
\alias{.valid_family}
\title{Named character vector of valid statistical family names}
\format{
An object of class \code{character} of length 22.
}
\usage{
.valid_family
}
\description{
Named character vector of valid statistical family names
}
\details{
None. TODO: convert character values to functions?
}
\keyword{datasets}
