% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\docType{data}
\name{.valid_invtforms}
\alias{.valid_invtforms}
\title{list of valid inverse transforms}
\format{
An object of class \code{list} of length 14.
}
\usage{
.valid_invtforms
}
\description{
list of valid inverse transforms
}
\details{
Given the function name \code{fcn} as a string, its inverse transform function is obtained by
\code{.valid_invtforms[fcn]}.
}
\keyword{datasets}
