% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscConstants_Enums.R
\docType{data}
\name{.valid_link}
\alias{.valid_link}
\title{Named character vector of valid statistical family link function names}
\format{
An object of class \code{character} of length 8.
}
\usage{
.valid_link
}
\description{
Named character vector of valid statistical family link function names
}
\details{
None. TODO: convert character values to functions?
}
\keyword{datasets}
