% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscConstants_Enums.R
\docType{data}
\name{.valid_simcode}
\alias{.valid_simcode}
\title{Named character vector of valid simulation options}
\format{
An object of class \code{character} of length 3.
}
\usage{
.valid_simcode
}
\description{
Named character vector of valid simulation options
}
\details{
None.
}
\keyword{datasets}
