% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscConstants_Enums.R
\docType{data}
\name{.valid_smooths}
\alias{.valid_smooths}
\title{Character vector of valid mgcv smooth function names}
\format{
An object of class \code{character} of length 2.
}
\usage{
.valid_smooths
}
\description{
Character vector of valid mgcv smooth function names
}
\details{
None.
}
\keyword{datasets}
