% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{elem_prod}
\alias{elem_prod}
\title{Calculate the element-wise product of two variables}
\usage{
elem_prod(x, y)
}
\arguments{
\item{x}{\itemize{
\item variable (vector, matrix, etc.)
}}

\item{y}{\itemize{
\item variable conformable to x
}}
}
\value{
(unnamed) object x*y
}
\description{
ADMB-equivalent to calculate the element-wise product of two variables.
}
\details{
Substitution for ADMB \code{elem_prod} function.
}
