% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsFunctions.R
\name{expand.DimsMap}
\alias{expand.DimsMap}
\title{Expand a dimensions map by additional dimensions}
\usage{
\method{expand}{DimsMap}(m1, m2)
}
\arguments{
\item{m1}{- dimensions map to expand}

\item{m2}{- dimensions map to expand by}
}
\value{
a dimensions map with expanded dimensions (a tibble with class "DimsMap")
}
\description{
Function to expand a dimensions map by additional dimensions.
}
\details{
This function uses [dplyr::cross_join()] to cross join the two dimension maps to create
a map with expanded dimensions. The two maps should not have any dimensions in common.

This is *not* a S3 generic method, so you must use `expand.DimsMap(m1,m2)`.

If both maps are "dense", the result is a "dense" map.
If either is "sparse", the result is a "sparse" map.
The other map attributes are appropriately expanded.

IMPORTANT NOTE: the (sparse or dense) index of `m1` is replicated `n` times, where `n` is
the number of rows in `m2`. It is the USER'S RESPONSIBILITY to re-number the index column as
appropriate for further use.
}
