% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged_array.R
\name{expand.ragged_array}
\alias{expand.ragged_array}
\title{Expand a ragged_array by a dimensions map}
\usage{
\method{expand}{ragged_array}(x, y)
}
\arguments{
\item{x}{\itemize{
\item the ragged array to convert
}}

\item{y}{\itemize{
\item the dimensions map used to expand \code{x}
}}
}
\value{
a ragged_array
}
\description{
S3 method to expand a ragged_array using a dimensions map.
}
\details{
\code{x}'s dimension map (it's "dfrDims" attribute) is expanded by \code{y} using
expand.DimsMap, then the vector underlying \code{x} is suitably replicated across
the dimensions in \code{y}.

This is an S3 generic method, so just use \code{expand(x,y)}.
}
