% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{expand}
\alias{expand}
\title{Generic S3 method to expand an object by another}
\usage{
expand(x, y)
}
\arguments{
\item{x}{- object to expand}

\item{y}{- object to expand by}
}
\value{
object of `class(x)` expanded by object y
}
\description{
Generic S3 method to expand an object by another.
}
