% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ModelFrames.R
\name{expandFunctionsToModelFrame}
\alias{expandFunctionsToModelFrame}
\title{Expand a functions dataframe to a model frame}
\usage{
expandFunctionsToModelFrame(dfr)
}
\arguments{
\item{dfr}{\itemize{
\item the functions dataframe
}}
}
\value{
a \code{DimsMap} model frame for the functions
}
\description{
Function to expand a functions dataframe to a model frame.
}
