% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Parameters.R
\name{expandParamDataframe_columns}
\alias{expandParamDataframe_columns}
\title{Expand parameter values dataframe for missing columns}
\usage{
expandParamDataframe_columns(dfr)
}
\description{
Expand parameter values dataframe for missing columns
}
