% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInfo_FisheriesRates.R
\name{extractInfo_FisheriesRates}
\alias{extractInfo_FisheriesRates}
\title{Extract fisheries rates information from info list}
\usage{
extractInfo_FisheriesRates(lst, dims = NULL, verbose = TRUE)
}
\arguments{
\item{lst}{\itemize{
\item parameter info list from \code{\link[=readInfo_FisheriesRates]{readInfo_FisheriesRates()}}
}}

\item{dims}{\itemize{
\item \code{dims} list from \code{setupModelDims}
}}
}
\value{
a list (see details)
}
\description{
Function to extract fisheries rates information from info list.
}
\details{
The output list has elements
\itemize{
}
}
