% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInfoType1.R
\name{extractInfoType1}
\alias{extractInfoType1}
\title{Extract function-type parameter info from a list}
\usage{
extractInfoType1(
  lst,
  dms = NULL,
  process_type = "",
  xtra_cols = character(0),
  verbose = TRUE
)
}
\arguments{
\item{lst}{\itemize{
\item function-type parameter info list from a \code{readParamInfo_}
}}

\item{dms}{\itemize{
\item appropriate \code{dms...} DimsMap from \code{setupModelDims}
}}

\item{process_type}{\itemize{
\item string describing process (for verbose output)
}}

\item{xtra_cols}{\itemize{
\item character vector of "extra" columns in the "function" section
}}

\item{verbose}{\itemize{
\item flag (TRUE/FALSE) to print diagnostic info
}}
}
\value{
a list (see details)
}
\description{
Function to extract function-type parameter info from a list.
}
\details{
The output list has elements

\itemize{
\item{option - format option}
\item{Fcns - list with function-level info}
\item{dfrCmbs - dataframe with all parameter combinations in vertical format}
\item{dfrUniqCmbs - dataframe with unique parameter combinations in vertical format}
}
}
