% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Txt.R
\name{extractLines}
\alias{extractLines}
\title{Extract character vector from section of longer vector}
\usage{
extractLines(txt, start, end)
}
\arguments{
\item{txt}{\itemize{
\item long character vector
}}

\item{start}{\itemize{
\item string to regard as marking the start of section (not included in the section)
}}

\item{end}{\itemize{
\item string to regard as marking the end of the section (not included in the section)
}}
}
\value{
a character vector of the "enclosed" section
}
\description{
Function to extract character vector from section of longer vector.
}
\details{
The text lines at which the \code{start} and \code{end} are found are not included in
the "section" returned by this function.
}
