% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParameters_Allometry.R
\name{extractParameters_Allometry}
\alias{extractParameters_Allometry}
\title{Extract allometry parameters from parameter info list}
\usage{
extractParameters_Allometry(lst, dims = NULL, verbose = TRUE)
}
\arguments{
\item{lst}{- parameter info list from [readParamInfo_Allometry()]}

\item{dims}{- dataframe with dimensions (columns) to expand}
}
\value{
a list
}
\description{
Function to extract allometry parameters from parameter info list.
}
\details{
TODO: fill in.
}
