% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParamInfo_Allometry.R
\name{extractParamInfo_Allometry}
\alias{extractParamInfo_Allometry}
\title{Extract allometry parameters from parameter info list}
\usage{
extractParamInfo_Allometry(lst, dims = NULL, verbose = TRUE)
}
\arguments{
\item{lst}{- parameter info list from [readParamInfo_Allometry()]}

\item{dims}{- `dims` list from `setupModelDims`}
}
\value{
a list (see details)
}
\description{
Function to extract allometry parameters from parameter info list.
}
\details{
The format of the returned list depends on the `option` specified in `lst`.
If `lst$option` is "data" (i.e., the input parameter values are fixed values of weight-at-size
specified at least for every population category), the output list has elements
\itemize{
\item{option - format option}
\item{params - vector of weight-at-size values}
\item{map - vector for the RTMB `map` list indicating the parameters are fixed; i.e. not to be estimated}
\item{dfrIdx2Pars - dataframe with columns "pidx", dimension names (y,s,r,x,m,p,z), and parameter specs IV,LB,UB,phz,PriorType,Pr1,Pr2
indicating the un-expanded mapping from `pidx` to the parameter input values ("IV"). `pidx` is the index into the parameter vector.
The dimension columns are incidental and reflect the specification from the input text file.}
\item{dfrDims2Pars - dataframe with columns "pidx" and the dimension names (y,s,r,x,m,p,z).
The dimension levels are fully expanded such that each row corresponds to a single "multidimensional" level.
The value of `pidx` for each row indicates the index of the associated weight-at-size value in the parameter vector.}
}

If `lst$option` is "function", the output list has elements
\itemize{
}
}
