% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParamInfoFunctionType1.R
\name{extractParamInfoFunctionType1}
\alias{extractParamInfoFunctionType1}
\title{Extract function-type parameter info from a list}
\usage{
extractParamInfoFunctionType1(
  lst,
  dms = NULL,
  process_type = "",
  xtra_cols = character(0),
  verbose = TRUE
)
}
\arguments{
\item{lst}{\itemize{
\item function-type parameter info list from a \code{readParamInfo_}
}}

\item{dms}{\itemize{
\item appropriate \code{dms...} DimsMap from \code{setupModelDims}
}}

\item{process_type}{\itemize{
\item string describing process (for verbose output)
}}

\item{xtra_cols}{\itemize{
\item character vector of "extra" columns in the "function" section
}}

\item{verbose}{\itemize{
\item flag (TRUE/FALSE) to print diagnostic info
}}
}
\value{
a list (see details)
}
\description{
Function to extract function-type parameter info from a list.
}
\details{
The output list has elements

\itemize{
\item{option - format option}
\item{Fcns - list with function-level info}
\item{MPs - list with main parameters-level info}
\item{OPs - list with offset parameters-level info}
\item{DPs - list with devs (fixed) parameter vectors-level info}
\item{REs - list with random effects-level info}
\item{PECs - list with main parameter-level info}
\item{FECs - list with main parameter-level info}
\item{dfrFPs - dataframe with functional priors info}
\item{dfrCmbs - dataframe with all parameter combinations in vertical format}
\item{dfrUniqCmbs - dataframe with unique parameter combinations in vertical format}
\item{dfrHCs - dataframe with all parameter combinations in "horizontal" format}
\item{dfrUHCs - dataframe with unique parameter combinations in "horizontal" format}
}
}
