% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParamInfoFunctionType2.R
\name{extractParamInfoFunctionType2}
\alias{extractParamInfoFunctionType2}
\title{Extract function-type parameter info for processes defined by transition matrices from a list}
\usage{
extractParamInfoFunctionType2(
  lst,
  dms = NULL,
  process_type = "",
  tmCats = NULL,
  verbose = TRUE
)
}
\arguments{
\item{lst}{\itemize{
\item function-type parameter info list from a \code{readParamInfo_}
}}

\item{dms}{\itemize{
\item appropriate \code{dms...} DimsMap from \code{setupModelDims}
}}
}
\value{
a list (see details)
}
\description{
Function to extract function-type parameter info for processes defined by transition matrices from a list.
}
\details{
The output list has elements

\itemize{
\item{option - format option}
\item{params - vector of weight-at-size values}
}
}
