% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Formulas.R
\name{extractSmoothVars}
\alias{extractSmoothVars}
\title{Extract the variables associated with a 1-d "smooth" function}
\usage{
extractSmoothVars(txt)
}
\arguments{
\item{txt}{\itemize{
\item text string defining the "smooth" function call
}}
}
\value{
character vector with the name of the variables
}
\description{
Extract the variables associated with a 1-d "smooth" function

Function to extract the variables associated with a 1-d "smooth" function.
}
\details{
The function text must have the form "smooth(...,k=...)", where "smooth" is the
\code{mgcv} smooth function (\code{s} or \code{ti}); the first "..." indicates the comma-delimited variables for the smooth,
where the first variable is numeric and any other (optional) variable must be a factor; and the second ...
is the number of knots followed by any other (comma-delimited) inputs to the smooth function.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code: smooth specification with "by" variable
txt = "s(x,k=10,by='r')"; #--x is numeric
res = extractSmoothVar(txt); # res == "x"
# example code: factor smooths specification
txt = "s(x,y,k=10,bs='fs',by='r')"; #--x is numeric, y is a factor
res = extractSmoothVar(txt); # res == c("x","y")
\dontshow{\}) # examplesIf}
}
