% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Txt.R
\name{extractTextValue}
\alias{extractTextValue}
\title{Extract a (character) value from a line of text}
\usage{
extractTextValue(txt, comment = "#")
}
\arguments{
\item{txt}{- input text string (1-element character vector)}

\item{comment}{- string used to split text into value and comments (default="#")}
}
\value{
a 1-element character vector
}
\description{
Function to extract a (character) value from a line of text.
}
\details{
The 1-line input text is split into a character vector using [stringr::str_split_1()] and
the first element of the result is retained. Whitespace preceding and following the retained
text is also removed (but "interior" whitespace is retained).
}
\examples{
# example code
txt = "'hows this'  #--input text of interest";
extractTextValue(txt)

}
