% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{factorize}
\alias{factorize}
\title{Convert columns in a model frame to factors if included in a formula}
\usage{
factorize(frmla, model_frame, char.only = TRUE)
}
\arguments{
\item{frmla}{\itemize{
\item formula or term
}}

\item{model_frame}{\itemize{
\item the dataframe to convert
}}

\item{char.only}{\itemize{
\item flag to convert only character columns to factors
}}
}
\description{
Function to convert columns in a dataframe to factors if included in a formula.
}
\details{
Copied from \code{reformulas:::factorize()} (char.only option changed to TRUE).
}
