% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{fillQ.ar1}
\alias{fillQ.ar1}
\title{Calculate the values to fill a precision matrix (Q) with an AR1 structure
across parameters within a group.}
\usage{
fillQ.ar1(covpars, ng, idx, sequential = TRUE)
}
\arguments{
\item{covpars}{\itemize{
\item vector of input covariance parameters
}}

\item{ng}{\itemize{
\item number of groups (levels) in grouping factor
}}

\item{idx}{\itemize{
\item index vector mapping \code{covpars} values to Q
}}
}
\value{
vector of values for the "x" slot in Q
}
\description{
Calculate the values to fill a precision matrix (Q) with an AR1 structure
across parameters within a group.
}
\details{
The values in \code{covpars} are standard deviations (\code{sigma}s) on the ln-scale and
correlation parameters on the symmetric logit scale (-Inf,Inf) corresponding to (-1,1) on
the arithmetic scale. These are converted to variances and correlation s before being mapped to Q.

The ordering of values within \code{covapar} is s1,c1,s2,c2,..., where s1 is sigma for the first group,
c is the correlation within the first group.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
ng = 2; #--number of groups (levels) in grouping factor
np = 5; #--number of (RE) parameters
#--number of covariance parameters is 2/group (i.e., sigma, rho)
Qt = buildTemplateQ.ar1(np,ng,sequential=TRUE); #--template for actual Q
idx = Qt@x; #--"map" from covpars to Q
##--R context
###--because `sequential`=TRUE when building `Qt`, the covariance parameters are
###--ordered as c(sigma's, rho's).
covpars = c(log(c(1,0.75))/2,symlogit(c(-0.5,0.5)));
Q = Qt;
Q@x = fillQ.ar1(covpars,ng,idx);
Q
##--RTMB context
covpars = RTMB::AD(c(log(c(1,0.75))/2,symlogit(c(-0.5,0.5))),force=TRUE); #--advector
Q = RTMB::AD(Qt,force=TRUE);   #--adsparse matrix
Q@x = fillQ.ar1(covpars,ng,idx);
as.matrix(Q);
\dontshow{\}) # examplesIf}
}
