% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{fillQ.diag}
\alias{fillQ.diag}
\title{Calculate the values to fill a diagonal precision matrix (\code{Q})}
\usage{
fillQ.diag(covpars, idx)
}
\arguments{
\item{covpars}{\itemize{
\item vector of input covariance parameters
}}

\item{idx}{\itemize{
\item index vector mapping \code{covpars} values to \code{Q}
}}
}
\value{
vector of values (inverse variances) for the "x" slot in \code{Q}
}
\description{
Calculate the values to fill a diagonal precision matrix (\code{Q})
}
\details{
The values in \code{covpars} are standard deviations (\code{sigma}s) on the ln-scale.
These are converted to variances before being mapped to \code{Q}. Covariance parameters are ordered
by RE parameters within each group.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
np = 2; #--number of (RE) parameters/group
ng = 5;  #--number of groups (levels) in grouping factor
np = 2; #--number of (RE) parameters/group
#--number of covariance parameters/group = 2 (1 variance parameter for each RE parameter/group)
Qt = buildTemplateQ.diag(np,ng); #--template for actual Q
idx = Qt@x;                      #--"map" from covpars to Q
##--R context
covpars = log(c(1,0.75))/2; #--std dev parameters on ln-scale (1 per RE parameter)
Q = Qt;
Q@x = fillQ.diag(covpars,idx);
Q
##--RTMB context
covpars = RTMB::AD(log(c(1,0.5))/2,force=TRUE); #--advector
Q = RTMB::AD(Qt,force=TRUE);   #--adsparse matrix
Q@x = fillQ.diag(covpars,idx);
as.matrix(Q);
\dontshow{\}) # examplesIf}
}
