% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{fillQ.us}
\alias{fillQ.us}
\title{Calculate the values to fill a precision matrix (Q) with general structure}
\usage{
fillQ.us(covpars, np, idx, byDiagonal = TRUE)
}
\arguments{
\item{covpars}{\itemize{
\item vector of input covariance parameters
}}

\item{np}{\itemize{
\item number of RE parameters
}}

\item{idx}{\itemize{
\item index vector mapping \code{covpars} values to Q
}}
}
\value{
vector of values for the "x" slot in Q
}
\description{
Calculate the values to fill a precision matrix (Q) with general structure
}
\details{
The values in \code{covpars} are standard deviations (\code{sigma}s) on the ln-scale and
correlation parameters on the symmetric logit scale (-Inf,Inf) corresponding to (-1,1) on
the arithmetic scale. These are converted to variances and correlations before being mapped to Q.

If byDiagonal=TRUE, the ordering of values within \code{covpar} proceeds along the main diagonal
(\code{np} ln-scale std. dev. terms), then along the sub-diagonals (\code{choose(np+1,2)-np} logit-scale
correlation terms).
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
ng = 2; #--number of groups (levels) in grouping factor
np = 5; #--number of (RE) parameters
#--number of covariance parameters is `choose(np+1,2)`
Qt = buildTemplateQ.us(np,ng,byDiagonal=TRUE); #--template for actual Q
idx = Qt@x; #--"map" from covpars to Q
##--R context
###--because `byDagonal`=TRUE when building `Qt`, the covariance parameters are
###--ordered as c(sigma's, rho's).
covpars = c(log(rep_len(c(1,0.75),np))/2,symlogit(rep_len(c(-0.5,0.5),choose(np+1,2)-np)));
Q = Qt;
Q@x = fillQ.us(covpars,np,idx);
Q
##--RTMB context
covpars = RTMB::AD(c(log(c(1,0.75))/2,symlogit(c(-0.5,0.5))),force=TRUE); #--advector
Q = RTMB::AD(Qt,force=TRUE);   #--adsparse matrix
Q@x = fillQ.us(covpars,np,idx);
as.matrix(Q);
\dontshow{\}) # examplesIf}
}
