% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{findKeyword}
\alias{findKeyword}
\title{Find the line index corresponding to a keyword in a vector of text lines}
\usage{
findKeyword(lns, kw)
}
\arguments{
\item{lns}{- vector of text to check}

\item{kw}{- key word to check}
}
\value{
index of `lns` at first occurrence of `kw`, or NULL if not found
}
\description{
Function to find the line index corresponding to a keyword in a vector of text lines.
}
