% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{findNextLine}
\alias{findNextLine}
\title{Find the next uncommented line}
\usage{
findNextLine(lns, iln)
}
\arguments{
\item{lns}{\itemize{
\item character vector to search
}}

\item{iln}{\itemize{
\item index at which to start search
}}
}
\value{
index into \code{lns} at which the next uncommented line occurs
}
\description{
Function to find the next uncommented line in a vector of character strings.
}
