% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_SMs.R
\name{findSmTrmClasses}
\alias{findSmTrmClasses}
\title{Get valid "smooth" class names}
\usage{
findSmTrmClasses()
}
\value{
character vector with names of "smooths" in .valid_smooths
}
\description{
Get valid "smooth" class names
}
