% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{findValidCovStructs}
\alias{findValidCovStructs}
\title{Get valid RE covariance structure names}
\usage{
findValidCovStructs()
}
\value{
character vector with names of RE covariance structures in .valid_covstruct
}
\description{
Get valid RE covariance structure names
}
