% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWatZ.R
\name{getAllometry}
\alias{getAllometry}
\title{Get allometry info as a list}
\usage{
getAllometry(dims, opts, params, verbose = FALSE)
}
\arguments{
\item{dims}{-}

\item{opts}{-}

\item{params}{- (full) parameters list}

\item{verbose}{- flag (T/F) to print diagnostics}
}
\value{
list with elements
\itemize{
  \item{wAtZ - 3-d array (y,s,c) with weight-at-size by year, season, pop class}
}
}
\description{
Function to get allometry info as a list. Used in the objective function.
}
