% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Matrices.R
\name{getBlockDiagInfo}
\alias{getBlockDiagInfo}
\title{Get the starting rows/columns of blocks in a block-diagonal matrix}
\usage{
getBlockDiagInfo(x)
}
\arguments{
\item{lst}{\itemize{
\item a matrix
}}
}
\value{
a vector of starting rows/columns of blocks in the block-diagonal matrix
}
\description{
Function to get the starting rows/columns of blocks in a block-diagonal matrix.
}
\details{
If the matrix is not block-diagonal (i.e., it does not yet have a "block_start" attribute),
then its "block" starts at the first row (using a 0-based index). If the matrix is block-diagonal, its
"block_start" attribute (a 0-based vector of rows at which blocks start) is returned.
}
