% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Matrices.R
\name{getBlockDiagStarts}
\alias{getBlockDiagStarts}
\title{Get the starting rows/columns of blocks for a list of matrices to be sub-matrices in a "super" block-diagonal matrix}
\usage{
getBlockDiagStarts(lst)
}
\arguments{
\item{lst}{\itemize{
\item a list of block-diagonal matrices
}}
}
\value{
a vector of starting rows/columns of blocks in the "super" block-diagonal matrix
}
\description{
Function to get the starting rows/columns of blocks for a list of matrices to be sub-matrices in a "super" block-diagonal matrix.
}
