% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Canonicals.R
\name{getCanonicalFunctionsTbl}
\alias{getCanonicalFunctionsTbl}
\title{Convert CTL file "functions" \code{tibble} to canonical format}
\usage{
getCanonicalFunctionsTbl(tbl)
}
\arguments{
\item{tbl}{\itemize{
\item \code{tibble} to convert
}}
}
\value{
a tibble with all "functions" ("fcn_idx","fcn_id","function","frame","params","vars","description") columns.
}
\description{
Convert CTL file "functions" tibble to canonical format.
}
\details{
Converts a "functions" \code{tibbl}e to canonical format (i.e., al columns). Missing columns
in the original tibble are added with \code{NA}s as values.
}
