% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Canonicals.R
\name{getCanonicalParamEquationsTbl}
\alias{getCanonicalParamEquationsTbl}
\title{Convert CTL file "params equations" \code{tibble} to canonical format}
\usage{
getCanonicalParamEquationsTbl(tbl)
}
\arguments{
\item{tbl}{\itemize{
\item \code{tibble} to convert
}}
}
\value{
a tibble with all "params equations" ("par_idx", "fcn_idx", "par_id", "feEQs", "feCons", "feLink", "reEQs", "reDisp", "reLink", "ecEQs", "ecLink") columns.
}
\description{
Convert CTL file "params equations" tibble to canonical format.
}
\details{
Converts a "params equations" \code{tibbl}e to canonical format (i.e., al columns). Missing columns
in the original tibble are added with \code{NA}s as values.
}
