% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Dataframe.R
\name{getColumnsByType}
\alias{getColumnsByType}
\title{Get the names of the columns with a given type from a dataframe}
\usage{
getColumnsByType(dfr, type = "all")
}
\arguments{
\item{dfr}{\itemize{
\item the dataframe
}}
}
\value{
character vector of names for columns with class matching \code{type}
}
\description{
Function to get the names of the columns with a given type from a dataframe.
}
