% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Dataframe.R
\name{getColumnTypes}
\alias{getColumnTypes}
\title{Get the classes (types) of columns in a dataframe}
\usage{
getColumnTypes(dfr)
}
\value{
a named character vector with values indicating the column class and names indicating the column
}
\description{
Function to get the classes (types) of columns in a dataframe.
}
