% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{getCorrMatFromTheta}
\alias{getCorrMatFromTheta}
\title{Get a correlation (or covariance) matrix from a Cholesky factor (lower triangle) \code{theta} parameterization}
\usage{
getCorrMatFromTheta(t, strictly.lower.triangle = TRUE, return.cov = FALSE)
}
\arguments{
\item{t}{\itemize{
\item a "\code{theta} vector for a Cholesky matrix, by row
}}

\item{strictly.lower.triangle}{\itemize{
\item if TRUE, \code{theta} corresponds to a strictly lower triangle Cholesky factor; if FALSE, it includes values along the main diagonal
}}

\item{return.cov-}{flag to return covariance matrix if strictly.lower.triangle=FALSE (default=FALSE)}
}
\value{
if \code{strictly.lower.triangle} is TRUE (default), the corresponding correlation matrix; otherwise, a covariance matrix
}
\description{
Function to get a correlation (or covariance) matrix from a Cholesky factor (lower triangle) \code{theta} parameterization.
}
\details{
If \code{strictly.lower.triangle} is FALSE, the first \code{n} elements of the input \code{theta} are interpreted as the elements of a \code{n} x \code{n}
diagonal scaling matrix, with values equal to \code{log(sd)/2}, where \code{sd^2} is the vector of variances scaling the correlation
matrix to a covariance matrix.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
##--ordinary R
tp = 0.3144855;
getCorrMatFromTheta(tp);
t = getThetaFromCovMat(getCorrMatFromTheta(tp),corrs.only=FALSE);
getCorrMatFromTheta(t,strictly.lower.triangle=FALSE);

S = matrix(c(2,0.3,0.3,2),nrow=2);
t = getThetaFromCovMat(S,corrs.only=FALSE);
Sp = getCorrMatFromTheta(t,strictly.lower.triangle=FALSE,return.cov=TRUE);

#--round trip
t - as.theta.vcov(getCorrMatFromTheta(t),TRUE);

##--with RTMB
require(RTMB);
tp = RTMB::MakeTape(getCorrMatFromTheta,numeric(1));
tp(t);
getCorrMatFromTheta(RTMB::AD(t));
#--can't do roundtrip as example because of `chol` call in `getThetaFromCovMat`.
\dontshow{\}) # examplesIf}
}
