% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsFunctions.R
\name{getDenseDimsIndex}
\alias{getDenseDimsIndex}
\title{Get the 1-d index associated with a vector of dimension indices for a fully-crossed expansion}
\usage{
getDenseDimsIndex(v, n, rev = FALSE)
}
\arguments{
\item{v}{\itemize{
\item vector of dimension indices
}}

\item{n}{\itemize{
\item vector of the number of corresponding dimension levels
}}

\item{rev}{\itemize{
\item flag to get the index for the reverse order
}}
}
\value{
an integer, the 1-d index into the fully-crossed dimensions array
}
\description{
Function to get the 1-d index associated with a vector of dimension indices
for a fully-crossed expansion of the dimensions.
}
\details{
If \code{rev} is false (the default), the index is calculated with the
first dimension cycling the fastest; otherwise, the index is calculated with the
last dimension cycling the fastest.
}
\examples{
#--examples for 3 dimensions, each with 5 levels
require(rtmbGMACS);
getDenseDimsIndex(c(1,1,1),c(5,5,5));
getDenseDimsIndex(c(5,1,1),c(5,5,5));
getDenseDimsIndex(c(1,5,1),c(5,5,5));
getDenseDimsIndex(c(1,1,5),c(5,5,5));
getDenseDimsIndex(c(5,5,1),c(5,5,5));
getDenseDimsIndex(c(1,5,5),c(5,5,5));
getDenseDimsIndex(c(5,5,5),c(5,5,5));
getDenseDimsIndex(c(1,1,1),c(5,5,5),TRUE);
getDenseDimsIndex(c(5,1,1),c(5,5,5),TRUE);
getDenseDimsIndex(c(1,5,1),c(5,5,5),TRUE);
getDenseDimsIndex(c(1,1,5),c(5,5,5),TRUE);
getDenseDimsIndex(c(5,5,1),c(5,5,5),TRUE);
getDenseDimsIndex(c(1,5,5),c(5,5,5),TRUE);
getDenseDimsIndex(c(5,5,5),c(5,5,5),TRUE);
}
