% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParamsInfo_FEs.R
\name{getFEs}
\alias{getFEs}
\title{Get fixed effects portion of formula}
\usage{
getFEs(formula, returnTerms = FALSE, debug = FALSE)
}
\arguments{
\item{formula}{\itemize{
\item formula to extract fixed effects from
}}
}
\value{
fixed effects formula
}
\description{
Function to get fixed effects portion of formula.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
getFEs(~1+a:b+(r|x)+s(x));
getFEs(~a:b);
getFEs(~(r|x)+s(x));   #--has implied intercept
getFEs(~0+(r|x)+s(x)); #--explicitly has no intercept
getFEs(~1+a:b+(r|x)+s(x),returnTerms=TRUE);
\dontshow{\}) # examplesIf}
}
