% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Formulas.R
\name{getFormulaAsText}
\alias{getFormulaAsText}
\title{Convert a formula to a text string}
\usage{
getFormulaAsText(f)
}
\arguments{
\item{f}{\itemize{
\item formula object
}}
}
\value{
a text string representation of the formula
}
\description{
Function to convert a formula to a text string .
}
\details{
None.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
##--one-sided formula
getFormulaAsText(~1+x*m+s(y));
##--two-sided formula
getFormulaAsText(z~1+x*m+s(y));
\dontshow{\}) # examplesIf}
}
