% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Matrices.R
\name{getLwrTriLenFromMatDim}
\alias{getLwrTriLenFromMatDim}
\title{From dimension of square matrix, compute length of strictly lower triangle}
\usage{
getLwrTriLenFromMatDim(matdim, strictly.lower.triangle = TRUE)
}
\arguments{
\item{matdim}{\itemize{
\item dimension (number of rows or columns) of square matrix
}}

\item{strictly.lower.triangle}{\itemize{
\item flag (TRUE/FALSE; default=TRUE)
}}
}
\value{
the length of a vector defining the (strictly, if \code{strictly.lower.triangle}=FALSE) lower triangle.
}
\description{
From dimension of square matrix, compute length of strictly lower triangle
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
t1 = getLwrTriLenFromMatDim(8);
t2 = getLwrTriLenFromMatDim(8,strictly.lower.triangle=FALSE);
\dontshow{\}) # examplesIf}
}
