% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Matrices.R
\name{getMatDimFromLwrTriLen}
\alias{getMatDimFromLwrTriLen}
\title{From length of (strictly) lower triangle, compute dimension of matrix}
\usage{
getMatDimFromLwrTriLen(ntri, strictly.lower.triangle = TRUE)
}
\arguments{
\item{ntri}{\itemize{
\item length of vector determining the (strictly) lower triangle
}}

\item{strictly.lower.triangle}{\itemize{
\item flag (TRUE/FALSE)
}}
}
\value{
the row or column dimension of the related square matrix
}
\description{
From length of (strictly) lower triangle, compute dimension of matrix
}
\details{
If strictly.lower.triangle=TRUE, the triangular matrix dimension
is strictly lower (i.e., below the main diagonal), if FALSE, it includes the
main diagonal.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
getMatDimFromLwrTriLen(getLwrTriLenFromMatDim(8));
getMatDimFromLwrTriLen(getLwrTriLenFromMatDim(8,strictly.lower.triangle=FALSE),strictly.lower.triangle=FALSE);
\dontshow{\}) # examplesIf}
}
