% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Parameters.R
\name{getParametersInitialValues}
\alias{getParametersInitialValues}
\title{Get initial values for a RTMB parameter}
\usage{
getParametersInitialValues(lst, dfrIVsp, links = NULL, verbose = FALSE)
}
\arguments{
\item{lst}{\itemize{
\item 
}}

\item{dfrIVsp}{\itemize{
\item dataframe with initial model parameter values
}}

\item{links}{\itemize{
\item links list (from \code{\link[=getLinkFcn]{getLinkFcn()}}) for transformation from/to model to/from RTMB parameter scales, or NULL
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
vector with initial values for RTMB parameter
}
\description{
Function to get initial values for a RTMB parameter.
}
\details{
TODO!
}
