% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Parameters.R
\name{getParametersPriorInfo}
\alias{getParametersPriorInfo}
\title{Create dataframe with priors info for parameters from a parameter values dataframe}
\usage{
getParametersPriorInfo(dfr)
}
\description{
Create dataframe with priors info for parameters from a parameter values dataframe
}
