% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Parameters.R
\name{getParamValsDataframe_colnames}
\alias{getParamValsDataframe_colnames}
\title{Get canonical names for a parameter values dataframe}
\usage{
getParamValsDataframe_colnames()
}
\description{
Get canonical names for a parameter values dataframe
}
