% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{getREs}
\alias{getREs}
\title{Extract RE terms from a formula}
\usage{
getREs(
  term,
  specials = names(.valid_covstruct),
  returnTerms = FALSE,
  debug = FALSE
)
}
\arguments{
\item{term}{\itemize{
\item formula or term to extract from
}}

\item{specials}{\itemize{
\item character vector of text strings to consider "special"
}}

\item{returnTerms}{\itemize{
\item flag to return a list of terms rather than a formula
}}

\item{debug}{\itemize{
\item flag to print debugginh info
}}
}
\value{
a formula, list, or NULL
}
\description{
Function to extract RE terms from a formula
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
res = getREs(~1+a:b+(r|x)+s(x));#--FEs+"ordinary" REs+a smooth
res = getREs(~a:b);             #--FEs only
res = getREs(~ar1(r|x)+s(x));   #--"special" REs plus smooth
res = getREs(~s(x));            #--smooth only
\dontshow{\}) # examplesIf}
}
