% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Matrices.R
\name{getRowColIndices_band}
\alias{getRowColIndices_band}
\title{Get a matrix of row/column indices along the diagonals of an n x n matrix.}
\usage{
getRowColIndices_band(n, diags)
}
\arguments{
\item{n}{-size of square matrix}

\item{diag}{\itemize{
\item vector of integer offsets from the main diagonal
}}
}
\value{
integer matrix of row/column indices
}
\description{
Get a matrix of row/column indices along the diagonals of an n x n matrix.
}
