% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Matrices.R
\name{getRowColIndices_diag}
\alias{getRowColIndices_diag}
\title{Get a matrix of row/column indices along the diagonal of an n x n matrix.}
\usage{
getRowColIndices_diag(n, diag)
}
\arguments{
\item{n}{-size of square matrix}

\item{diag}{\itemize{
\item integer offset from main diagonal
}}
}
\value{
integer matrix of row/column indices
}
\description{
Get a matrix of row/column indices along the diagonal of an n x n matrix.
}
