% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_SMs.R
\name{getSMs}
\alias{getSMs}
\title{Extract a formula for smooth terms from a formula}
\usage{
getSMs(formula, targets = .valid_smooths, returnTerms = FALSE, debug = FALSE)
}
\arguments{
\item{formula}{\itemize{
\item formula (or term) from which to extract smooth terms
}}

\item{targets}{\itemize{
\item character vector of smooth function names (default=.valid_smooths)
}}

\item{returnTerms}{\itemize{
\item flag to return a list of individual terms rather than a formula
}}

\item{debug}{\itemize{
\item flag to print debugging info
}}
}
\value{
formula with only smooth terms, or list of smooth terms, or NULL if no smooths
}
\description{
Extract a formula for smooth terms from a formula
}
\details{
Valid smooths are found in rtmbGMACS:::.valid_smooths.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
f = ~ s(z) + s:x;
sms = getSMs(f,targets="s",debug=TRUE);

f = ~ s:x;
sms = getSMs(f);

f = ~ 1 + s:x + s(x, bs = "tp") + ti(y, bs = "gp") + s(z) + (g|p)
sms = getSMs(f,returnTerms=TRUE);

f = ~ s(x,y, bs = "gp") + ti(z) + (g|p)
sms = getSMs(f,"s");
\dontshow{\}) # examplesIf}
}
