% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParamsInfo_FEs.R
\name{getTermsFEs}
\alias{getTermsFEs}
\title{Get terms in a fixed effects formula}
\usage{
getTermsFEs(trm, spc = "", debug = TRUE)
}
\arguments{
\item{spc}{\itemize{
\item spacer for debug printing
}}

\item{debug}{\itemize{
\item flag to print debugging info
}}

\item{formula}{\itemize{
\item formula to extract terms from
}}
}
\value{
list of terms from fixed effects formula in REVERSE order
}
\description{
Function to get terms in a fixed effects formula.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
getTermsFEs(~a*b+1);
getTermsFEs(~1);
getTermsFEs(~0+a:b+x);
getTermsFEs(~a:b + x + 0);
\dontshow{\}) # examplesIf}
}
