% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{getTermsREs}
\alias{getTermsREs}
\title{Extract RE terms from a formula}
\usage{
getTermsREs(
  term,
  specials = .valid_covstruct,
  pass = 0,
  termno = 0,
  debug = FALSE
)
}
\arguments{
\item{term}{\itemize{
\item formula or term to extract from
}}

\item{pass}{\itemize{
\item counter for debugging
}}

\item{termno}{\itemize{
\item counter for debugging
}}

\item{debug}{\itemize{
\item flag to print debugginh info
}}
}
\value{
a formula, list, or NULL
}
\description{
Function to extract RE terms from a formula
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
res = getTermsREs(~1+a:b+(r|x)+s(x));#--FEs + "ordinary" REs + a smooth
res = getTermsREs(~a:b);             #--FEs only
res = getTermsREs(~ar1(r|x)+s(x));   #--"special" REs + a smooth
res = getTermsREs(~s(x));            #--smooth only
\dontshow{\}) # examplesIf}
}
