% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_SMs.R
\name{getTermsSMs}
\alias{getTermsSMs}
\title{Extract smooth terms from a formula}
\usage{
getTermsSMs(trm, targets = .valid_smooths, spc = "", debug = FALSE)
}
\arguments{
\item{targets}{\itemize{
\item character vector of smooth function names (default=.valid_smooths)
}}

\item{debug}{\itemize{
\item flag to print debugging info
}}

\item{formula}{\itemize{
\item formula (or term) from which to extract smooth terms
}}
}
\value{
formula with only smooth terms, or list of smooth terms, or NULL if no smooths
}
\description{
Extract smooth terms from a formula
}
\details{
Valid smooths are found in rtmbGMACS:::.valid_smooths.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
f = ~ s(z) + s:x;
sms = getTermsSMs(f,debug=TRUE);

f = ~ s:x;            #--no smooths
sms = getTermsSMs(f);

f = ~ 1 + s:x + s(x, bs = "tp") + s(y, bs = "gp") + ti(z) + (g|p)
sms = getTermsSMs(f);
sms = getTermsSMs(f,targets="ti");

f = ~ ti(x,y, bs = "gp") + s(z) + (g|p)
sms = getTermsSMs(f);
\dontshow{\}) # examplesIf}
}
