% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{getThetaFromCovMat}
\alias{getThetaFromCovMat}
\title{Get the \code{theta} (Cholesky factor) parameterization of a covariance matrix}
\usage{
getThetaFromCovMat(S, corrs.only = FALSE)
}
\arguments{
\item{S}{a covariance matrix}

\item{corrs.only}{return only values corresponding to the correlation matrix parameters?}
}
\value{
the corresponding \code{theta} parameter vector
}
\description{
Function to get \code{theta} (Cholesky factor) parameterization of
a covariance matrix; from Balint Tamasi, TMB user's group list.
}
\details{
If \code{corrs.only}=TRUE, \code{theta} corresponds to a strictly lower triangle matrix.
If \code{corrs.only}=FALSE, then \code{theta} corresponds to a lower triangle matrix. If \code{n} is the length of the main diagonal of \code{S}, then
the first \code{n} elements of the returned \code{theta} vector are given by
\code{log(diag(S))/2}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
##--ordinary R
S = matrix(c(2,0.3,0.3,2),nrow=2);
getThetaFromCovMat(S);
getCorrMatFromTheta(getThetaFromCovMat(S),strictly.lower.triangle=FALSE)
getThetaFromCovMat(S,corrs.only=TRUE);

##--with RTMB
require(RTMB);

###--NOT the way  to test RTMB context(!!)
####--error arises from RTMB approach to making R Cholesky factorization RTMB-compatible
S = matrix(c(2,0.3,0.3,2),nrow=2)
S = RTMB::AD(S,force=TRUE);
getThetaFromCovMat(S)                  #--generates C stack error!

###--the following run fine
S = matrix(c(2,0.3,0.3,2),nrow=2)
tp = MakeTape(getThetaFromCovMat,S);
tp(S)
tp = MakeTape(function(x){getThetaFromCovMat(x,corrs.only=TRUE)},S);
tp(S)
\dontshow{\}) # examplesIf}
}
