% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{\%i\%}
\alias{\%i\%}
\title{infix interaction operator to combine levels from two factors}
\usage{
f1 \%i\% f2
}
\arguments{
\item{f1}{\itemize{
\item factor
}}

\item{f2}{\itemize{
\item factor
}}

\item{order}{\itemize{
\item flag to re-order levels
}}
}
\value{
factor with merged levels from \code{f1} and \code{f2}
}
\description{
Operator to combine levels from two factors.
}
\details{
Copied from reformulas:::\verb{\%i\%}.
}
