% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth.R
\name{grwPwrLaw1}
\alias{grwPwrLaw1}
\title{Calculate growth}
\usage{
grwPwrLaw1(pGrA, zGrA, pGrB, zGrB, pGrBeta, zBs)
}
\arguments{
\item{pGrA}{\itemize{
\item mean post-molt size at zSclGrA
}}

\item{zGrA}{\itemize{
\item pre-molt size at yielding pGrA as mean post-molt size
}}

\item{pGrB}{\itemize{
\item mean post-molt size at zSclGrB
}}

\item{zGrB}{\itemize{
\item pre-molt size at yielding pGrB as mean post-molt size
}}

\item{pGrBeta}{\itemize{
\item gamma distribution scale parameter for post-molt variability
}}

\item{zBs}{\itemize{
\item pre-molt sizes at which to calculate matrix
}}
}
\value{
growth matrix
}
\description{
Function to calculate growth.
}
\details{
The formula used is

mnZs = pGrA\*exp(log(pGrB/pGrA)/log(zGrB/zGrA)\*log(zBs/zGrA));
}
\examples{
# example code
grM = grwPwrLaw1(33,25,150,125,1,seq(25,180,5));

}
