% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth.R
\name{grwPwrLaw2}
\alias{grwPwrLaw2}
\title{Calculate size-dependent growth}
\usage{
grwPwrLaw2(pGrA, zGrA, pGrB, zGrB, pGrBeta, zBs)
}
\arguments{
\item{zBs}{\itemize{
\item pre-molt sizes at which to calculate growth
}}

\item{pLnM}{\itemize{
\item log-scale base mortality
}}

\item{pZ0}{\itemize{
\item reference size (fixed parameter)
}}
}
\value{
object with same dimensions as \code{z}.
}
\description{
Function to calculate size-dependent growth.
}
\details{
The formula for mean growth used is

mnZs = exp(grA+grB*log(zBs));
}
\examples{
# example code
z = seq(25,100,5);
M = natMortZ(log(0.2),100,z);

}
