% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_FEs.R
\name{hasFEs}
\alias{hasFEs}
\title{Test if formula includes fixed effects}
\usage{
hasFEs(formula)
}
\arguments{
\item{formula}{\itemize{
\item formula to test
}}
}
\value{
logical (T/F)
}
\description{
Function to test if formula includes fixed effects.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
hasFEs(~1+a:b+(r|x)+s(x));
hasFEs(~a:b);
hasFEs(~(r|x)+s(x));   #--has implied intercept
hasFEs(~0+(r|x)+s(x)); #--explicitly has no intercept
\dontshow{\}) # examplesIf}
}
