% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{hasREs}
\alias{hasREs}
\title{Test if formula includes random effects}
\usage{
hasREs(formula)
}
\arguments{
\item{formula}{\itemize{
\item formula to test
}}
}
\value{
logical (T/F)
}
\description{
Function to test if formula includes random effects.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
hasREs(~1+a:b+(r|x)+s(x));
hasREs(~a:b);          #--FEs only
hasREs(~(r|x)+s(x));   #--REs plus smooth
hasREs(~s(x));         #--smooth only
\dontshow{\}) # examplesIf}
}
