% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_SMs.R
\name{hasSMs}
\alias{hasSMs}
\title{Test if formula includes \code{mgcv}-type smooths}
\usage{
hasSMs(formula)
}
\arguments{
\item{formula}{\itemize{
\item formula to test
}}
}
\value{
logical (T/F)
}
\description{
Function to test if formula includes \code{mgcv}-type smooths.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
hasSMs(~1+a:b+(r|x)+s(x));
hasSMs(~a:b+s:x);      #--FEs only
hasSMs(~(r|x)+s(x));   #--REs plus smooth
hasSms(~s(x));         #--smooth only
hasSMs(~ar1(r|x));     #--"special" REs
\dontshow{\}) # examplesIf}
}
