% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{identifyVars}
\alias{identifyVars}
\title{Identify variables in a character string, formula, or call}
\usage{
identifyVars(trm)
}
\arguments{
\item{trm}{\itemize{
\item character string, formula, or call
}}
}
\value{
character vector with variables as character strings
}
\description{
Identify variables in a character string, formula, or call
}
\details{
If \code{trm} is a formula or call, \code{\link[=deparse1]{deparse1()}} is called on it
to obtain a character string. \code{trm} is then split using the regex
"[[~\*\|\+\-:/]]" and non-numeric terms and empty elements are
removed from the resulting character vector.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code

identifyVars(~0 + cos(x) + y/m | z:g+w*p)         #--formula
identifyVars("~0 + cos(x) + y/m | z:g+w*p")       #--character string
identifyVars("fred = 0 + cos(x) + y/m | z:g+w*p") #--equation as text
\dontshow{\}) # examplesIf}
}
