% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{identity}
\alias{identity}
\title{Identity function}
\usage{
identity(x)
}
\arguments{
\item{x}{\itemize{
\item object to "convert"
}}
}
\value{
the input object
}
\description{
Function that simply returns the input object
}
\details{
This function that simply returns the input.
}
