% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsFunctions.R
\name{is.DenseDimsMap}
\alias{is.DenseDimsMap}
\title{Test if an object is a dense DimsMap}
\usage{
is.DenseDimsMap(x)
}
\arguments{
\item{x}{\itemize{
\item object to test
}}
}
\value{
logical
}
\description{
Function to test if an object is a dense DimsMap.
}
\details{
None.
}
