% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsFunctions.R
\name{is.DimsMap}
\alias{is.DimsMap}
\title{Test if an object is a DimsMap}
\usage{
is.DimsMap(x)
}
\arguments{
\item{x}{\itemize{
\item object to test
}}
}
\value{
logical
}
\description{
Function to test if an object is a DimsMap.
}
\details{
None.
}
