% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged_array.R
\name{is.ragged_array}
\alias{is.ragged_array}
\title{Test if an object is a ragged_array}
\usage{
is.ragged_array(x)
}
\arguments{
\item{x}{- the object}
}
\value{
TRUE or FALSE
}
\description{
Function to test if an object is a ragged_array.
}
\details{
Returns TRUE of the object inherits from class "ragged_array".
}
