% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsFunctions.R
\name{is.SparseDimsMap}
\alias{is.SparseDimsMap}
\title{Test if an object is a sparse DimsMap}
\usage{
is.SparseDimsMap(x)
}
\arguments{
\item{x}{- object to test}
}
\value{
logical
}
\description{
Function to test if an object is a sparse DimsMap.
}
\details{
None.
}
