% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{isKeyWord}
\alias{isKeyWord}
\title{Check if a string matches a key word}
\usage{
isKeyWord(str, kws = c("list", "vector", "matrix", "dataframe"))
}
\arguments{
\item{str}{\itemize{
\item string to check
}}

\item{kws}{\itemize{
\item vector of key words to check against
}}
}
\value{
TRUE or FALSE.
}
\description{
Function to check if a string matches a key word.
}
\details{
uses \code{\link[stringr:str_trim]{stringr::str_trim()}} to trim starting/ending whitespace.
}
