% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{kg2g}
\alias{kg2g}
\title{Convert values from kilograms to grams}
\usage{
kg2g(x)
}
\arguments{
\item{x}{\itemize{
\item vector of values to convert
}}
}
\value{
vector of transformed values
}
\description{
Function to convert values from kilograms to grams.
}
