% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{kg2kt}
\alias{kg2kt}
\title{Convert values from kilograms to 1,000's t}
\usage{
kg2kt(x)
}
\arguments{
\item{x}{\itemize{
\item vector of values to convert
}}
}
\value{
vector of transformed values
}
\description{
Function to convert values from kilograms to 1,000's t.
}
