% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{kt2kg}
\alias{kt2kg}
\title{Convert values from 1,000s t to kilograms}
\usage{
kt2kg(x)
}
\arguments{
\item{x}{\itemize{
\item vector of values to convert
}}
}
\value{
vector of transformed values
}
\description{
Function to convert values from 1,000s t to kilograms.
}
