% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Catchability.R
\name{lnQ}
\alias{lnQ}
\title{Calculate catchability \code{Q} as a function of size}
\usage{
lnQ(z, pLnQ, verbose = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to calculate \code{Q}
}}

\item{pLnQ}{\itemize{
\item vector of ln-scale values of \code{Q}
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
arithmetic-scale vector of catchabilities the same size as \code{z}
}
\description{
Function to calculate catchability \code{Q} as a function of size.
}
\details{
This function simply exponentiates \code{pLnQ} and returns an arithmetic-scale vector.
\code{pLnQ} can be a vector, but it must be conformable to \code{z}.
}
