% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Misc.R
\name{logistic}
\alias{logistic}
\title{Convert values to the logistic scale}
\usage{
logistic(x)
}
\arguments{
\item{x}{- numeric object}
}
\value{
1.0/(1.0+exp(-x)), with same names and class as `x`
}
\description{
Function to convert values to the logistic scale.
}
\details{
inverse of logit function.
}
