% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Misc.R
\name{logit}
\alias{logit}
\title{Convert values to the logit scale}
\usage{
logit(x)
}
\arguments{
\item{x}{- numeric object}
}
\value{
log(x/(1-x)), with same names and class as `x`
}
\description{
Function to convert values to the logit scale.
}
\details{
inverse of logistic function. Values of `x` outside the
(0,1) range generate NaNs.
}
