% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{makeFac}
\alias{makeFac}
\title{Convert a vector to a factor}
\usage{
makeFac(v, char.only = TRUE)
}
\arguments{
\item{v}{\itemize{
\item the vector to convert
}}

\item{char.only}{\itemize{
\item flag to convert \code{v} only if it is a character vector
}}
}
\description{
Function to convert a vector to a factor
}
\details{
Copied from \code{reformulas:::makeFac()} (char.only option changed to TRUE).
}
