% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFileTemplates.R
\name{makeFileTemplate_ParamFEs}
\alias{makeFileTemplate_ParamFEs}
\title{Create a template for the information necessary to specify the FE components of a model parameter}
\usage{
makeFileTemplate_ParamFEs(
  x,
  model_frame = NULL,
  contrasts = NULL,
  drop.unused.levels = TRUE,
  tab_size = 4,
  debug = TRUE
)
}
\arguments{
\item{x}{\itemize{
\item formula, possibly containing fixed effects, or list from \code{\link[=createParamInfo_FEs]{createParamInfo_FEs()}}
}}

\item{model_frame}{\itemize{
\item model frame for parameter (required if \code{x} is a formula)
}}

\item{contrasts}{\itemize{
\item a list of contrasts (if not NULL)
}}

\item{drop.unused.levels}{\itemize{
\item flag to drop unused levels
}}

\item{tab_size}{\itemize{
\item number of spaces to convert tabs to
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging info
}}
}
\value{
a character vector
}
\description{
Function to create a template for the information necessary to specify the FE components of a model parameter.
}
