% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFileTemplates.R
\name{makeFileTemplate_ParamsRE}
\alias{makeFileTemplate_ParamsRE}
\title{Create a template for the information necessary to specify the RE components of a model parameter}
\usage{
makeFileTemplate_ParamsRE(
  formula,
  model_frame,
  contrasts = NULL,
  drop.unused.levels = TRUE,
  debug = TRUE
)
}
\arguments{
\item{formula}{\itemize{
\item formula, possibly containing random effects
}}

\item{model_frame}{\itemize{
\item model frame for parameter
}}

\item{contrasts}{\itemize{
\item a list of contrasts (if not NULL)
}}
}
\value{
a character vector
}
\description{
Function to create a template for the information necessary to specify the RE components of a model parameter.
}
