% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Formulas.R
\name{makeRHSFormula}
\alias{makeRHSFormula}
\title{Create a RHS formula from a call or a (list of) language object(s)}
\usage{
makeRHSFormula(terms)
}
\arguments{
\item{terms}{\itemize{
\item a call or a (list of) language object(s)
}}
}
\value{
a RHS formula
}
\description{
Create a RHS formula from a call or a (list of) language object(s)
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
f = ~ 1 + x + s(x, bs = "tp") + s(y, bs = "gp") + s(z)
bars = reformulas::findbars_x(f,target="s",default.special=NULL);
makeRHSFormula(bars);
\dontshow{\}) # examplesIf}
}
