% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{mfexp}
\alias{mfexp}
\title{Exponentiate a variable}
\usage{
mfexp(x)
}
\arguments{
\item{x}{\itemize{
\item variable to exponentiate
}}
}
\value{
vector x*x with same names as x
}
\description{
ADMB-equivalent  to exponentiate a variable.
}
\details{
Substitution for ADMB \code{mfexp} function.
}
