% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{mkPrecMat}
\alias{mkPrecMat}
\title{Make a precision matrix}
\usage{
mkPrecMat(n, pars, type, sparse = TRUE, force = FALSE)
}
\arguments{
\item{n}{\itemize{
\item dimension of the matrix
}}

\item{pars}{\itemize{
\item vector of parameter values determining the precision matrix
}}

\item{type}{\itemize{
\item type of precision matrix
}}

\item{sparse}{\itemize{
\item flag to create a sparse or dense matrix
}}

\item{force}{\itemize{
\item force an RTMB adsparse representation
}}
}
\value{
a precision matrix (a \link[Matrix:Matrix]{Matrix::Matrix} type or \link[RTMB:RTMB-package]{RTMB::RTMB} "adsparse" type)
}
\description{
Function to make a precision matrix.
}
\details{
If \code{force} is FALSE, the returned matrix will be an "adsparse" matrix only if
\code{pars} is an "advector"; otherwise, it will be a \link[Matrix:Matrix]{Matrix::Matrix} type.

The \code{type}s of precision matrices that can be created using this function are currently:
\itemize{
\item{"diag" - diagonal iid matrix with variance \code{V} along the diagonal}
\item{"iid" - synonym for "diag"}
\item{"ar1" - AR1 precision matrix with variance \code{V} and correlation coefficient \eqn{\rho} (latter parameterized as \eqn{\rho=\code{symlogistic(par)}})}
\item{"us" - unstructured precision matrix}
}

Variances \code{V} are generally parameterized on the ln scale such that \eqn{V = exp(2*par)}.
}
