% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{mkReTermInfoList}
\alias{mkReTermInfoList}
\title{Make an info list for a RE term}
\usage{
mkReTermInfoList(
  re_term,
  model_frame,
  drop.unused.levels = TRUE,
  sparse = NULL
)
}
\arguments{
\item{re_term}{a language object of the form  (effect | groupvar)}

\item{model_frame}{model frame}

\item{drop.unused.levels}{\itemize{
\item flag to drop unused levels from model matrix
}}

\item{sparse}{(logical) set up sparse model matrices?}
}
\value{
list, see details
}
\description{
Function to make an info list for a RE term.
}
\details{
Hacked from \code{reformulas:::mkBlist()}.

Return list has elements:
\itemize{
\item{re_term - expression for RE term}
\item{re_term_lbl - character string for RE term}
\item{factorized_model_frame - model frame with non-numeric columns as factors}
\item{ff - factors}
\item{termZt - transposed Z for term}
\item{ngrps - number of factor groups in term}
\item{nprs - number of RE parameters (i.e., per group) in term}
\item{npvs - number of RE parameter values in term}
\item{pr_lbls - labels for parameters}
\item{pv_lbls - labels for parameter values}
\item{covstr - covariance structure string}
\item{Qtp - template for precision matrix}
\item{fillQ - function to fill precision matrix}
}
}
