% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{namedList}
\alias{namedList}
\title{Create a list with names from input elements}
\usage{
namedList(...)
}
\arguments{
\item{...}{\itemize{
\item comma-separated objects to be included in the resulting list
}}
}
\value{
a named list
}
\description{
Function to create a list with names from input elements
}
