% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNaturalMortality.R
\name{natMort}
\alias{natMort}
\title{Calculate natural mortality}
\usage{
natMort(pLnM)
}
\arguments{
\item{pLnM}{\itemize{
\item ln-scale natural mortality rate
}}
}
\value{
mortality rate on arithmetic scale
}
\description{
Function to calculate natural mortality.
}
\details{
The returned object is the same dimensions and type as the
input \code{pLnM} object.The formula used is

M = exp(pLnM)
}
\examples{
# example code
M = natMort(log(0.2));

}
