% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNaturalMortality.R
\name{natMortZ}
\alias{natMortZ}
\title{Calculate size-dependent natural mortality}
\usage{
natMortZ(pLnM, pZ0, z)
}
\arguments{
\item{pLnM}{\itemize{
\item log-scale base mortality
}}

\item{pZ0}{\itemize{
\item reference size (fixed parameter)
}}

\item{z}{\itemize{
\item sizes at which to calculate mortality rates
}}
}
\value{
object with same dimensions as \code{z}.
}
\description{
Function to calculate size-dependent natural mortality.
}
\details{
The returned object is the same dimensions and type as the
input \code{z} object.If \code{pLnM} and \code{pZ0} are not scalars, they should be
conformable to the shape of \code{z}.

The formula used is

M(z) = exp{pLnM)*(pZ0/z)}
}
\examples{
# example code
z = seq(25,100,5);
M = natMortZ(log(0.2),100,z);

}
