% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged_array.R
\name{new_ragged_array}
\alias{new_ragged_array}
\title{Create a new ragged_array object}
\usage{
new_ragged_array(x, dfrDims)
}
\arguments{
\item{x}{- vector (or other object that can be converted to a vector using `as.vector(x)`)}

\item{dfrDims}{- dimensions dataframe (a DimsMap)}
}
\value{
a vector of class "ragged_array" with a "dfrDims" attribute
}
\description{
Internal constructor to create a new ragged_array object.
}
\details{
A ragged_array object is essentially a vector with an attribute (a tibble)
that maps the index of the vector to a set of dimensions.
If `length(as.vector(x))` is < `nrow(dfrDims)`, it is replicated using `rep_len(as.vector(x),nrow(dfrDims))`
to a length of `nrow(dfrDims)`. If `length(x)` is > `nrow(dfrDims)`, it is truncated to the
latter.

This function is not exported. Use `ragged_array(x,dfrDims)` instead.

See https://adv-r.hadley.nz/s3.html#s3-constructor for details on constructors.
}
