% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{none}
\alias{none}
\title{Alias for the \code{identity} \code{\link[=identity]{identity()}} function}
\usage{
none(x)
}
\arguments{
\item{x}{\itemize{
\item object to "convert"
}}
}
\value{
the input object
}
\description{
This function simply returns the input object
}
\details{
This function simply returns the input.
}
