% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function.R
\name{obj_fun}
\alias{obj_fun}
\title{Calculate the objective function for a gmacs model}
\usage{
obj_fun(params)
}
\arguments{
\item{params}{\itemize{
\item named list of parameter values (p=1 for now)
}}
}
\value{
The objective function (negative log-likelihood) value given the data and parameters
}
\description{
Function to calculate the objective function for a gmacs model.
}
\details{
Input data and options should be organized in the calling environment as a list of lists
with the name "inputs". \code{inputs} should have the structure
\itemize{
\item{dims - a list with model dimensions maps}
\item{data - a named list with data-related entries}
\item{options - a named list with selected options}
\item{testing - a logical to indicate the function is run in a testing environment}
}

Example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{inputs = list(data=NULL,options=NULL,testing=TRUE);
params = list(dummy=0.0);
obj = MakeADFun(objfun,params,silent=TRUE);
opt = nlminb(obj$par, obj$fn, obj$gr);
sdreport(obj);
}\if{html}{\out{</div>}}
}
