% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function.R
\name{objfcn}
\alias{objfcn}
\title{Calculate the objective function}
\usage{
objfcn(params)
}
\arguments{
\item{params}{- named list of parameter values (p=1 for now)}
}
\value{
The objective function (negative log-likelihood) value given the data and parameters
}
\description{
Function to calculate the objective function
}
\details{
Input data should be organized in the calling environment as a list with
the name "data"
}
