% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{parseDataframe}
\alias{parseDataframe}
\title{Parse multiple text lines into a dataframe}
\usage{
parseDataframe(lns, verbose = FALSE)
}
\arguments{
\item{lns}{- the text lines parse}

\item{verbose}{- flag (T/F) to print debugging info}
}
\value{
dataframe (a [tibble::tibble()]) parsed values
}
\description{
Function to parse multiple text lines into a dataframe until reaching an ">EOD<" marker.
}
\details{
IMPORTANT: A counter with the name `iln` should be defined in the global environment
while this function runs. The parsing of lines in `lns` uses the value of iln
at the time the function is called as the starting index in `lns`.
When the function exits, the value of `iln` indicates the index of the last element
examined.
}
