% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{parseMatrix}
\alias{parseMatrix}
\title{Parse multiple text lines into a matrix}
\usage{
parseMatrix(lns, verbose = FALSE)
}
\arguments{
\item{lns}{\itemize{
\item the text lines parse
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging info
}}
}
\value{
matrix of parsed values
}
\description{
Function to parse multiple text lines into a matrix until reaching an ">EOM<" marker.
}
\details{
IMPORTANT: A counter with the name \code{iln} should be defined in the global environment
while this function runs. The parsing of lines in \code{lns} uses the value of iln
at the time the function is called as the starting index in \code{lns}.
When the function exits, the value of \code{iln} indicates the index of the last element
examined.
}
