% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{parseTextToDataframe}
\alias{parseTextToDataframe}
\title{Parse a text vector to create a dataframe}
\usage{
parseTextToDataframe(lns, iln, nlns, col_names = TRUE, show_col_types = FALSE)
}
\arguments{
\item{lns}{\itemize{
\item character vector of text lines to parse
}}

\item{iln}{\itemize{
\item index to line at which to start parsing
}}

\item{nlns}{\itemize{
\item number of lines to parse
}}

\item{col_names}{\itemize{
\item flag (T/F) or column names (see \code{\link[readr:read_delim]{readr::read_delim()}})
}}

\item{show_col_types}{\itemize{
\item flag (T/F) to show column types (see \code{\link[readr:read_delim]{readr::read_delim()}})
}}
}
\value{
a tibble (see \link[tibble:tibble]{tibble::tibble})
}
\description{
Parse a text vector to create a dataframe
}
\details{
Uses \code{\link[readr:read_delim]{readr::read_delim()}} to parse the vector of text lines after
removing all trailing comments and replacing all white space with a
single space.
}
