% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Text.R
\name{parseToDigits}
\alias{parseToDigits}
\title{Parse strings in a character vector to digits}
\usage{
parseToDigits(x)
}
\arguments{
\item{x}{\itemize{
\item character vector to parse
}}
}
\value{
integer vector
}
\description{
Function to parse strings in a character vector to digits.
}
\details{
For each element in \code{x}, everything except for digits is removed and
the resulting character value is converted to an integer.
}
\examples{
parseToDigits(c("ab23","cd42","a42 b23"))

}
