% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{parseVal}
\alias{parseVal}
\title{Parse a string from an input data file to a value}
\usage{
parseVal(con, nl, ln = NULL)
}
\arguments{
\item{str}{- string to parse}

\item{type}{- expected value type (not used currently)}
}
\value{
a value with mode as determined by [readr::parse_guess()]
}
\description{
Function to parse a string from an input data file to a value.
}
\details{
Splits the string by whitespace and drops anything after the comment character,
using [scan()] with `comment.char`="#" and `strip.white`=TRUE followed by [readr::parse_guess()].
}
