% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged_array.R
\name{print.ragged_array}
\alias{print.ragged_array}
\title{print a ragged_array}
\usage{
\method{print}{ragged_array}(x)
}
\arguments{
\item{x}{- the ragged array to print}
}
\value{
none
}
\description{
S3 method to print a ragged_array (as a tibble).
}
\details{
Prints the ragged array as a tibble with the values of the vector
in the "val" column. This is an S3 generic method, so just use `print(x)`.
}
