% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth_PrM2M.R
\name{prM2M_AscLogistic}
\alias{prM2M_AscLogistic}
\title{Calculate a size-dependent, ascending logistic function for the probability of molting to maturity}
\usage{
prM2M_AscLogistic(mdZ, wdZ, zBs)
}
\arguments{
\item{mdZ}{\itemize{
\item size at which curve starts to descend from 1
}}

\item{wdZ}{\itemize{
\item width of descent (standard deviation of normal curve)
}}

\item{zBs}{\itemize{
\item pre-molt sizes at which to calculate vector
}}

\item{dZ}{\itemize{
\item size bin width to use to set
}}
}
\value{
object with same dimensions as \code{zBs}.
}
\description{
Function to calculate a size-dependent growth, ascending logistic function for the probability of molting to maturity.
}
\details{
The formula used is

$$prM(zBs) = exp(zBs-mdZ) \over (1+exp(zBs-mdZ))$$
}
\examples{
# example code
z = seq(25,100,5);
prM2M = prM2M_AscLogistic(55,30,z);

}
