% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth_PrM2M.R
\name{prM2M_AscNormal}
\alias{prM2M_AscNormal}
\title{Calculate a size-dependent, ascending normal probability of molting to maturity}
\usage{
prM2M_AscNormal(mdZ, wdZ, zBs)
}
\arguments{
\item{mdZ}{\itemize{
\item size at which curve ascends to 1
}}

\item{wdZ}{\itemize{
\item width of ascent (standard deviation of normal curve)
}}

\item{zBs}{\itemize{
\item pre-molt sizes at which to calculate vector
}}

\item{dZ}{\itemize{
\item size bin width to use to set
}}
}
\value{
object with same dimensions as \code{zBs}.
}
\description{
Function to calculate a size-dependent growth, descending normal probability of molting to maturity.
}
\details{
The formula used is

$$prM(zBs > mdZ) = 1.0$$
$$prM(zBs \le mdZ) = exp(0.5*((zBs-mdZ) \over wdZ)^2)$$
}
\examples{
# example code
z = seq(25,100,5);
prM2M = prM2M_AscNormal(55,30,z);

}
