% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth_PrMolt.R
\name{prMolt_Constant}
\alias{prMolt_Constant}
\title{Calculate the size-independent probability of undergoing a molt}
\usage{
prMolt_Constant(pM, zBs)
}
\arguments{
\item{zBs}{\itemize{
\item pre-molt sizes at which to calculate vector
}}

\item{prM}{\itemize{
\item size-independent probability of undergoing a molt
}}
}
\value{
object same size as zBs
}
\description{
Function to size-independent calculate the probability of undergoing a molt.
}
\details{
The returned object is the same size/shape as \code{zBs}, with the \code{prM}
for all elements
}
\examples{
# example code
prM = prMolt_Constant(0.5,seq(25,100,5));

}
