% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth_PrMolt.R
\name{prMolt_DscNormal}
\alias{prMolt_DscNormal}
\title{Calculate a size-dependent, descending normal probability of undergoing a molt}
\usage{
prMolt_DscNormal(mdZ, wdZ, zBs)
}
\arguments{
\item{mdZ}{\itemize{
\item size at which curve starts to descend from 1
}}

\item{wdZ}{\itemize{
\item width of descent (standard deviation of normal curve)
}}

\item{zBs}{\itemize{
\item pre-molt sizes at which to calculate vector
}}

\item{dZ}{\itemize{
\item size bin width to use to set
}}
}
\value{
object with same dimensions as \code{zBs}.
}
\description{
Function to calculate a size-dependent growth, descending normal probability of undergoing a molt.
}
\details{
The formula used is

$$prM(zBs < mdZ) = 1.0$$
$$prM(zBs \ge mdZ) = exp(-0.5*((zBs-mdZ) \over wdZ)^2)$$
}
\examples{
# example code
z = seq(25,100,5);
prM = prMolt_DscNormal(55,30,z);

}
