% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRecruitment_CategoryProportions.R,
%   R/calcRecruitment_SizeDistribution.R
\name{prRecZ1}
\alias{prRecZ1}
\title{Calculate a gamma-distrbuted size istribution at recruitment}
\usage{
prRecZ1(mnZ, wdZ, zMn, zMx, zBs, dZ)

prRecZ1(mnZ, wdZ, zMn, zMx, zBs, dZ)
}
\arguments{
\item{mnZ}{\itemize{
\item mean size at recruitment
}}

\item{wdZ}{\itemize{
\item width (std. dev of gamma distribution)
}}

\item{zMn}{\itemize{
\item minimum size with non-zero probability of recruitment (see Details)
}}

\item{zMx}{\itemize{
\item maximum size with non-zero probability of recruitment (see Details)
}}

\item{zBs}{\itemize{
\item sizes at which to calculate vector
}}

\item{dZ}{\itemize{
\item size bin width to use to set
}}
}
\value{
object with same dimensions as \code{zBs}.

object with same dimensions as \code{zBs}.
}
\description{
Function to calculate a gamma-distrbuted size istribution at recruitment.

Function to calculate a gamma-distributed size distribution at recruitment.
}
\details{
The formula used is

$$prZ(zBs < zMn or zBs > zMx)   = 0.0$$
otherwise, for mnZ <=zBs<=zMx,
$$prZmn = pgamma(mnZ-dZ/2,shape=(mnZ^2)/(wdZ^2),scale=(wdZ^2)/mnZ)$$
$$prZmx = pgamma(zMx+dZ/2,shape=(mnZ^2)/(wdZ^2),scale=(wdZ^2)/mnZ)$$
$$prZ(zBs) = pgamma(zBs+dZ/2,shape=(mnZ^2)/(wdZ^2),scale=(wdZ^2)/mnZ) -
pgamma(zBs-dZ/2,shape=(mnZ^2)/(wdZ^2),scale=(wdZ^2)/mnZ)$$
$$prZ(zBs) = prZ(zBs \le zMx)/(prZMx-prZmn)$$
}
\section{cat("dZ:",dZ,"\n");}{

}

\section{cat("ZBs\n");}{

}

\section{print(zBs);}{

}

\section{cat("mnZ\n");}{

}

\section{print(mnZ);}{

}

\section{cat("wdZ\n");}{

}

\section{print(wdZ);}{

}

\section{cat("zMn\n");}{

}

\section{print(zMn);}{

}

\section{cat("zMx\n");}{

}

\section{print(zMx);}{
#--calculate squarewave to "box" size range
sqw = squarewave(zMn,zMx,zBs,dZ)
}

\section{cat("sqw\n");}{

}

\section{print(sqw);}{

}

\section{cat("1-sqw\n");}{

}

\section{print(1-sqw);}{
shp = (mnZ^2)/(wdZ^2);
}

\section{cat("shp\n");}{

}

\section{print(shp);}{
scl = (wdZ^2)/mnZ;
}

\section{cat("scl\n");}{

}

\section{print(scl);}{
prZmn = pgamma(zMn-dZ/2,shape=shp,scale=scl);
prZmx = pgamma(zMx+dZ/2,shape=shp,scale=scl);
prZ   = pgamma(zBs+dZ/2,shape=shp,scale=scl) -
pgamma(zBs-dZ/2,shape=shp,scale=scl);
}

\section{cat("prZ before scaling\n")}{

}

\section{print(prZ);}{



}

\section{print(prZmx- prZmn);}{
#--normalize and apply squarewave window so sum  across zBs is 1 for each population category
prZ = sqw*prZ/(prZmx- prZmn);
}

\section{cat("prZ after scaling\n")}{

}

\section{cat("sum(prZ):",sum(prZ),"\n");}{

}

\examples{
# example code
zBs = seq(25,100,5);
dZ  = z[2]-z[1];
prZ = prRecZ1(35,20,55,25,80,zBs,dZ);

# example R code
zBs = seq(25,100,5);
dZ  = z[2]-z[1];
prZ = prRecZ1(35,20,55,25,80,zBs,dZ);

}
