% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADMB_DataFileFunctions.R
\name{readADMB_DataFile}
\alias{readADMB_DataFile}
\title{Read a slightly modified ADMB GMACS input data file}
\usage{
readADMB_DataFile(fn)
}
\arguments{
\item{fn}{- path to input data file}
}
\value{
list
}
\description{
Function to read a slightly modified ADMB GMACS input data file.
}
\details{
The input ADMB GMACS file should be modified in the following manner:
\itemize{
\item{The initial section (prior to the Catch Data section) should be removed or commented out.}
\item{The comment character "#" should be removed from the header line for input Catch Data, Relative Abundance Indices, Size Comps, Growth Data, and Environmental Indices dataframes}
}

The output list has an attribute "type" = "input data list" with elements
\itemize{
  \item{dfrCDF - list of catch data dataframes}
  \item{dfrID  - list of index data dataframes}
  \item{dfrZCD - list of size comps dataframes}
  \item{dfrGrD - list of growth (molt increment) data dataframes}
  \item{dfrMOD - list of maturity ogive data dataframes}
  \item{dfrTD - list of tagging data dataframes}
  \item{dfrED - list of environmental data dataframes}
}
}
