% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADMB_DatFile.R
\name{readADMB_DatFile}
\alias{readADMB_DatFile}
\title{Read an ADMB gmacs.dat file}
\usage{
readADMB_DatFile(fn, nFlts)
}
\arguments{
\item{fn}{\itemize{
\item filename for the admb gmacs.dat file
}}

\item{nFlts}{\itemize{
\item number of fleets defined
}}
}
\value{
list of options for running an ADMB gmacs model, with attribute "type" = "admb dat file inputs"
}
\description{
Function to read an ADMB gmacs.dat file (\emph{not} an ADMB gmacs model data file).
}
\details{
The returned list has the following elements:
\itemize{
\item {optMSY}
\item {dfrOptsFixedF - dataframe indicating which fleets have fixed Fs for OFL calculations}
\item {dfrOptsTBs - dataframe with time blocks for OFL calculations}
\item {dfrOptsOFL - dataframe with OFL calculation options}
\item {dfrOptsPrj - dataframe with projection options}
\item {dfrOptsHSs - dataframe with state harvest strategy options}
\item {optMaxMCMC - max number of MCMC draws}
\item {optFullDiagnostics - flag to print full diagnostics}
}

If it doesn't already exist, a counter \code{iln} is created in the global environment to keep
track of parsing progress. If it already exists, a copy of the current value is made and
\code{iln} is set to 1. Upon function exit, \code{iln} is reset to its original value if it existed before the function was called;
otherwise it is removed.
}
