% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADMB_PrjFile.R
\name{readADMB_PrjFile}
\alias{readADMB_PrjFile}
\title{Read an ADMB gmacs prj file}
\usage{
readADMB_PrjFile(fn, nFlts)
}
\arguments{
\item{fn}{\itemize{
\item filename for admb gmacs prj file
}}

\item{nFlts}{\itemize{
\item number of fleets defined
}}
}
\value{
list of options for projections, with attribute "type" = "admb prj file inputs"
}
\description{
Function to read an ADMB gmacs prj file.
}
\details{
The returned list has the following elements:
\itemize{
\item {optMSY}
\item {dfrOptsFixedF - dataframe indicating which fleets have fixed Fs for OFL calculations}
\item {dfrOptsTBs - dataframe with time blocks for OFL calculations}
\item {dfrOptsOFL - dataframe with OFL calculation options}
\item {dfrOptsPrj - dataframe with projection options}
\item {dfrOptsHSs - dataframe with state harvest strategy options}
\item {optMaxMCMC - max number of MCMC draws}
\item {optFullDiagnostics - flag to print full diagnostics}
}
}
