% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADMB_DataFileFunctions.R
\name{readCatchData}
\alias{readCatchData}
\title{Read the Catch Data section of an input data file}
\usage{
readCatchData(lns)
}
\value{
a tibble (see [tibble::tibble]) with attribute "type" = "catch data dataframe"
}
\description{
Read the Catch Data section of an input data file
}
