% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADMB_DataFileFunctions.R
\name{readCatchDataFrameFormat1}
\alias{readCatchDataFrameFormat1}
\title{Read a catch data frame in GMACS input format 1}
\usage{
readCatchDataFrameFormat1(lnsp)
}
\description{
Read a catch data frame in GMACS input format 1
}
