% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDataFile.R
\name{readDataFile}
\alias{readDataFile}
\title{Read a slightly modified ADMB GMACS input data file}
\usage{
readDataFile(fn)
}
\arguments{
\item{fn}{- path to input data file}
}
\value{
list
}
\description{
Function to read a slightly modified ADMB GMACS input data file.
}
\details{
The input ADMB GMACS file should be modified in the following manner:
\itemize{
\item{The initial section (prior to the Catch Data section) should be removed or commented out.}
\item{The comment character "#" should be removed from the header line for input Catch Data, Relative Abundance Indices, Size Comps, Growth Data, and Environmental Indices dataframes}
}
}
