% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADMB_DataFileFunctions.R
\name{readEnvironmentalData}
\alias{readEnvironmentalData}
\title{Read the Environmental Data section of an input data file}
\usage{
readEnvironmentalData(lns)
}
\value{
a tibble (see \link[tibble:tibble]{tibble::tibble}) with attribute "type" = "environmental data dataframes"
}
\description{
Read the Environmental Data section of an input data file
}
