% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile_CTL.R
\name{readFile_CTL}
\alias{readFile_CTL}
\title{Read CTL file TODO: finish this!}
\usage{
readFile_CTL(fn)
}
\arguments{
\item{fn}{\itemize{
\item path/name to the CTL file
}}
}
\value{
very long list
}
\description{
Function to read the CTL file.
}
