% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADMB_DataFileFunctions.R
\name{readGrowthData}
\alias{readGrowthData}
\title{Read the Growth Data section of an input data file}
\usage{
readGrowthData(lns)
}
\value{
a tibble (see \link[tibble:tibble]{tibble::tibble}) with attribute "type" = "growth data dataframe"
}
\description{
Read the Growth Data section of an input data file
}
