% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADMB_DataFileFunctions.R
\name{readIndexData}
\alias{readIndexData}
\title{Read the Index Data section of an input data file}
\usage{
readIndexData(lns)
}
\value{
a tibble (see \link[tibble:tibble]{tibble::tibble}) with attribute "type" = "relative abundance data dataframe"
}
\description{
Read the Index Data section of an input data file
}
