% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADMB_DataFileFunctions.R
\name{readIndexDataFrameFormat1}
\alias{readIndexDataFrameFormat1}
\title{Read an index data frame in GMACS input format 1}
\usage{
readIndexDataFrameFormat1(lnsp)
}
\description{
Read an index data frame in GMACS input format 1
}
