% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInfo_FisheriesRates.R
\name{readInfo_FisheriesRates}
\alias{readInfo_FisheriesRates}
\title{Read fisheries rates info}
\usage{
readInfo_FisheriesRates(conn, verbose = FALSE)
}
\arguments{
\item{conn}{\itemize{
\item connection (or filename) to read from
}}

\item{verbose}{\itemize{
\item flag to print extra info
}}
}
\value{
nested list with elements (see \strong{Details})
}
\description{
Function to read fisheries rates info.
}
\details{
TBD.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
if (FALSE){
  conn="inputSpecs_FisheriesRates.txt";
  lstResults = readInfo_FisheriesRates(conn,verbose=TRUE);
}
\dontshow{\}) # examplesIf}
}
