% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInfoType1.R
\name{readInfoType1}
\alias{readInfoType1}
\title{Read a model info section}
\usage{
readInfoType1(lns, verbose = FALSE)
}
\arguments{
\item{verbose}{\itemize{
\item flag to print extra info
}}

\item{conn}{\itemize{
\item connection (or filename) to read from
}}
}
\value{
nested list with elements (see \strong{Details})
}
\description{
Function to read a model info section
}
\details{
The returned list has elements
\itemize{
\item{Fcns - list with elements}
\itemize{
\item{n - number of functions defined}
\item{dfr - dataframe with function specs}
}
}
}
