% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParamInfo_InitialN.R
\name{readParamInfo_InitialN}
\alias{readParamInfo_InitialN}
\title{Read initial N-related parameter info}
\usage{
readParamInfo_InitialN(conn, debug = FALSE)
}
\arguments{
\item{conn}{\itemize{
\item connection (or filename) to read from
}}

\item{debug}{\itemize{
\item flag to print extra info
}}
}
\value{
nested list with elements (see \strong{Details})
}
\description{
Function to read initial N-related parameter info.
}
\details{
The returned list has elements
\itemize{
\item{option - initial N calculation option}
\item{lnTotRec - ln-scale total recruitment (only if \code{option}="free_parameters_v2")}
\item{dfr - dataframe, with parameter specifications (if \code{option}!="zero_pop")}
}
}
