% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDataFile.R
\name{readRelativeAbundanceData}
\alias{readRelativeAbundanceData}
\title{Read Relative Abundance Data section of input data file}
\usage{
readRelativeAbundanceData(lns)
}
\description{
Read Relative Abundance Data section of input data file
}
