% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDataFile.R
\name{readRelativeAbundanceDataFrameFormat1}
\alias{readRelativeAbundanceDataFrameFormat1}
\title{Read a catch data frame in GMACS input format 1}
\usage{
readRelativeAbundanceDataFrameFormat1(lnsp)
}
\description{
Read a catch data frame in GMACS input format 1
}
