% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADMB_DataFileFunctions.R
\name{readSizeComps}
\alias{readSizeComps}
\title{Read the Size Comps section of an input data file}
\usage{
readSizeComps(lns)
}
\value{
a tibble (see [tibble::tibble]) with attribute "type" = "size comps dataframe"
}
\description{
Read the Size Comps section of an input data file
}
