% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDataFile.R
\name{readSizeCompsDataFrameFormat1}
\alias{readSizeCompsDataFrameFormat1}
\title{Read a catch data frame in GMACS input format 1}
\usage{
readSizeCompsDataFrameFormat1(lnsp)
}
\description{
Read a catch data frame in GMACS input format 1
}
