% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRecruitment_TimeSeries.R
\name{recTS_exp}
\alias{recTS_exp}
\title{Calculate a gamma-distrbuted size istribution at recruitment}
\usage{
recTS_exp(pLnR)
}
\arguments{
\item{mnZ}{\itemize{
\item mean size at recruitment
}}

\item{wdZ}{\itemize{
\item width (std. dev of gamma distribution)
}}

\item{zMn}{\itemize{
\item minimum size with non-zero probability of recruitment
}}

\item{zMx}{\itemize{
\item maximum size with non-zero probability of recruitment
}}

\item{zBs}{\itemize{
\item sizes at which to calculate vector
}}

\item{dZ}{\itemize{
\item size bin width to use to set
}}
}
\value{
object with same dimensions as \code{zBs}.
}
\description{
Function to calculate a gamma-distrbuted size istribution at recruitment.
}
\details{
The formula used is

$$recTS(zBs < zMn or zBs > zMx)   = 0.0$$
otherwise, for mnZ <=zBs<=zMx,
$$recTSmn = pgamma(mnZ-dZ/2,shape=(mnZ^2)/(wdZ^2),scale=(wdZ^2)/mnZ)$$
$$recTSmx = pgamma(zMx+dZ/2,shape=(mnZ^2)/(wdZ^2),scale=(wdZ^2)/mnZ)$$
$$recTS(zBs) = pgamma(zBs+dZ/2,shape=(mnZ^2)/(wdZ^2),scale=(wdZ^2)/mnZ) -
pgamma(zBs-dZ/2,shape=(mnZ^2)/(wdZ^2),scale=(wdZ^2)/mnZ)$$
$$recTS(zBs) = recTS(zBs \le zMx)/(recTSMx-recTSmn)$$
}
\examples{
# example code
zBs = seq(25,100,5);
dZ  = z[2]-z[1];
recTS = prRecZ1(35,20,55,25,80,zBs,dZ);

}
