% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Txt.R
\name{removeCommentLines}
\alias{removeCommentLines}
\title{Remove strings starting with a comment character}
\usage{
removeCommentLines(strv, comment = "#")
}
\arguments{
\item{strv}{- character vector to remove comment lines from}

\item{comment}{- character(s) to regard as indicating a comment}
}
\value{
a character vector with elements that don't start with a comment character
}
\description{
Function to remove strings starting with a comment character.
}
\examples{
txt = c("#this is a comment",
        "  #so is this,even thought the line starts with blank space",
        "this is not a comment",
        "this is not either, but the first is line 3.");
removeCommentLines(txt); #--should be \code{txt[3:4]}

}
