% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Formulas.R
\name{removeSmoothsFromFormulaText}
\alias{removeSmoothsFromFormulaText}
\title{Remove \code{mgcv} "smooths" from text defining a formula}
\usage{
removeSmoothsFromFormulaText(txt)
}
\arguments{
\item{txt}{\itemize{
\item length 1 character vector with resulting text describing a formula
}}
}
\value{
length 1 character vector with smooth functions removed
}
\description{
Function to remove \code{mgcv} "smooths" from text defining a formula.
}
\details{
This function is used when processing a formula (as text) to remove any
\code{mgcv} smooth functions \code{s} and \code{ti} from the text. \code{mgcv} smooth functions must be
handled separately in terms of creating a model matrix from a formula and a model frame.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
txt = "~1 + yblk + s(z,k=10) + s(m,k=20,by='yblk') +ti(z)";
res = removeSmoothsFromFormulaText(txt); #--res == "~1+yblk";
\dontshow{\}) # examplesIf}
}
