% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{replaceTerm}
\alias{replaceTerm}
\title{Replace part (or all) of a formula term}
\usage{
replaceTerm(term, target, repl)
}
\arguments{
\item{term}{\itemize{
\item the term in which to make the replacement
}}

\item{target}{\itemize{
\item the part to replace
}}

\item{repl}{\itemize{
\item the replacement
}}
}
\value{
term with part replaced
}
\description{
Function to replace part (or all) of a formula term.
}
\details{
Copied from \code{reformulas:::replaceTerm()}.
}
