% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Formulas.R
\name{reviseFormula}
\alias{reviseFormula}
\title{Remove all single-value variables included in a model frame from a formula}
\usage{
reviseFormula(f, dfrMF)
}
\arguments{
\item{f}{\itemize{
\item model formula or equivalent text string (1-element character vector)
}}

\item{dfrMF}{\itemize{
\item basis dataframe from which to extract the model frame for \code{f}
}}
}
\description{
Function to remove all single-value variables included in a model frame from a formula.
}
\details{
\code{R} formulae cannot be used with variables with only one value in the model frame due to the apparent lack
of a method to create 1-element contrasts when creating the model matrix (\code{R} throws an error when any
1-value variables are included in a formula). Based on the input model frame (\code{dfrMF}), this function removes any
1-value variables included in the formula \code{f} and returns the revised formula (or the original, if no variables in the model frame had only one value)

This function throws an error if all variables in \code{f} have only one value and \code{f} does not include an intercept term.
}
