% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{rmChars}
\alias{rmChars}
\title{Remove all instances of a pattern from a character vector}
\usage{
rmChars(str, pattern = "")
}
\arguments{
\item{str}{\itemize{
\item character vector to search
}}

\item{pattern}{\itemize{
\item regex of pattern to remove
}}
}
\value{
character vector with pattern removed
}
\description{
Function to remove all instances of a pattern from a character vector.
}
