% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{selClmpSplineRight}
\alias{selClmpSplineRight}
\title{Calculates a selectivity curve using a right-clamped spline function}
\usage{
selClmpSplineRight(z, params, consts, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute function values}

\item{params}{- logit-scale vector of function parameters (values at knots)}

\item{consts}{- vector of knots}
}
\value{
- named vector of selectivity values at `z`
}
\description{
Function to calculate a selectivity curve using a right-clamped spline function.
}
\details{
Calculates a selectivity curve using a right-clamped spline function based on
[RTMB::splinefun()]. The values in the `params` vector are the logit-scale
values at the knots. The first `n` values in the `consts` vector are the knots,
where `n` is th length of `params`. The spline is "clamped" at the righthand end
by adding `nr` extra knots at the end with values equal to the
last (upper knots) values in `param`. The separation of the extra knots is
taken as `1/(nr+1)` the minimum separation in the `z` vector. Any values extrapolated
for `z` values greater than the last extra knot
will be the same as the the inverse logit-transformed value of the last value
of the `param` vector.
}
