% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{selSplineClmpd}
\alias{selSplineClmpd}
\title{Calculates a selectivity curve using a clamped spline function}
\usage{
selSplineClmpd(z, params, consts, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute function values}

\item{params}{- logit-scale vector of function parameters (values at knots)}

\item{consts}{- vector of knots}
}
\value{
- named vector of selectivity values at `z`
}
\description{
Function to calculate a selectivity curve using a clamped spline function.
}
\details{
Calculates a selectivity curve using a clamped spline function based on
[RTMB::splinefun()]. The values in the `params` vector are the logit-scale
values at the knots. The first `n` values in the `consts` vector are the knots,
where `n` is th length of `params`. The spline is "clamped" at both ends
by adding `nr` extra knots at each end with values equal to the first (lower knots)
or last (upper knots) values in `param`. The separation of the extra knots is
taken as `1/(nr+1)` times the minimum separation in the `z` vector. Any values extrapolated
for `z` values less than the first extra knot (or greater than the last extra knot)
will be the same as the the inverse logit-transformed value of the first (last) value
of the `param` vector.
}
