% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{selSplineClmpdRight}
\alias{selSplineClmpdRight}
\title{Calculates a selectivity curve using a right-clamped spline function}
\usage{
selSplineClmpdRight(z, params, consts, debug = FALSE)

selSplineClmpdRight(z, params, consts, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute function values
}}

\item{params}{\itemize{
\item logit-scale vector of function parameters (values at knots)
}}

\item{consts}{\itemize{
\item vector of knots
}}

\item{knots}{\itemize{
\item (AD constant) vector of knots
}}
}
\value{
\itemize{
\item vector of selectivity values at \code{z}
}

\itemize{
\item named vector of selectivity values at \code{z}
}
}
\description{
Function to calculate a selectivity curve using a right-clamped spline function.

Function to calculate a selectivity curve using a right-clamped spline function.
}
\details{
The parameter pRefZ is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.

Calculates a selectivity curve using a right-clamped spline function based on
\code{\link[RTMB:splinefun]{RTMB::splinefun()}}. The values in the \code{params} vector are the logit-scale
values at the knots. The first \code{n} values in the \code{knots} vector are the knots,
where \code{n} is th length of \code{params}. The spline is "clamped" at the righthand end
by adding \code{nr} extra knots at the end with values equal to the
last (upper knots) values in \code{param}. The separation of the extra knots is
taken as \code{1/(nr+1)} the minimum separation in the \code{z} vector. Any values extrapolated
for \code{z} values greater than the last extra knot
will be the same as the the inverse logit-transformed value of the last value
of the \code{param} vector.

Calculates a selectivity curve using a right-clamped spline function based on
\code{\link[RTMB:splinefun]{RTMB::splinefun()}}. The values in the \code{params} vector are the logit-scale
values at the knots. The first \code{n} values in the \code{consts} vector are the knots,
where \code{n} is th length of \code{params}. The spline is "clamped" at the righthand end
by adding \code{nr} extra knots at the end with values equal to the
last (upper knots) values in \code{param}. The separation of the extra knots is
taken as \code{1/(nr+1)} the minimum separation in the \code{z} vector. Any values extrapolated
for \code{z} values greater than the last extra knot
will be the same as the the inverse logit-transformed value of the last value
of the \code{param} vector.
}
