% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{splitForm_RE}
\alias{splitForm_RE}
\title{Split formula containing random effect terms}
\usage{
splitForm_RE(re_form, defaultCovStruct = "diag", debug = FALSE)
}
\arguments{
\item{re_form}{\itemize{
\item a formula containing random effect terms, or a single RE term
}}

\item{defaultCovStruct}{\itemize{
\item covariance structure name for non-"special" RE terms (default="diag")
}}

\item{debug}{\itemize{
\item flag (TRUE/FALSE) to print debugging info
}}
}
\value{
NULL, or a list containing elements
\itemize{
\item{\code{reTrms} named list with each RE term as an element}
\item{\code{reTrmFormulas} named list of \code{x | g} formulas for each term}
\item{\code{reTrmAddArgs} named list (see details)}
\item{\code{reTrmCovTypes} named character vector of covariance structure types}
}
}
\description{
Parse a formula into random effect terms, identifying covariance structures
}
\details{
Hacked from \code{\link[reformulas:splitForm]{reformulas::splitForm()}} to consider only RE terms.
reformulas::splitForm taken from Steve Walker's lme4ord,
ultimately from the flexLambda branch of the \code{lme4} package
\url{https://github.com/stevencarlislewalker/lme4ord/blob/master/R/formulaParsing.R}.

The names associated with all elements in the returned list are the same names as used for the \code{reTrms} list.

Each element in the \code{reTrmAddArgs} list is a language object, which itself has a list-like structure (use \code{[[} to access elements).
For each language object, the first element will be the covariance structure type, as a \code{symbol}.
Any additional elements represent additional arguments determining the covariance structure.
Additional arguments of the form "x=value"can be recovered as \code{value = el[["x"]]}, where \code{el} is list element.
}
\examples{
sf = splitForm_RE(re_form=~x+y)                     ## no REs
sf = splitForm_RE(re_form=~x+y+(f|g))               ## no covariance structure specified
sf = splitForm_RE(re_form=~x+y+ar1(f|g))            ## covariance structure specified for RE term
sf = splitForm_RE(re_form=~x+y+(diag(f|g)))         ## 'hidden' term with covariance structure specified
sf = splitForm_RE(re_form=~x+y+((ar1(f|g))))        ## doubly-'hidden' term with covariance structure specified
sf = splitForm_RE(re_form=~x+y+(f|g)+cs(1|g))       ## combination of terms with covariance structures un/specified
sf = splitForm_RE(re_form=~x+y+(1|f/g))                      ## 'slash' term: equivalent to (1|g:f) + (1|f)
sf = splitForm_RE(re_form=~x+y+(1|f/g/h))                    ## 'slash' term: equivalent to (1|h:g:f) + (1|g:f) + (1|f)
sf = splitForm_RE(re_form=~x+y+(1|(f/g)/h))                  ## 'slash' term: equivalent to (1|h:g:f) + (1|g:f) + (1|f)
sf = splitForm_RE(re_form=~x+y+(f|g)+cs(1|g)+cs(a|b,stuff))  ## complex covariance structures specified
sf = splitForm_RE(re_form=~(((x+y))))                        ## lots of parentheses
sf = splitForm_RE(re_form=~1+cs(f|g,n=2))                    ## covariance structure with one additional argument
sf = splitForm_RE(re_form=~1+cs(f|g,n=2,stuff))              ## covariance structure with multiple additional arguments (TODO: returns NULL!)
sf = splitForm_RE(re_form=~1+cs(f|g,n=2,m=3))                ## covariance structure with multiple additional arguments (TODO: returns NULL!)
sf = splitForm_RE(re_form=~1+cs(f|g,list(n=2,m=3)))          ## covariance structure with multiple additional arguments given as list (WORKS!)
sf = splitForm_RE(re_form=~1+s(x, bs = "tp"))                ## smooth special (ignored)

}
\author{
Steve Walker (hacked by William Stockhausen)
}
