% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_SMs.R
\name{splitForm_SM}
\alias{splitForm_SM}
\title{Split formula containing special random effect terms}
\usage{
splitForm_SM(formula, debug = FALSE, sm_fcns = findSmTrmClasses())
}
\arguments{
\item{formula}{a formula containing special random effect terms}

\item{debug}{flag (TRUE/FALSE) to print debugging info}

\item{sm_fcns}{\itemize{
\item vector of names of valid "smooth" functions
}}
}
\value{
a list containing elements
\itemize{
\item{\code{smTrmFormulas} list of \code{x | g} formulas for each term}
\item{\code{smTrmAddArgs} list of function+additional arguments, i.e. \code{list()} (non-special), \code{foo()} (no additional arguments), \code{foo(addArgs)} (additional arguments)}
\item{\code{smTrmClasses} (vector of special functions/classes, as character)}
}
}
\description{
Parse a formula into smooth terms,
treating 'special' terms (of the form foo(x|g[,m])) appropriately
}
\details{
Hacked from \code{\link[reformulas:splitForm]{reformulas::splitForm()}} to consider only RE terms.
}
\examples{
splitForm_SM(~x+y)                     ## fixed effects only
splitForm_SM(~x+y+(f|g))               ## fixed and random effects
splitForm_SM(~x+y+diag(f|g))           ## one special
splitForm_SM(~1+diag(f|g)+s(x, bs = "tp"))
splitForm_SM(~1+diag(f|g)+s(x, k=7, bs = "tp"))

}
\author{
Steve Walker
}
