% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_REs.R
\name{splitTerm_RE}
\alias{splitTerm_RE}
\title{Split a RE term into component parts}
\usage{
splitTerm_RE(term, specials = findValidCovStructs())
}
\arguments{
\item{term}{\itemize{
\item term to split
}}

\item{specials}{\itemize{
\item character vector of "specials" to handle
}}
}
\value{
a list
}
\description{
Split a RE term into component parts

Function to split a RE term into component parts.
}
\details{
The returned list has elements
\itemize{
\item{covstr - string indicating covariance type}
\item{re - random effects term}
\item{group - grouping factor}
}
}
