% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Txt.R
\name{splitText}
\alias{splitText}
\title{Split text string into a character vector}
\usage{
splitText(txt, split = "\\\\n")
}
\arguments{
\item{txt}{\itemize{
\item text string (1-elenent character vector) to split
}}

\item{split}{\itemize{
\item string used to split text (default="\\n)
}}
}
\value{
a character vector
}
\description{
Function to split text string to a character vector.
}
\details{
The text is plit into a character vector using \code{\link[stringr:str_split]{stringr::str_split_1()}}.
}
