% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{squarewave}
\alias{squarewave}
\title{Create a "square wave"}
\usage{
squarewave(zL, zR, zs, dz = 1, shftfac = 2, sclfac = 20)
}
\arguments{
\item{zL}{\itemize{
\item smallest size at which function is non-zero
}}

\item{zR}{\itemize{
\item largest size at which function is non-zero
}}

\item{dz}{\itemize{
\item width to adjust zs by (typically, binwidth; default=1)
}}

\item{shftfac}{\itemize{
\item factor to scale dz by when adjusting location by so w(z==z0) = 1 (default=20)
}}

\item{sclfac}{\itemize{
\item factor to scale dz by to obtain \code{plogis} scale
}}

\item{-zs}{\itemize{
\item object with sizes at which to calculate the function
}}
}
\value{
object same size as zs
}
\description{
Function to create a "square wave".
}
\details{
Provides a differentiable function such that \verb{w(zL<=z<=zR) = 1} and
0 otherwise.
}
