% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{squarewave_right}
\alias{squarewave_right}
\title{Create a "square wave" opened to the right}
\usage{
squarewave_right(z0, zs, dz = 1, shftfac = 2, sclfac = 20)
}
\arguments{
\item{z0}{\itemize{
\item smallest size at which function is non-zero
}}

\item{dz}{\itemize{
\item width to adjust zs by (typically, binwidth; default=1)
}}

\item{shftfac}{\itemize{
\item factor to scale dz by when adjusting location by so w(z==z0) = 1 (default=20)
}}

\item{sclfac}{\itemize{
\item factor to scale dz by to obtain \code{plogis} scale
}}

\item{-}{zs - object with sizes at which to calculate the function}
}
\value{
object same size as zs
}
\description{
Function to create a "square wave" opened to the right.
}
\details{
Provides a differentiable function such that \code{w(z>=z0) = 1} and
\code{w(z<z0) = 0}. Uses \code{\link[=plogis]{plogis()}} to calculate the logistic curve, with
\code{location = z0} and \code{scale = dz/sclfac}. The \code{zs} are shifted to the left
by \code{dz/shftfac} so that w(z==z0) = 1 and w(z>z0) = 1.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
z = 0:100;
w = squarewave_right(50,z);
if (require(ggplot2)){
  ggplot(data.frame(z=z,w=w),aes(x=z,y=w)) + geom_point() + geom_line();
}
\dontshow{\}) # examplesIf}
}
