% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{stackedLogistic1}
\alias{stackedLogistic1}
\title{Calculates a 5-parameter "stacked" logistic selectivity curve}
\usage{
stackedLogistic1(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute function values}

\item{params}{- vector of function parameters}

\item{refZ}{- ignored}
}
\value{
- named vector of selectivity values
}
\description{
Function to calculate a 5-parameter "stacked" logistic selectivity curve.
}
\details{
Calculates 5-parameter "stacked" logistic selectivity curve parameterized by
\itemize{
     \item params[1]: weighting factor on the first curve
     \item params[2]: size at inflection point for 1st logistic curve
     \item params[3]: sd for 1st logistic curve
     \item params[4]: size at inflection point for 2nd logistic curve
     \item params[5]: sd for the 2nd logistic curve
}
}
