% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{stackedLogistic2}
\alias{stackedLogistic2}
\title{Calculates a 6-parameter "stacked" logistic selectivity curve}
\usage{
stackedLogistic2(z, params, refZ = 0, debug = FALSE)

stackedLogistic2(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute function values
}}

\item{params}{\itemize{
\item vector of function parameters
}}

\item{refZ}{\itemize{
\item ignored
}}

\item{pRefZ}{\itemize{
\item ignored
}}
}
\value{
\itemize{
\item vector of selectivity values
}

\itemize{
\item named vector of selectivity values
}
}
\description{
Function to calculate a 6-parameter "stacked" logistic selectivity curve.

Function to calculate a 6-parameter "stacked" logistic selectivity curve.
}
\details{
The parameter pRefZ is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.

Calculates 6-parameter "stacked" logistic selectivity curve parameterized by
\itemize{
\item params\link{1}: asymptote for 1st logistic curve
\item params\link{2}: size at inflection point for 1st logistic curve
\item params\link{3}: sd for 1st logistic curve
\item params\link{4}: asymptote for complete stacked curve
\item params\link{5}: size at inflection point for 2nd logistic curve
\item params\link{6}: sd for the 2nd logistic curve
}

Calculates 6-parameter "stacked" logistic selectivity curve parameterized by
\itemize{
\item params\link{1}: asymptote for 1st logistic curve
\item params\link{2}: size at inflection point for 1st logistic curve
\item params\link{3}: sd for 1st logistic curve
\item params\link{4}: asymptote for complete stacked curve
\item params\link{5}: size at inflection point for 2nd logistic curve
\item params\link{6}: sd for the 2nd logistic curve
}
}
