% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopDynamics.R
\name{stepPopNs}
\alias{stepPopNs}
\title{Step the population numbers one tine interval}
\usage{
stepPopNs(y, s, N_z)
}
\arguments{
\item{y}{- year index}

\item{s}{- season index}
}
\description{
Function to step the population numbers one time interval
}
\details{
The population is proagated forward one time step. Potential processes involved
in the process are:
 - molting/change in shell condition
 - maturation
 - growth
 - natural mortality
 - fishing mortality
 -
}
