% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged_array.R
\name{str.ragged_array}
\alias{str.ragged_array}
\title{Display structure of a ragged_array}
\usage{
\method{str}{ragged_array}(x, ...)
}
\arguments{
\item{x}{- the object}
}
\value{
TRUE or FALSE
}
\description{
Function to display structure of a ragged_array.
}
\details{
Returns TRUE of the object inherits from class "ragged_array".
}
