% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged_array.R
\name{[.ragged_array}
\alias{[.ragged_array}
\title{Subset a ragged_array object}
\usage{
\method{[}{ragged_array}(x, i = NULL)
}
\arguments{
\item{x}{\itemize{
\item the ragged array object
}}

\item{i}{\itemize{
\item a vector of indices by which to subset \code{x}
}}
}
\value{
the subsetted ragged_array object
}
\description{
Subset a ragged_array object by its vector index.
}
\details{
The "dfrDims" attribute for the returned ragged_array object is
the "dfrDims" attribute of \code{x}, but appropriately subsetted based on its row index
to match the subsetted vector underlying \code{x}.
}
\examples{
# simple subsetting
x = ragged_array(1:5);
x[2:3];


}
