% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged_array.R
\name{[<-.ragged_array}
\alias{[<-.ragged_array}
\title{Assign values to a ragged_array object}
\usage{
\method{[}{ragged_array}(v) <- value
}
\arguments{
\item{v}{- the values to assign}

\item{x}{- the LHS ragged array object in the assignment}

\item{y}{- the indices into x, or a dimsMap}
}
\value{
x, the LHS object
}
\description{
Assign values to a ragged_array object.
}
\details{
TODO!
}
