% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Misc.R
\name{symlogistic}
\alias{symlogistic}
\title{Calculate a 0-symmetric logistic function between -1 and 1}
\usage{
symlogistic(x)
}
\arguments{
\item{x}{- numeric vector at which to calculate the symmetric logistic function}
}
\value{
a numeric vector with the same size and names (if any) as `x`
}
\description{
Function to calculate a logistic function centered on 0
}
\details{
This is the inverse to `symlogit` (see examples)
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
##--R context
x = seq(-5,5,0.1);
y = symlogistic(x);
head(y);
xp = symlogit(y) - x
##--RTMB context
x = RTMB::AD(seq(-5,5,0.1),force=TRUE);
y = symlogistic(x);
head(y)
\dontshow{\}) # examplesIf}
}
