% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Misc.R
\name{symlogit}
\alias{symlogit}
\title{Calculate a 0-symmetric logit function}
\usage{
symlogit(x)
}
\arguments{
\item{x}{- numeric vector with values from -1.0 to 1.0 at which to calculate the symmetric logit function}
}
\value{
a numeric vector with the same size and names (if any) as `x`
}
\description{
Function to calculate a logit function centered on 0
}
\details{
This is the inverse to `symlogistic` (see examples)
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
##--R context
x = seq(-0.99,0.99,0.01);
y = symlogit(x);
head(y);
xp = symlogistic(y) - x
##--RTMB context
x = RTMB::AD(seq(-0.49,0.49,0.01),force=TRUE);
y = symlogit(x);
head(y)
\dontshow{\}) # examplesIf}
}
