% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Transforms.R
\name{tf_apply}
\alias{tf_apply}
\title{Convert values using a transform function}
\usage{
tf_apply(tform, x)
}
\arguments{
\item{tform}{\itemize{
\item name of transform function
}}

\item{x}{\itemize{
\item vector of values to convert
}}
}
\value{
vector of transformed values
}
\description{
Function to convert values sing a transform function.
}
