% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Transforms.R
\name{tf_g2kg}
\alias{tf_g2kg}
\title{Convert values from grams to kilograms}
\usage{
tf_g2kg(x)
}
\arguments{
\item{x}{\itemize{
\item vector of values to convert
}}
}
\value{
vector of transformed values
}
\description{
Function to convert values from grams to kilograms.
}
