% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Transforms.R
\name{tf_log}
\alias{tf_log}
\title{Transform values to log-space}
\usage{
tf_log(x)
}
\arguments{
\item{x}{\itemize{
\item vector of values to convert
}}
}
\value{
vector of transformed values
}
\description{
Function to convert values to logspace.
}
