% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Transforms.R
\name{tf_none}
\alias{tf_none}
\title{Return the input with no changes}
\usage{
tf_none(x)
}
\arguments{
\item{x}{- vector of values to "convert"}
}
\value{
vector of "transformed" values
}
\description{
A pass-through transformation function (doesn't change values of input).
}
\details{
This constitutes a function that simply returns the input.
}
