% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tform_inv}
\alias{tform_inv}
\title{Compute inverse transforms on a vector of values}
\usage{
tform_inv(fcn, val, force = FALSE)
}
\arguments{
\item{fcn}{\itemize{
\item vector of function names used to apply "forward" transform
}}

\item{val}{\itemize{
\item vector of values to apply inverse transform(s) to
}}
}
\value{
numeric (or adsparse) vector of transformed values
}
\description{
Function to compute inverse transforms on a vector of values.
}
\details{
If the length of \code{fcn} is not equal to the length of \code{val}, it
is repeated until this condition is reached.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
tfcns = c('symlogistic','symlogit',   'ident','log');
ifcns = c('symlogit',   'symlogistic','ident','exp');
vals = c(0.5,0.5,10,10);
rest = tform(tfcns,vals);
tform(ifcns,rest);
tform_inv(fcns,rest);
rest = tform(tfcns,vals,force=TRUE);
tform(ifcns,rest,force=TRUE);
tform_inv(fcns,rest,force=TRUE);
\dontshow{\}) # examplesIf}
}
