% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Txt.R
\name{whichDims}
\alias{whichDims}
\title{Get which model dimensions are identified in a character vector}
\usage{
whichDims(char_vec)
}
\arguments{
\item{char_vec}{- character vector to check for model dimension names}
}
\value{
character vector with identifed model dimension names
}
\description{
Function to get which model dimensions are identified in a character vector.
}
