% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Alls.R
\name{alls_ExpandInDataframe}
\alias{alls_ExpandInDataframe}
\title{Expand "all" levels in a dataframe}
\usage{
alls_ExpandInDataframe(dfr, lstAlls, verbose = FALSE)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe with dimensions to expand from "all" to the individual levels
}}

\item{lstAlls}{\itemize{
\item list in format returned by \code{\link[=alls_GetLevels]{alls_GetLevels()}}
}}
}
\value{
expanded dataframe
}
\description{
Function to expand "all" levels in a dataframe.
}
\details{
The \code{lstAlls} object can either be a list of dataframes returned from \code{\link[=alls_GetLevels]{alls_GetLevels()}} mapping
"all"s to each model dimension or a "roll your own" list.

The "roll your own"
list can be in the format returned from \code{\link[=alls_GetLevels]{alls_GetLevels()}} (i.e., a dataframe of levels for each "dimension" to expand)
or as a named list with a vector of the values to expand to for each "dimension". In either case, the names of the list
elements correspond to the associated "dimension" to expand.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
dfr = tibble::tibble(x=1:3,y="all");
#--`lstAlls` in format returned from [alls_GetLevels()]
lstAlls = list(y=tibble::tibble(value=as.character(paste0("y",c(1:2,1:2))),y=c("all","all",paste0("y",as.character(1:2)))));
alls_ExpandInDataframe(dfr,lstAlls);

lstAlls = list(y=c("y1","y2"));
alls_ExpandInDataframe(dfr,lstAlls);
\dontshow{\}) # examplesIf}
}
