% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Alls.R
\name{alls_ExpandInDataframe}
\alias{alls_ExpandInDataframe}
\title{Expand "all" levels in a dataframe}
\usage{
alls_ExpandInDataframe(dfr, lstAlls, verbose = FALSE)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe with dimensions to expand from "all" to levels
}}

\item{lstAlls}{\itemize{
\item list returned by \code{\link[=alls_GetLevels]{alls_GetLevels()}}
}}
}
\value{
expanded dataframe
}
\description{
Function to expand "all" levels in a dataframe.
}
