% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Alls.R
\name{alls_GetLevels}
\alias{alls_GetLevels}
\title{Get maps from "all" dimension levels to individual levels}
\usage{
alls_GetLevels(dfrDims, verbose = FALSE)
}
\arguments{
\item{dfrDims}{- dataframe with model dimensions to use for expansion}
}
\value{
list with a dataframe for each dimension mapping "all" to every value (and every value to itself).
}
\description{
Function to get maps from "all" dimension levels to individual levels.
}
