% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{asclogistic}
\alias{asclogistic}
\title{Calculate an ascending logistic function parameterized by z50 and slope}
\usage{
asclogistic(z, z50, slope, fsz = 0)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{z50}{- size at which selectivity  = 0.5 (logit-scale mean)}

\item{slope}{- slope at z50}

\item{fsz}{- if fsz>0, fsz=fully-selected size. if fsz<0, function is normalized to max. if fsz=0, no re-scaling is done}
}
\value{
vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending logistic function parameterized by z50 and slope.
}
\details{
None.
}
