% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{asclogistic}
\alias{asclogistic}
\title{Calculate an ascending logistic function parameterized by z50 and slope}
\usage{
asclogistic(z, params, fsz = 0, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{params}{- 2-element vector with selectivity function parameters}

\item{debug}{- flag (T/F) to print debugging messages}

\item{refZ}{- reference size}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending logistic function parameterized by z50 and slope.
}
\details{
The parameter values are
\itemize{
\item{params[1] - z50   - size at which selectivity = 0.5 (logit-scale mean)}
\item{params[2] - slope - slope at z50}
}

If `refZ`>0, `refZ`=fully-selected size. if `refZ`<0, function is normalized to max.
if `refZ`=0, no re-scaling is done.
}
