% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{asclogistic}
\alias{asclogistic}
\title{Calculate an ascending logistic function parameterized by z50 and slope}
\usage{
asclogistic(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivities
}}

\item{params}{\itemize{
\item 2-element vector with selectivity function parameters
}}

\item{refZ}{\itemize{
\item reference size
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
named vector with selectivity values at the elements of \code{z}
}
\description{
Function to calculate an ascending logistic function parameterized by z50 and slope.
}
\details{

}
