% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{asclogistic1}
\alias{asclogistic1}
\title{Calculate an ascending logistic function parameterized by z50 and width (1/slope)}
\usage{
asclogistic1(z, params, refZ = 0, debug = FALSE)

asclogistic1(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivities
}}

\item{params}{\itemize{
\item 2-element vector with selectivity function parameters
}}

\item{refZ}{\itemize{
\item reference size
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}

\item{pZ50}{\itemize{
\item size at 50\\% selectivity
}}

\item{pWdZ}{\itemize{
\item width (1/slope) at \code{pZ50}
}}

\item{pRefZ}{\itemize{
\item reference size
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
named vector with selectivity values at the elements of \code{z}

named vector with selectivity values at the elements of \code{z}
}
\description{
Function to calculate an ascending logistic function parameterized by z50 and width.

Function to calculate an ascending logistic function parameterized by z50 and width.
}
\details{
The parameter pRefZ is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.

If \code{pRefZ}>0, \code{pRefZ}=fully-selected size.
If \code{pRefZ}<0, function is normalized to max.
If \code{pRefZ}=0, no re-scaling is done.


}
