% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{asclogistic5095}
\alias{asclogistic5095}
\title{Calculate an ascending logistic function parameterized by z50 and z95}
\usage{
asclogistic5095(z, params, refZ = 0, debug = FALSE)

asclogistic5095(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivities
}}

\item{params}{\itemize{
\item 2-element vector with selectivity function parameters
}}

\item{refZ}{\itemize{
\item reference size
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}

\item{pZ50}{\itemize{
\item ize at which selectivity = 0.50 (logit-scale mean)
}}

\item{pZ95}{\itemize{
\item size at which selectivity = 0.95
}}

\item{pRefZ}{\itemize{
\item reference size
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
vector with selectivity values at the elements of z

named vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending logistic function parameterized by z50 and z95.

Function to calculate an ascending logistic function parameterized by z50 and z95.
}
\details{
The parameter pRefZ is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.

If \code{pRefZ}>0, \code{pRefZ}=fully-selected size.
If \code{pRefZ}<0, function is normalized to max.
If \code{pRefZ}=0, no re-scaling is done.

The parameter values are
\itemize{
\item{params\link{1} - z50: size at which selectivity = 0.50 (logit-scale mean)}
\item{params\link{2} - z95: size at which selectivity = 0.95}
}

If \code{refZ}>0, \code{refZ}=fully-selected size.
If \code{refZ}<0, function is normalized to max.
If \code{refZ}=0, no re-scaling is done.
}
