% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{asclogistic5095}
\alias{asclogistic5095}
\title{Calculate an ascending logistic function parameterized by z50 and z95}
\usage{
asclogistic5095(z, z50, z95, fsz = 0)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{z50}{- size at which selectivity  = 0.5 (logit-scale mean)}

\item{z95}{- size at which selectivity  = 0.95}

\item{fsz}{- if fsz>0, fsz=fully-selected size. if fsz<0, function is normalized to max. if fsz=0, no re-scaling is done}

\item{debug}{- flag (T/F) to print debugging messages}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending logistic function parameterized by z50 and z95.
}
\details{
None.
}
