% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{asclogistic50D95}
\alias{asclogistic50D95}
\title{Calculate an ascending logistic function parameterized by z50 and D95 (=z95-z50)}
\usage{
asclogistic50D95(z, z50, D95, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{z50}{- size at which selectivity  = 0.5 (logit-scale mean)}

\item{D95}{- z95-z50}

\item{refZ}{- reference size}

\item{debug}{- flag (T/F) to print debugging messages}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending logistic function parameterized by z50 and D95.
}
\details{
None.
}
