% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{asclogistic50D95}
\alias{asclogistic50D95}
\title{Calculate an ascending logistic function parameterized by z50 and D95 (=z95-z50)}
\usage{
asclogistic50D95(z, params, refZ = 0, debug = FALSE)

asclogistic50D95(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivity values
}}

\item{params}{\itemize{
\item 2-element vector with selectivity function parameters
}}

\item{refZ}{\itemize{
\item reference size
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}

\item{pZ50}{\itemize{
\item size at which selectivity = 0.5 (logit-scale mean)
}}

\item{pZ9550}{\itemize{
\item z95-z50: difference between sizes at 95\\% and 50\\%-selected
}}

\item{pRefZ}{\itemize{
\item reference size
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
vector with selectivity values at the elements of z

named vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending logistic function parameterized by z50 and D95.

Function to calculate an ascending logistic function parameterized by z50 and D95.
}
\details{
The parameter pRefZ is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.

If \code{pRefZ}>0, \code{pRefZ}=fully-selected size.
If \code{pRefZ}<0, function is normalized to max.
If \code{pRefZ}=0, no re-scaling is done.


}
