% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{ascnormal}
\alias{ascnormal}
\title{Calculate an ascending normal selectivity function parameterized by size-at-1 and width}
\usage{
ascnormal(z, params, fsz = 0)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{params}{- selectivity parameters}

\item{fsz}{- if fsz>0, fsz=fully-selected size. if fsz<0, function is normalized to max. if fsz=0, no re-scaling is done}
}
\value{
vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending normal selectivity function parameterized by size-at-1 and width.
}
\details{
The parameters are
\itemize{
\item{p[1] - size at which ascending limb hits 1 (mean of normal distribution)}
\item{p[2] - width of ascending limb (standard deviation ofnormal distribution)}
}
}
