% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{ascnormal1}
\alias{ascnormal1}
\title{Calculate an ascending normal selectivity curve}
\usage{
ascnormal1(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivity values
}}

\item{params}{\itemize{
\item 2-element vector of selectivity function parameters
}}

\item{refZ}{\itemize{
\item reference size (not used)
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending normal selectivity curve.
}
\details{
The parameters are
\itemize{
\item{params\link{1} - width of ascending limb (standard deviation of a normal distribution)}
\item{params\link{2} - size at which ascending limb hits 1 (mean of a normal distribution)}
}
}
