% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{ascnormal2}
\alias{ascnormal2}
\title{Calculate ascending normal selectivity curve}
\usage{
ascnormal2(
  z,
  params = NULL,
  ascSref = NULL,
  ascZ1 = NULL,
  refZ = 0,
  debug = FALSE
)

ascnormal2(
  z,
  params = NULL,
  ascSref = NULL,
  ascZ1 = NULL,
  refZ = 0,
  debug = FALSE
)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivity values
}}

\item{params}{\itemize{
\item 2-element vector of selectivity function parameters
}}

\item{ascSref}{\itemize{
\item selectivity at size = refZ (if \code{params} not given)
}}

\item{ascZ1}{\itemize{
\item size at which ascending limb reaches 1 (if \code{params} not given)
}}

\item{refZ}{\itemize{
\item size at which function reaches params\link{2}
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}

\item{pAscZ1}{\itemize{
\item size at which ascending limb reaches 1
}}

\item{pAscRefS}{\itemize{
\item selectivity at size = pRefZ
}}

\item{pRefZ}{\itemize{
\item reference size (a constant) at which function reaches pAscRefS
}}

\item{dZ}{\itemize{
\item bin size reference for join (square wave) function
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
vector with selectivity values at the elements of z

named vector with selectivity values at the elements of z
}
\description{
Function to calculate ascending normal selectivity curve.

Function to calculate ascending normal selectivity curve.
}
\details{
The parameter pRefZ is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.
This function uses a left square wave \code{\link[=squarewave_left]{squarewave_left()}} that rises
from 0 to 1 "at" pAscZ1 to create a differentiable function.

The parameter vector has values
\itemize{
\item{params\link{1}: ascSref - selectivity at size = refZ}
\item{params\link{2}: ascZ1 - size at which ascending limb reaches 1}
}
}
