% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{ascnormal2a}
\alias{ascnormal2a}
\title{Calculate an ascending normal selectivity curve}
\usage{
ascnormal2a(z, params, refS = 0.5, debug = FALSE)

ascnormal2a(z, params, refS = 0.5, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute function values
}}

\item{params}{\itemize{
\item vector of function parameters
}}

\item{refS}{\itemize{
\item reference selectivity value (default=0.5)
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}

\item{pAscZ1}{\itemize{
\item size at which ascending limb reaches 1
}}

\item{pZatRefS}{\itemize{
\item size at which selectivity = pRefS
}}

\item{pRefS}{\itemize{
\item reference selectivity value (constant)
}}

\item{dZ}{\itemize{
\item bin size reference for join (square wave) function
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
vector with selectivity values at the elements of z

named vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending normal selectivity curve

Function to calculate an ascending normal selectivity curve
}
\details{
The parameter pRefS is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.
This function uses a left square wave \code{\link[=squarewave_left]{squarewave_left()}} that rises
from 0 to 1 "at" pAscZ1 to create a differentiable function.

The parameter vector has values
\itemize{
\item{params\link{1}: size at which selectivity = refS}
\item{params\link{2}: size at which ascending limb reaches 1}
}
}
