% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{ascnormal2b}
\alias{ascnormal2b}
\title{Calculate an ascending normal selectivity curve}
\usage{
ascnormal2b(z, params, refS = 0, debug = FALSE)
}
\arguments{
\item{z}{- dvector of sizes at which to compute function values}

\item{params}{- dvar_vector of function parameters}

\item{refS}{- selectivity at params[1]-params[2]}
}
\value{
- named vector of selectivity values
}
\description{
Function to calculate an ascending normal selectivity curve.
}
\details{
Function is parameterized by
\itemize{
     \item params[1]: size at which ascending limb reaches 1
     \item params[2]: delta from size at 1 to size at which selectivity=refS
}
}
