% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{ascnormal2b}
\alias{ascnormal2b}
\title{Calculate an ascending normal selectivity curve}
\usage{
ascnormal2b(z, params, refS = 0, debug = FALSE)

ascnormal2b(z, params, refS = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item dvector of sizes at which to compute function values
}}

\item{params}{\itemize{
\item dvar_vector of function parameters
}}

\item{refS}{\itemize{
\item selectivity at params\link{1}-params\link{2}
}}

\item{pAscZ1}{\itemize{
\item size at which ascending limb reaches 1
}}

\item{pDZ2RefS}{\itemize{
\item delta from size at 1 to size at which selectivity=pRefS
}}

\item{pRefS}{\itemize{
\item selectivity at pMnZ-pDZ2RefS
}}
}
\value{
\itemize{
\item vector of selectivity values
}

\itemize{
\item named vector of selectivity values
}
}
\description{
Function to calculate an ascending normal selectivity curve.

Function to calculate an ascending normal selectivity curve.
}
\details{
The parameter pRefS is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.
This function uses a left square wave \code{\link[=squarewave_left]{squarewave_left()}} that rises
from 0 to 1 "at" pAscZ1 to create a differentiable function.

Function is parameterized by
\itemize{
\item params\link{1}: size at which ascending limb reaches 1
\item params\link{2}: delta from size at 1 to size at which selectivity=refS
}
}
