% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllometry.R
\name{calcAllometry}
\alias{calcAllometry}
\title{Calculate allometry for all model categories across time}
\usage{
calcAllometry(dims, info, params, verbose = FALSE)
}
\arguments{
\item{dims}{- dimensions list}

\item{info}{- info list (output list from [extractParamInfo_Allometry()])}

\item{params}{- RTMB parameters list with allometry-specific elements}

\item{verbose}{- flag to print diagnostic info}
}
\value{
list
}
\description{
Function to calculate allometryfor all model categories across time.
}
