% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanGrowthFunctions.R
\name{calcMnGrowthPwrLaw1}
\alias{calcMnGrowthPwrLaw1}
\title{Power law model 1 for mean growth}
\usage{
calcMnGrowthPwrLaw1(z, params, consts = NULL, debug = FALSE)
}
\arguments{
\item{z}{- vector of pre-molt sizes (bin mid-points)}

\item{params}{- vector of parameters (potentially estimated)}

\item{consts}{- vector of constants (not estimated)}

\item{debug}{- debugging flag (T/F)}
}
\value{
named vector of mean molt increments or post-molt size
}
\description{
Function to calculate power law model 1 for mean growth
}
\details{
Mean growth (molt increment *or* post-molt size) `g` is described as
a power law function of pre-molt size `z` by
g(z|p, c) = exp(p[1]) * z^p[2]
where `p` represents the parameter vector.
exp(`p[1]`) represents the mean growth at pre-molt size z=1.

The interpretation of `g` as representing a molt increment or post-molt size
is up to the user.
}
