% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MeanGrowth.R
\name{calcMnGrowthPwrLaw3}
\alias{calcMnGrowthPwrLaw3}
\title{Power law model 2 for mean growth}
\usage{
calcMnGrowthPwrLaw3(z, params, consts, debug = FALSE)

calcMnGrowthPwrLaw3(z, params, consts, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of pre-molt sizes (bin mid-points)
}}

\item{params}{\itemize{
\item vector of parameters (potentially estimated)
}}

\item{consts}{\itemize{
\item vector of constants (not estimated)
}}

\item{debug}{\itemize{
\item debugging flag (T/F)
}}
}
\value{
named vector of mean molt increments or post-molt size

named vector of mean molt increments or post-molt size
}
\description{
Function to calculate power law model 2 for mean growth

Function to calculate power law model 3 for mean growth
}
\details{
Mean growth (molt increment \emph{or} post-molt size) \code{g} is described as
a power law function of pre-molt size \code{z} by
g(z|p, c) = p\link{1} * (z/c\link{1})^(log(p\link{2}/p\link{1})/log(c\link{2}/c\link{1}))
where \code{p} represents the parameter vector and \code{c} represents the consts vector.
\code{p[1]} is the mean growth at pre-molt size \code{c[1]}.
\code{p[2]} is the mean growth at pre-molt size \code{c[2]}.

The interpretation of \code{g} as representing a molt increment or post-molt size
is up to the user.

Mean growth (molt increment \emph{or} post-molt size) \code{g} is described as
a power law function of pre-molt size \code{z} by
g(z|p, c) = p\link{1} * (z/c\link{1})^p\link{2}
where \code{p} represents the parameter vector and \code{c} represents the consts vector.
\code{p[1]} is the mean growth at pre-molt size \code{c[1]}.

The interpretation of \code{g} as representing a molt increment or post-molt size
is up to the user.
}
