% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRecruitment.R
\name{calcRecruitment}
\alias{calcRecruitment}
\title{Calculate recruitment-related quantitites}
\usage{
calcRecruitment(dims, options, params, verbose)
}
\arguments{
\item{dims}{\itemize{
\item model dimensions list
}}

\item{options}{\itemize{
\item model options list
}}

\item{params}{\itemize{
\item model parameters list
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
list with
}
\description{
Function to calculate recruitment-related quantities.
}
\details{
Options for calculating recruitment-related quantities include:
\itemize{
\item{?? - ??}
}
Returned list includes:
\itemize{
\item{?? - ??}
}
}
