% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSelectivity.R
\name{calcSelectivity}
\alias{calcSelectivity}
\title{Calculate selectivity functions for all model categories across time}
\usage{
calcSelectivity(dims, info, params, verbose = FALSE)
}
\arguments{
\item{dims}{\itemize{
\item dimensions list
}}

\item{info}{\itemize{
\item info list (output list from \code{\link[=extractParamInfo_Selectivity]{extractParamInfo_Selectivity()}})
}}

\item{params}{\itemize{
\item RTMB parameters list with elements specific to selectivity
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
A list of size-specific selectivity arrays by selectivity function
}
\description{
Function to calculate the selectivity functions for all model categories across time.
}
\details{
If \code{f_}, \code{y_}, and \code{s_} are the selectivity function, year, and season of interest,
then the indices into the population vector and associated selectivity values are
given by \code{ic_} and \code{selVals}
selVals = lst[\link{f_}][y_,s_,ic_];
}
