% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSelFunctions.R, R/rsim_calcSelFcns.R
\name{calcSelFcns}
\alias{calcSelFcns}
\title{Calculate all selectivity/retention functions}
\usage{
calcSelFcns(mc, showPlot = TRUE)

calcSelFcns(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{showPlot}{- flag to show plots}

\item{dims}{- model dimensions list}

\item{options}{- model options list}

\item{params}{- model parameters list}

\item{verbose}{- flag to print diagnostic info}
}
\value{
list with

array sel_cz with dimensions 'pc', 'z'. The label associated with
each pc can be obtained as mc$dims$selfcns$lbls[pc].
}
\description{
Function to calculate all selectivity/retention functions.

Function to calculate selectivity/retention functions.
}
\details{
Options for calculating selectivity/retention functions include:
\itemize{
 \item{?? - ??}
}

None.
}
