% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSelFunctions.R, R/rsim_calcSelFcns.R
\name{calcSelFcns}
\alias{calcSelFcns}
\title{Calculate all selectivity/retention functions}
\usage{
calcSelFcns(mc, showPlot = TRUE)

calcSelFcns(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{\itemize{
\item model configuration object
}}

\item{showPlot}{\itemize{
\item flag to show plots
}}

\item{dims}{\itemize{
\item model dimensions list
}}

\item{options}{\itemize{
\item model options list
}}

\item{params}{\itemize{
\item model parameters list
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
list with

array sel_cz with dimensions 'pc', 'z'. The label associated with
each pc can be obtained as mc$dims$selfcns$lbls\link{pc}.
}
\description{
Function to calculate all selectivity/retention functions.

Function to calculate selectivity/retention functions.
}
\details{
Options for calculating selectivity/retention functions include:
\itemize{
\item{?? - ??}
}

None.
}
