% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{const_sel}
\alias{const_sel}
\title{Calculate a constant-valued selectivity curve}
\usage{
const_sel(z, params, debug = FALSE)
}
\arguments{
\item{z}{- sizes at which to compute selectivity values}

\item{params}{- parameter vector}

\item{debug}{- flag (T/F) to print debugging messages}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate a constant-valued selectivity curve
}
\details{
The parameter vector has values
\itemize{
 \item{params[1]: the constant value}
}
}
