% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParamInfo.R
\name{createParamInfo}
\alias{createParamInfo}
\title{Get parameter information for fixed effects, random effects, and smooths in a formula}
\usage{
createParamInfo(
  formula,
  model_frame,
  contrasts = NULL,
  drop.unused.levels = FALSE,
  offset = NULL,
  sparse = TRUE
)
}
\arguments{
\item{formula}{\itemize{
\item a formula from which to extract parameter info
}}

\item{model_frame}{\itemize{
\item the model frame
}}

\item{contrasts}{\itemize{
\item a list of contrasts for fixed effects
}}

\item{offset}{\itemize{
\item the name of the column in \code{model_frame} to be used as a fixed offset (or NULL)
}}

\item{sparse}{\itemize{
\item (logical) return sparse model matrices?
}}
}
\value{
a list composed of
\itemize{
\item{parameter}{the name of the model parameter}
\item{formula}{the input formula}
\item{model_frame - the input model frame}
\item{contrasts - list of contrasts used to determine model matrices}
\item{drop.unused.levels - flag (T/F) to drop unused levels from design matrices}
\item{offset - the name of the model_frame column used as a fixed offset}
\item{piFEs}{parameter info list for fixed effects from \code{\link[=createParamsInfo_FEs]{createParamsInfo_FEs()}}}
\item{piREs}{parameter info list for random effects from \code{\link[=createParamsInfo_REs]{createParamsInfo_REs()}}}
\item{piSMs}{parameter info list for smooths from \code{\link[=createParamsInfo_SMs]{createParamsInfo_SMs()}}}
}
}
\description{
Function to get parameter information for fixed effects, random effects, and smooths in a formula.
}
\details{
The
}
