% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{dbllogistic}
\alias{dbllogistic}
\title{Calculate a double logistic selectivity curve}
\usage{
dbllogistic(z, params, fsz = 0, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivity curve}

\item{params}{- 4-element vector of selectivity function parameters}

\item{debug}{- flag (T/F) to print debugging messages}

\item{refZ}{- reference size}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate a double logistic selectivity curve.
}
\details{
The parameter values are
\itemize{
\item{params[1] - ascZ50   - ascending logistic size at which selectivity  = 0.5 (logit-scale mean)}
\item{params[2] - ascSlope - ascending logistic slope at z50}
\item{params[3] - dscZ50   - descending logistic size at which selectivity  = 0.5 (logit-scale mean)}
\item{params[4] - dscSlope - descending logistic slope at z50}
}

If `refZ`>0, `refZ`=fully-selected size. if `refZ`<0, function is normalized to max.
if `refZ`=0, no re-scaling is done.
}
