% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{dbllogistic5095}
\alias{dbllogistic5095}
\title{Calculate a double logistic function parameterized by z50 and D95=z95-z50 for ascending/descending limbs}
\usage{
dbllogistic5095(z, ascZ50, ascZ95, dscZ95, dscZ50, fsz = 0)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{ascZ50}{- ascending z50}

\item{ascZ95}{- z95 on ascending limb}

\item{dscZ95}{- z95 on descending limb}

\item{dscZ50}{- descending z50}

\item{fsz}{- if fsz>0, fsz=fully-selected size. if fsz<0, function is normalized to max. if fsz=0, no re-scaling is done}
}
\value{
vector with selectivity values at the elements of z
}
\description{
Function to calculate a double logistic function parameterized by z50, z95 on ascending/descending limbs.
}
\details{
None.
}
