% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{dbllogistic5095}
\alias{dbllogistic5095}
\title{Calculate a double logistic function parameterized by z50 and D95=z95-z50 for ascending/descending limbs}
\usage{
dbllogistic5095(z, params, refZ = 0, debug = FALSE)

dbllogistic5095(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute selectivities
}}

\item{params}{\itemize{
\item 4-element parameter vector
}}

\item{refZ}{\itemize{
\item reference size
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging messages
}}

\item{pAscZ50}{\itemize{
\item ascending limb size at 50\\% selected
}}

\item{pAscZ95}{\itemize{
\item ascending limb size at 95\\% selected
}}

\item{pDscZ95}{\itemize{
\item descending limb size at 95\\% selected
}}

\item{pDscZ50}{\itemize{
\item descending limb size at 50\\% selected
}}

\item{pRefZ}{\itemize{
\item reference size
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging messages
}}
}
\value{
vector with selectivity values at the elements of z

named vector with selectivity values at the elements of z
}
\description{
Function to calculate a double logistic function parameterized by z50, z95 on ascending/descending limbs.

Function to calculate a logistic function parameterized by z50, z95 on ascending/descending limbs.
}
\details{
The parameter pRefZ is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.

If \code{pRefZ}>0, pRefZ is the fully-selected size.
if \code{pRefZ}<0, function is normalized to max.
if \code{pRefZ}=0, no re-scaling is done.


}
