% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{dbllogistic5095}
\alias{dbllogistic5095}
\title{Calculate a double logistic function parameterized by z50 and D95=z95-z50 for ascending/descending limbs}
\usage{
dbllogistic5095(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{params}{- 4-element parameter vector}

\item{refZ}{- reference size}

\item{debug}{- flag (T/F) to print debugging messages}
}
\value{
named vector with selectivity values at the elements of z
}
\description{
Function to calculate a double logistic function parameterized by z50, z95 on ascending/descending limbs.
}
\details{
The parameters are
\itemize{
 \item{params[1] - ascending limb size at 50\% selected}
 \item{params[2] - ascending limb size at 95\% selected}
 \item{params[3] - descending limb size at 95\% selected}
 \item{params[4] - descending limb size at 50\% selected}
}

If `refZ`>0, fsz=fully-selected size. if `refZ`<0, function is normalized to max.
if `refZ`=0, no re-scaling is done.
}
