% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{dblnormal4a}
\alias{dblnormal4a}
\title{Calculates a 4-parameter normal selectivity curve}
\usage{
dblnormal4a(z, params, refZ, debug = FALSE)

dblnormal4a(z, params, refZ, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute function values
}}

\item{params}{\itemize{
\item vector of function parameters
}}

\item{refZ}{\itemize{
\item max possible size (e.g., max(z))
}}

\item{pAscZ1}{\itemize{
\item size at which ascending limb reaches 1
}}

\item{pAscWd}{\itemize{
\item width of ascending limb
}}

\item{pDscSclDZ}{\itemize{
\item scaled increment to pAscZ1 at which descending limb departs from 1
}}

\item{pDscWd}{\itemize{
\item width of descending limb
}}

\item{pRefZ}{\itemize{
\item max possible size (e.g., max(z))
}}

\item{verbose}{\itemize{
\item flag to print debugging info
}}
}
\value{
\itemize{
\item named vector of selectivity values
}

\itemize{
\item named vector of selectivity values
}
}
\description{
Function to calculate a 4-parameter normal selectivity curve.

Function to calculate a 4-parameter normal selectivity curve.
}
\details{
The parameter pRefZ is a constant and must be identified as such in the
\code{map} list when MakeADFun'ing an objective function that uses this function.

Calculates 4-parameter normal function parameterized by
\itemize{
\item params\link{1}: size at which ascending limb reaches 1
\item params\link{2}: width of ascending limb
\item params\link{3}: scaled increment to params\link{1} at which descending limb departs from 1
\item params\link{4}: width of descending limb
}
}
