% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{dblnormal6}
\alias{dblnormal6}
\title{Calculates a 6-parameter normal selectivity curve}
\usage{
dblnormal6(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute function values}

\item{params}{- vector of function parameters}

\item{refZ}{- ignored}
}
\value{
- named vector of selectivity values
}
\description{
Function to calculate a 6-parameter normal selectivity curve.
}
\details{
Calculates 6-parameter normal function parameterized by
\itemize{
     \item params[1]: size at which ascending limb reaches 1
     \item params[2]: width of ascending limb
     \item params[3]: size at which descending limb departs from 1
     \item params[4]: width of descending limb
     \item params[5]: floor of ascending limb
     \item params[6]: floor of descending limb
}
}
