% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{elem_div}
\alias{elem_div}
\title{Calculate the element-wise division of two variables}
\usage{
elem_div(x, y)
}
\arguments{
\item{x}{- numerator variable (vector, matrix, etc.)}

\item{y}{- divisor variable conformable to x}
}
\value{
(unnamed) object x/y
}
\description{
ADMB-equivalent to calculate the element-wise division of two variables.
}
\details{
Substitution for ADMB `elem_div` function.
}
