% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscADMBfunctions.R
\name{elem_prod}
\alias{elem_prod}
\title{Calculate the element-wise product of two variables}
\usage{
elem_prod(x, y)
}
\arguments{
\item{x}{- variable (vector, matrix, etc.)}

\item{y}{- variable conformable to x}
}
\value{
(unnamed) object x*y
}
\description{
Function to calculate the element-wise product of two variables.
}
\details{
Substitution for ADMB `elem_prod` function.
}
