% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Text.R
\name{evalTextAsCode}
\alias{evalTextAsCode}
\title{Evaluate text as lines of code}
\usage{
evalTextAsCode(strv, frame = 0)
}
\arguments{
\item{strv}{\itemize{
\item character vector with lines of code to evaluate.
}}
}
\value{
nothing
}
\description{
Function to evaluate text as lines of code.
}
\details{
The lines of \code{code} in strv are evaluated in the frame specified by
\code{frame}. If \code{frame} is non-negative, it specifies the parent frame relative to the
caller (so \code{frame=0}, the deault, is evaluated in the caller's frame). If \code{frame} is negative,
the lines of code are evaluated in the environment specified by its absolute value.
}
\examples{
\dontshow{if ( FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
str = "
  v1  = 5;
  v2 = 1:v1;
  sum = list(x="contr.sum",m="contr.sum");
      ";
# example code
\dontshow{\}) # examplesIf}
}
