% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Text.R
\name{evalTextAsDataframe}
\alias{evalTextAsDataframe}
\title{Evaluate a character vector as a dataframe (a \code{tbl_df})}
\usage{
evalTextAsDataframe(char_vec, comment = "#")
}
\arguments{
\item{char_vec}{\itemize{
\item character vector to parse/evaluate as a dataframe
}}

\item{comment}{\itemize{
\item string indicating that what follows is a comment
}}
}
\value{
a dataframe (actually a \code{tbl_df}, see \code{\link[tibble:tibble]{tibble::tibble()}})
}
\description{
Function to evaluate a character vector as a dataframe (a \code{tbl_df}).
}
\details{
Each element in \code{char_vec} is regarded as a line of text (as in a text file).
The dataframe structure and column types are determined by \code{\link[readr:read_table]{readr::read_table()}}.

This function differs from \code{\link[=extractDataframe]{extractDataframe()}} in that, after comments are removed are
removed from the input character vector \code{char_vec}, the resulting
character vector is regarded as defining the dataframe. In \code{\link[=extractDataframe]{extractDataframe()}},
the input character vector may be a much larger section of text than what defines the
dataframe to be extracted, but the number of rows (\code{n}) to extract is known as, is the index of
the element to begin with (\code{start}).
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
str<-paste0('
  id        function  frame   params              description
  pwrLaw1   pwrLaw1   mfALL   pA,pB               w(z)_=_pA_*_(z^pB)
  pwrLaw2   pwrLaw2   mf2024  pLnA,pLnS,pB,pZ0    w(z)_=_exp(pLnA+pLnS+pB*ln(z/pZ0))
')
dfr = txt |> splitText() |> evalTextAsDataframe();
\dontshow{\}) # examplesIf}
}
