% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Text.R
\name{evalTextAsDigits}
\alias{evalTextAsDigits}
\title{Evaluate strings in a character vector as digits}
\usage{
evalTextAsDigits(x)
}
\arguments{
\item{x}{\itemize{
\item character vector to parse/evaluate
}}
}
\value{
integer vector
}
\description{
Function to evaluate strings in a character vector as digits.
}
\details{
For each element in \code{x}, everything except for digits is removed and
the resulting character value is converted to an integer.
}
\examples{
evalTextAsDigits(c("ab23","cd42","a42 b23"))

}
