% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Dataframe.R
\name{expandDataframe}
\alias{expandDataframe}
\title{Expand dataframe for expressions and "all" dimension levels}
\usage{
expandDataframe(dfr, lstAlls = NULL, verbose = TRUE)
}
\arguments{
\item{dfr}{- dataframe with expressions and dimensions to expand from "all" to levels}

\item{lstAlls}{- list returned by [alls_GetLevels()], or NULL}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe expanded first by lstAlls, then by evaluating expressions
}
\description{
Function to expand dataframe for expressions and "all" dimension levels.
}
\details{
Expressions are evaluated by column row-wise, by column index. Consequently,
columns evaluated later can refer to values in columns previously evaluated.
Values in expressions will be coerced to the the column type of the original
dataframe.
}
