% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Text.R
\name{extractDataframe}
\alias{extractDataframe}
\title{Extract a dataframe (a \code{\link[tibble:tibble]{tibble::tibble()}}) from a character vector}
\usage{
extractDataframe(txt, n, start, comment = "#")
}
\arguments{
\item{txt}{\itemize{
\item character vector to search
}}

\item{n}{\itemize{
\item number of non-comment lines to extract
}}

\item{start}{\itemize{
\item index into \code{txt} to regard as the start of the search
}}

\item{comment}{\itemize{
\item character indicating the following text in the string is a comment
}}
}
\value{
list with elements \code{txt} and \code{end} (see details)
}
\description{
Function to a dataframe (a \code{\link[tibble:tibble]{tibble::tibble()}}) from a character vector.
}
\details{
Each element in \code{txt} is regarded as a line of text (as in a text file).
The dataframe structure and column types are determined by \code{\link[readr:read_table]{readr::read_table()}}.
\itemize{
\item{txt - character vector of length \code{n} with comment lines removed (i.e., a subset of \code{txt})}
\item{end - index of the last element extracted in the input \code{txt} vector}
}

This function differs from \code{\link[=evalTextAsDataframe]{evalTextAsDataframe()}} in that
the input character vector may be a much larger section of text than what defines the
dataframe to be extracted, but the number of rows (\code{n}) to extract is known, as is the index of
the element to begin with (\code{start}). In \code{\link[=evalTextAsDataframe]{evalTextAsDataframe()}}, after comments are removed are
removed from the input character vector \code{char_vec}, the resulting
character vector is regarded as defining the dataframe.
}
