% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParamInfo_Allometry.R
\name{extractParamInfo_Allometry}
\alias{extractParamInfo_Allometry}
\title{Extract allometry parameters from parameter info list}
\usage{
extractParamInfo_Allometry(lst, dims = NULL, verbose = TRUE)
}
\arguments{
\item{lst}{- parameter info list from [readParamInfo_Allometry()]}

\item{dims}{- `dims` list from `setupModelDims`}
}
\value{
a list
}
\description{
Function to extract allometry parameters from parameter info list.
}
\details{
TODO: fill in.
}
