% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParamInfoFunctionType1.R
\name{extractParamInfoFunctionType1}
\alias{extractParamInfoFunctionType1}
\title{Extract function-type parameter info from a list}
\usage{
extractParamInfoFunctionType1(
  lst,
  dms = NULL,
  process_type = "",
  verbose = TRUE
)
}
\arguments{
\item{lst}{\itemize{
\item function-type parameter info list from a \code{readParamInfo_}
}}

\item{dms}{\itemize{
\item appropriate \code{dms...} DimsMap from \code{setupModelDims}
}}
}
\value{
a list (see details)
}
\description{
Function to extract function-type parameter info from a list.
}
\details{
The output list has elements

\itemize{
\item{option - format option}
\item{params - vector of weight-at-size values}
\item{map - vector for the RTMB \code{map} list indicating the parameters are fixed; i.e. not to be estimated}
\item{dfrIdx2Pars - dataframe with columns "pidx", dimension names (y,s,r,x,m,p,z), and parameter specs IV,LB,UB,phz,PriorType,Pr1,Pr2
indicating the un-expanded mapping from \code{pidx} to the parameter input values ("IV"). \code{pidx} is the index into the parameter vector.
The dimension columns are incidental and reflect the specification from the input text file.}
\item{dfrDims2Pars - dataframe with columns "pidx" and the dimension names (y,s,r,x,m,p,z).
The dimension levels are fully expanded such that each row corresponds to a single "multidimensional" level.
The value of \code{pidx} for each row indicates the index of the associated weight-at-size value in the parameter vector.}
}

If \code{lst$option} is "function", the output list has elements
\itemize{
}
}
