% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Text.R
\name{extractTextAfterString}
\alias{extractTextAfterString}
\title{Extract text after a string}
\usage{
extractTextAfterString(txt, str)
}
\arguments{
\item{txt}{- character vector from which to extract strings}

\item{str}{- string to extract after (see NOTE under details)}
}
\value{
character vector of extracted strings
}
\description{
Function to extract text after a string
}
\details{
For each element in the vector, this extracts the string following
the `after` string from each element of the input
character vector, or `NA` if no `after` string found.

NOTE: Characters with special regex meanings (e.g., ")","]") need to
be double-escaped to use in `str`.
}
\examples{
# example code
extractTextAfterString("test:fred[x,y]==0","==");
extractTextAfterString(c("fred[x,y]==0: FALSE","b[z0]: 1"),":");
extractTextAfterString("fred[x,y[z]]=0","=");
extractTextAfterString("fred[x,y[z]]=0","\\\\[");
extractTextAfterString("fred[x,y[z]]=0","fred\\\\[");

}
