% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Text.R
\name{extractTextBetweenBrackets}
\alias{extractTextBetweenBrackets}
\title{Extract text between first and last brackets}
\usage{
extractTextBetweenBrackets(txt)
}
\arguments{
\item{txt}{- character vector from which to extract strings}
}
\value{
character vector of extracted strings
}
\description{
Function to extract text between first and last brackets
}
\details{
For each element in the vector, this extracts the string between the
*first* "[" and *last* "]" from each element of the input
character vector, or `NA` if no matching brackets.
}
\examples{
# example code
extractTextBetweenBrackets("fred[x,y]=0");
extractTextBetweenParens(c("fred[x,y]=0","b[z0]"));
extractTextBetweenParens("fred[x,y[z]]=0");

}
