% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Text.R
\name{extractTextBetweenParens}
\alias{extractTextBetweenParens}
\title{Extract text between first and last parentheses}
\usage{
extractTextBetweenParens(txt)
}
\arguments{
\item{txt}{- character vector from which to extract strings}
}
\value{
character vector of extracted strings
}
\description{
Function to extract text between first and last parentheses
}
\details{
For each element in the vector, this extracts the string between the
*first* "(" and *last* ")" from each element of the input
character vector, or `NA` if no matching parentheses.
}
\examples{
# example code
extractTextBetweenParens("fred(x,y)=0");
extractTextBetweenParens(c("fred(x,y)=0","b(z0)"));
extractTextBetweenParens("fred(x,y(z))=0");

}
