% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Text.R
\name{extractTextSection}
\alias{extractTextSection}
\title{Extract a text section with only non-comment lines from a character vector}
\usage{
extractTextSection(txt, n = length(txt), start = 1, comment = "#")
}
\arguments{
\item{txt}{\itemize{
\item character vector to search
}}

\item{n}{\itemize{
\item number of non-comment lines to extract (default=length(txt))
}}

\item{start}{\itemize{
\item index into \code{txt} to regard as the start of the search (default=1)
}}

\item{comment}{\itemize{
\item character indicating the following text in the string is a comment
}}
}
\value{
list with elements \code{txt} and \code{end} (see details)
}
\description{
Function to extract a text section with only non-comment lines from a character vector.
}
\details{
Each element in \code{txt} is regarded as a line of text (as in a text file).
The function searches for the first \code{n} non-comment lines and extracts them as a new vector.
The returned list has elements
\itemize{
\item{txt - character vector of length \code{n} with comment lines removed (i.e., a subset of \code{txt})}
\item{end - index of the last element extracted in the input \code{txt} vector}
}
}
