% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDataFile.R
\name{findNextLine}
\alias{findNextLine}
\title{Find next uncommented line}
\usage{
findNextLine(lns, iln)
}
\arguments{
\item{lns}{- character vector to search}

\item{iln}{- index at which to start search}
}
\value{
index into `lns` at which the next uncommented line occurs
}
\description{
Function to find next uncommented line in a vector of character strings.
}
