% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWatZ.R, R/getAllometry.R
\name{getAllometry}
\alias{getAllometry}
\title{Get allometry info as a list}
\usage{
getAllometry(dims, opts, params, verbose = FALSE)

getAllometry(dims, opts, params, verbose = FALSE)
}
\arguments{
\item{dims}{- dimensions list}

\item{opts}{- options list}

\item{params}{- RTMB parameters list}

\item{verbose}{- flag to print diagnostic info}
}
\value{
list with elements
\itemize{
  \item{wAtZ - 3-d array (y,s,c) with weight-at-size by year, season, pop class}
}

list
}
\description{
Function to get allometry info as a list. Used in the objective function.

Function to get allometry
}
