% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWatZ.R
\name{getAllometry}
\alias{getAllometry}
\title{Get allometry info as a list}
\usage{
getAllometry(dims, opts, params, verbose = FALSE)
}
\arguments{
\item{dims}{\itemize{
\item 
}}

\item{opts}{\itemize{
\item 
}}

\item{params}{\itemize{
\item (full) parameters list
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostics
}}
}
\value{
list with elements
\itemize{
\item{wAtZ - 3-d array (y,s,c) with weight-at-size by year, season, pop class}
}
}
\description{
Function to get allometry info as a list. Used in the objective function.
}
