% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDFR_sdreport.R
\name{getDFR_sdreport}
\alias{getDFR_sdreport}
\title{Convert an sdreport object to a dataframe}
\usage{
getDFR_sdreport(obj, report = TRUE, model = "")
}
\arguments{
\item{obj}{\itemize{
\item RTMB sdreport object
}}

\item{report}{\itemize{
\item flag (T/F) to ??
}}

\item{model}{\itemize{
\item name of model
}}
}
\value{
dataframe
}
\description{
Function to convert an sdreport object to a dataframe.
}
\details{
TBD
}
