% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Text.R
\name{getDimNames}
\alias{getDimNames}
\title{Get model dimension names}
\usage{
getDimNames()
}
\value{
character vector with model dimension names
}
\description{
Function to get model dimension names.
}
