% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{getLinkFcn}
\alias{getLinkFcn}
\title{Return a list with the reuested link function and it's inverse}
\usage{
getLinkFcn(txt)
}
\arguments{
\item{x}{\itemize{
\item vector of values to "convert"
}}
}
\value{
vector of "transformed" values
}
\description{
Function to return a list with the reuested link function and it's inverse
}
\details{
This constitutes a function that simply returns the input.
}
