% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{getLinkFcn}
\alias{getLinkFcn}
\title{Return the input with no changes}
\usage{
getLinkFcn(txt)
}
\arguments{
\item{x}{\itemize{
\item vector of values to "convert"
}}
}
\value{
vector of "transformed" values
}
\description{
A pass-through transformation function (doesn't change values of input).
}
\details{
This constitutes a function that simply returns the input.
}
