% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{logit}
\alias{logit}
\title{Calculate variable on logit scale}
\usage{
logit(x)
}
\arguments{
\item{x}{- variable}
}
\value{
vector log(x/(1-x)) with same names as x
}
\description{
Function to calculate variable on logit scale.
}
\details{
inverse of logisitc function
}
