% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{logit}
\alias{logit}
\title{Convert values to the logit scale}
\usage{
logit(x)
}
\arguments{
\item{x}{\itemize{
\item numeric object
}}
}
\value{
log(x/(1-x)), with same names and class as \code{x}
}
\description{
Function to convert values to the logit scale.
}
\details{
inverse of logistic function. Values of \code{x} outside the
(0,1) range generate NaNs.
}
