% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFileTemplates.R
\name{makeFileTemplate_Param}
\alias{makeFileTemplate_Param}
\title{Create a template for the information necessary to specify all components of a model parameter}
\usage{
makeFileTemplate_Param(
  parameter,
  x,
  model_frame = NULL,
  contrasts = NULL,
  drop.unused.levels = TRUE,
  offset = NULL,
  debug = TRUE
)
}
\arguments{
\item{parameter}{\itemize{
\item parameter name
}}

\item{x}{\itemize{
\item formula or list from \code{\link[=createParamInfo]{createParamInfo()}}
}}

\item{model_frame}{\itemize{
\item model frame for parameter (required if \code{x} is a formula)
}}

\item{contrasts}{\itemize{
\item a list of contrasts (if not NULL)
}}

\item{drop.unused.levels}{\itemize{
\item flag to drop unused levels
}}
}
\value{
a character vector
}
\description{
Function to create a template for the information necessary to specify all components of a model parameter.
}
