% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFileTemplates.R
\name{makeFileTemplate_ParamFEs}
\alias{makeFileTemplate_ParamFEs}
\title{Create a template for the information necessary to specify the FE components of a model parameter}
\usage{
makeFileTemplate_ParamFEs(
  x,
  model_frame = NULL,
  contrasts = NULL,
  drop.unused.levels = TRUE,
  debug = TRUE
)
}
\arguments{
\item{x}{\itemize{
\item formula, possibly containing fixed effects, or list from \code{\link[=createParamInfo_FeEs]{createParamInfo_FeEs()}}
}}

\item{model_frame}{\itemize{
\item model frame for parameter (required if \code{x} is a formula)
}}

\item{contrasts}{\itemize{
\item a list of contrasts (if not NULL)
}}

\item{drop.unused.levels}{\itemize{
\item flag to drop unused levels
}}
}
\value{
a character vector
}
\description{
Function to create a template for the information necessary to specify the FE components of a model parameter.
}
