% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFileTemplates.R
\name{makeFileTemplate_ParamSMs}
\alias{makeFileTemplate_ParamSMs}
\title{Create a template for the information necessary to specify the smooths for a model parameter TODO: finish this function!}
\usage{
makeFileTemplate_ParamSMs(
  x,
  model_frame,
  contrasts = NULL,
  drop.unused.levels = TRUE,
  debug = TRUE
)
}
\arguments{
\item{x}{\itemize{
\item formula, possibly containing fixed effects, or list from \code{\link[=createParamInfo_SMs]{createParamInfo_SMs()}}
}}

\item{model_frame}{\itemize{
\item model frame for parameter
}}

\item{contrasts}{\itemize{
\item a list of contrasts (if not NULL)
}}
}
\value{
a character vector
}
\description{
Function to create a template for the information necessary to specify the smooths for a model parameter.
}
