% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{mkPrecMat.ar1}
\alias{mkPrecMat.ar1}
\title{Make a precision matrix for (possibly multiple) AR1 or iid processes}
\usage{
mkPrecMat.ar1(n, sigma=1, rho=0, sdX=NULL, sparse=TRUE)
}
\arguments{
\item{n}{integer vector > 0, number of elements in each AR1 process (no default)}

\item{sigma}{\itemize{
\item vector with the standard deviation for each white noise process
}}

\item{rho}{\itemize{
\item vector with the correlation coefficient between successive values for each AR1 process (default=0)
}}

\item{sdX}{\itemize{
\item vector with standard deviations of the observed process (default=1)
}}

\item{sparse}{\itemize{
\item (T/F) Should the matrix be of class 'dsCMatrix' (default=0)
}}
}
\value{
a precision matrix with block-diagonal AR1 structure and attributes "params",
a vector with the values for \code{n}, \code{sigma} and \code{rho}, and "qType", which is "diag"
if \code{rho} is 0 and "ar1" otherwise.
}
\description{
Function for creating a precision matrix for
(possibly multiple) AR1 or iid (if \code{rho}=0) processes
}
\details{
A precision matrix for multiple independent AR1 and/or iid processes can be
created by making \code{n}, \code{sigma} (or \code{sdX}), and \code{rho} vectors. \code{sigma} (or \code{sdX}), and \code{rho}
are expanded (or truncated) to the length of \code{n}.

Note that the variance associated with individual observations is

$$ var(X_i) = E(X^2_i)-\mu^2 = \frac{\sigma^2}{1-\rho^2} $$

and thus

$$ \sigma = \sqrt{{1-\rho^2}} \cdot sd(X) $$

and consequently the precision matrix can (optionally) be specified in terms of the standard
deviation of the observed process (\code{sdX}).
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
##--R context
Q <- mkPrecMat.ar1(n=5,sigma=2,rho=0.5,sparse=TRUE);
Q <- mkPrecMat.ar1(n=c(5,5),sigma=2,rho=0.5,sparse=TRUE);
C <- solve(Q); #--covariance matrix

Qp <- mkPrecMat.ar1(n=5,sigma=0,rho=0.5, sdX=2/sqrt(1-0.5^2), sparse=TRUE);


##--RTMB context
require(RTMB);
Q <- mkPrecMat.ar1(n=5,sigma=2, rho=0.5, sdX=0, sparse=TRUE, force=TRUE);
C <- solve(Q); #--covariance matrix
Qp <- mkPrecMat.ar1(n=5, sigma=0, rho=0.5, sdX=2/sqrt(1-0.5^2), sparse=TRUE, force=TRUE);
Q-Qp
Qp <- mkPrecMat.ar1(n=5, sigma=0, rho=RTMB::AD(0.5,force=TRUE), sdX=RTMB::AD(2/sqrt(1-0.5^2),force=TRUE), sparse=TRUE, force=FALSE);
\dontshow{\}) # examplesIf}
}
