% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_CovQs.R
\name{mkPrecMat}
\alias{mkPrecMat}
\title{Make a precision matrix}
\usage{
mkPrecMat(n, pars, type, sparse = TRUE, force = FALSE)
}
\arguments{
\item{n}{\itemize{
\item dimension of the matrix
}}

\item{pars}{\itemize{
\item vector of parameter values determining the precision matrix
}}

\item{type}{\itemize{
\item type of precision matrix
}}

\item{sparse}{\itemize{
\item flag to create a sparse or dense matrix
}}

\item{force}{\itemize{
\item force an RTMB adsparse representation
}}
}
\value{
a precision matrix (a \link[Matrix:Matrix]{Matrix::Matrix} type or \link[RTMB:RTMB-package]{RTMB::RTMB} "adsparse" type)
}
\description{
Function to make a precision matrix.
}
\details{

}
