% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{parseList}
\alias{parseList}
\title{Parse multiple text lines into a list with named elements}
\usage{
parseList(lns, verbose = FALSE)
}
\arguments{
\item{lns}{\itemize{
\item the text lines parse
}}

\item{names}{\itemize{
\item the names of the elements of the output list
}}
}
\value{
named list of parsed values
}
\description{
Function to parse multiple text lines into a list with named elements until reaching an ">EOL<" marker.
}
\details{
Each element in the text to be parsed is identified as to type
using a line that starts with "name: type", where type is one of "vector", "matrix",
or "dataframe".

IMPORTANT: A counter with the name \code{iln} should be defined in the global environment
while this function runs. The parsing of lines in \code{lns} uses the value of iln
at the time the function is called as the starting index in \code{lns}.
When the function exits, the value of \code{iln} indicates the index of the last element
examined.
}
