% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{parseVal}
\alias{parseVal}
\title{Parse a string from an input data file to a value}
\usage{
parseVal(con, nl, ln = NULL)
}
\arguments{
\item{str}{\itemize{
\item string to parse
}}

\item{type}{\itemize{
\item expected value type (not used currently)
}}
}
\value{
a value with mode as determined by \code{\link[readr:parse_guess]{readr::parse_guess()}}
}
\description{
Function to parse a string from an input data file to a value.
}
\details{
Splits the string by whitespace and drops anything after the comment character,
using \code{\link[=scan]{scan()}} with \code{comment.char}="#" and \code{strip.white}=TRUE followed by \code{\link[readr:parse_guess]{readr::parse_guess()}}.
}
