% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_ReadParse.R
\name{parseVector}
\alias{parseVector}
\title{Parse a 1-element text vector into a vector}
\usage{
parseVector(lns, verbose = FALSE)
}
\arguments{
\item{lns}{\itemize{
\item the text lines parse
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging info
}}
}
\value{
vector of parsed values
}
\description{
Function to parse a 1-element text vector into a vector.
}
\details{
IMPORTANT: A counter with the name \code{iln} should be defined in the global environment
while this function runs. The line \code{lns[iln]} is parsed to create the vector.
The value of \code{iln} is not incremented (because there was no search).
}
