% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDataFile.R
\name{readEnvironmentalData}
\alias{readEnvironmentalData}
\title{Read Growth Data section of input data file}
\usage{
readEnvironmentalData(lns)
}
\description{
Read Growth Data section of input data file
}
