% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADMB_DataFileFunctions.R
\name{readSizeCompsDataFrameFormat1}
\alias{readSizeCompsDataFrameFormat1}
\title{Read a size comps frame in GMACS input format 1}
\usage{
readSizeCompsDataFrameFormat1(lnsp)
}
\description{
Read a size comps frame in GMACS input format 1
}
