% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{selSpline}
\alias{selSpline}
\title{Calculates a selectivity curve using a spline function}
\usage{
selSpline(z, params, consts, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute function values}

\item{params}{- logit-scale vector of function parameters}

\item{consts}{- vector of knots}
}
\value{
- named vector of selectivity values
}
\description{
Function to calculate a selectivity curve using a spline function.
}
\details{
Calculates a selectivity curve using a spline function based on
[RTMB::splinefun()]. The values in the `params` vector are the logit-scale
values at the knots. The first `n` values in the `consts` vector are the knots,
where `n` is th length of `params`.
}
