% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{selSpline}
\alias{selSpline}
\title{Calculates a selectivity curve using a spline function}
\usage{
selSpline(z, params, consts, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute function values
}}

\item{params}{\itemize{
\item logit-scale vector of function parameters
}}

\item{consts}{\itemize{
\item vector of knots
}}
}
\value{
\itemize{
\item named vector of selectivity values
}
}
\description{
Function to calculate a selectivity curve using a spline function.
}
\details{
Calculates a selectivity curve using a spline function based on
\link[RTMB:splinefun-method]{RTMB::splinefun-method}. The values in the \code{params} vector are the logit-scale
values at the knots. The first \code{n} values in the \code{consts} vector are the knots,
where \code{n} is th length of \code{params}.
}
