% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions_Text.R
\name{skipCommentLines}
\alias{skipCommentLines}
\title{Identify first non-comment line in a character vector}
\usage{
skipCommentLines(txt, start, comment = "#")
}
\arguments{
\item{txt}{\itemize{
\item character vector to search
}}

\item{start}{\itemize{
\item index into \code{txt} to regard as the start of the search
}}

\item{comment}{\itemize{
\item character indicating the following text in the string is a comment
}}
}
\value{
the index of the first line in \code{txt} that does not start with a comment character
}
\description{
Function to identify the index of the first non-comment line in a character vector.
}
\details{
Each element in \code{txt} is regarded as a line of text (as in a text file).
The function searches for the first non-comment line and returns its index.
}
\examples{
txt = c("#this is a comment",
        "  #so is this,even thought the line starts with blank space",
        "this is not a comment",
        "this is not either, but the first is line 3.");
extractLines(txt,1); #--should be 3: txt[3] is first non-comment line
}
