% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscADMBfunctions.R
\name{square}
\alias{square}
\title{Square a variable}
\usage{
square(x)
}
\arguments{
\item{x}{- variable to square}
}
\value{
vector x*x with same names as x
}
\description{
Function to square a variable.
}
\details{
Substitution for ADMB `square` function.
}
