% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selectivity.R,
%   R/SelectivityFunctions.R
\name{stackedLogistic1}
\alias{stackedLogistic1}
\title{Calculates a 5-parameter "stacked" logistic selectivity curve}
\usage{
stackedLogistic1(z, params, refZ = 0, debug = FALSE)

stackedLogistic1(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute function values
}}

\item{params}{\itemize{
\item vector of function parameters
}}

\item{refZ}{\itemize{
\item ignored
}}

\item{pMnZ1}{\itemize{
\item size at inflection point for 1st logistic curve
}}

\item{pSdZ1}{\itemize{
\item sd for 1st logistic curve
}}

\item{pMnZ2}{\itemize{
\item size at inflection point for 2nd logistic curve
}}

\item{pSdZ2}{-sd for the 2nd logistic curve}

\item{pOmga}{\itemize{
\item weighting factor on the first curve
}}

\item{verbose}{\itemize{
\item flag to print debugging info
}}
}
\value{
\itemize{
\item vector of selectivity values
}

\itemize{
\item named vector of selectivity values
}
}
\description{
Function to calculate a 5-parameter "stacked" logistic selectivity curve.

Function to calculate a 5-parameter "stacked" logistic selectivity curve.
}
\details{
Calculates 5-parameter "stacked" logistic selectivity curve.

Calculates 5-parameter "stacked" logistic selectivity curve parameterized by
\itemize{
\item params\link{1}: weighting factor on the first curve
\item params\link{2}: size at inflection point for 1st logistic curve
\item params\link{3}: sd for 1st logistic curve
\item params\link{4}: size at inflection point for 2nd logistic curve
\item params\link{5}: sd for the 2nd logistic curve
}
}
