% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{stackedLogistic1}
\alias{stackedLogistic1}
\title{Calculates a 5-parameter "stacked" logistic selectivity curve}
\usage{
stackedLogistic1(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{\itemize{
\item vector of sizes at which to compute function values
}}

\item{params}{\itemize{
\item vector of function parameters
}}

\item{refZ}{\itemize{
\item ignored
}}
}
\value{
\itemize{
\item named vector of selectivity values
}
}
\description{
Function to calculate a 5-parameter "stacked" logistic selectivity curve.
}
\details{
Calculates 5-parameter "stacked" logistic selectivity curve parameterized by
\itemize{
\item params\link{1}: weighting factor on the first curve
\item params\link{2}: size at inflection point for 1st logistic curve
\item params\link{3}: sd for 1st logistic curve
\item params\link{4}: size at inflection point for 2nd logistic curve
\item params\link{5}: sd for the 2nd logistic curve
}
}
