% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{stackedLogistic2}
\alias{stackedLogistic2}
\title{Calculates a 6-parameter "stacked" logistic selectivity curve}
\usage{
stackedLogistic2(z, params, refZ = 0, debug = FALSE)
}
\arguments{
\item{z}{- vector of sizes at which to compute function values}

\item{params}{- vector of function parameters}

\item{refZ}{- ignored}
}
\value{
- named vector of selectivity values
}
\description{
Function to calculate a 6-parameter "stacked" logistic selectivity curve.
}
\details{
Calculates 6-parameter "stacked" logistic selectivity curve parameterized by
\itemize{
     \item params[1]: asymptote for 1st logistic curve
     \item params[2]: size at inflection point for 1st logistic curve
     \item params[3]: sd for 1st logistic curve
     \item params[4]: asymptote for complete stacked curve
     \item params[5]: size at inflection point for 2nd logistic curve
     \item params[6]: sd for the 2nd logistic curve
}
}
