% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopDynamics.R
\name{stepPopNs}
\alias{stepPopNs}
\title{Step the population numbers one season}
\usage{
stepPopNs(y, s, N_z)
}
\arguments{
\item{y}{\itemize{
\item year index
}}

\item{s}{\itemize{
\item season index
}}
}
\description{
Function to step the population numbers one season
}
\details{
The population is propagated forward one season (i.e.) time step. Potential processes involved
in the process are:
\itemize{
\item molting/change in shell condition
\item maturation
\item growth
\item natural mortality
\item fishing mortality
\item 
}
}
