% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged_array.R
\name{[<-.ragged_array}
\alias{[<-.ragged_array}
\title{Assign values to a ragged_array object}
\usage{
\method{[}{ragged_array}(x, y) <- value
}
\arguments{
\item{x}{- the LHS ragged array object in the assignment}

\item{y}{- a ragged array, vector of indices, or DimsMap by which to assign values from `value` into `x`}
}
\value{
the LHS ragged_array object
}
\description{
Assign values to ragged_array object by an index vector or DimsMap.
}
\details{
TODO!
}
