% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_other_functions.R
\name{tibbleRowToString}
\alias{tibbleRowToString}
\title{Allow Quarto-based vignettes}
\usage{
tibbleRowToString(tbl_rw)
}
\arguments{
\item{tbl_rw}{\itemize{
\item a row of a tibble
}}

\item{libname}{\itemize{
\item library name (?)
}}

\item{pkgname}{\itemize{
\item package name
}}
}
\value{
see \code{\link[tools:vignetteEngine]{tools::vignetteEngine()}}.

a string
}
\description{
Function to allow Quarto-based vignettes

Function to convert a tibble row to a string.
}
\details{
Uses \code{\link[tools:vignetteEngine]{tools::vignetteEngine()}} to create a Quarto-based vignette engine
for this package.
}
