% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{zero}
\alias{zero}
\title{Zero function}
\usage{
zero(x)
}
\arguments{
\item{x}{\itemize{
\item a numeric object to "convert"
}}
}
\value{
a numeric object of the same class and shape as the input filled with zeros
}
\description{
Function that simply returns the input object multiplied by zero.
}
\details{
This function simply returns a numeric object of the same class and shape as the input filled with zeros.
}
